% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cpp_core_documents.r
\name{a_to_alpha}
\alias{a_to_alpha}
\title{Calculate alpha angles from a-parameters}
\usage{
a_to_alpha(a)
}
\arguments{
\item{a}{the \emph{a}-parameter vector.}
}
\description{
\code{a_to_alpha} is a function for converting an a-parameter vector to an alpha angle vector.
The returned values are in the radian metric.
}
\examples{
a_to_alpha(c(1, 1))

}
