% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{addCountsToConstraintData}
\alias{addCountsToConstraintData}
\title{(Internal) Count number of pool items that match a constraint}
\usage{
addCountsToConstraintData(x, attrib)
}
\arguments{
\item{x}{a \code{\link{data.frame}} containing a single constraint data.}

\item{attrib}{an \code{\linkS4class{item_attrib}} object or a \code{\linkS4class{st_attrib}} object.}
}
\value{
\code{\link{addCountsToConstraintData}} returns an updated \code{\link{data.frame}}.
}
\description{
\code{\link{addCountsToConstraintData}} is an internal function for
counting the number of items in the pool that match a constraint.
}
\keyword{internal}
