% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{addScoreToConstraintData}
\alias{addScoreToConstraintData}
\title{(Internal) Sum scores of items in a solution that match a constraint}
\usage{
addScoreToConstraintData(x, attrib, item_idx, item_resp, item_ncat)
}
\arguments{
\item{x}{a \code{\link{data.frame}} containing a single constraint data.}

\item{attrib}{an \code{\linkS4class{item_attrib}} object.}

\item{item_idx}{item indices in the solution.}

\item{item_resp}{scores of items in the solution.}

\item{item_ncat}{number of score categories of items in the solution.}
}
\value{
\code{\link{addScoreToConstraintData}} returns an updated \code{\link{data.frame}}.
}
\description{
\code{\link{addScoreToConstraintData}} is an internal function for
summing the score of items in a solution that match a constraint.
}
\keyword{internal}
