% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{addSolutionToConstraintData}
\alias{addSolutionToConstraintData}
\title{(Internal) Count number of items in a solution that match a constraint}
\usage{
addSolutionToConstraintData(x, attrib, item_idx, all_values)
}
\arguments{
\item{x}{a \code{\link{data.frame}} containing a single constraint data.}

\item{attrib}{an \code{\linkS4class{item_attrib}} object.}

\item{item_idx}{item indices in the solution.}

\item{all_values}{for set-based assembly,
\code{TRUE} returns all values for each set,
\code{FALSE} returns descriptive statisistics.}
}
\value{
\code{\link{addSolutionToConstraintData}} returns an updated \code{\link{data.frame}}.
}
\description{
\code{\link{addSolutionToConstraintData}} is an internal function for
counting the number of items in a solution that match a constraint.
}
\keyword{internal}
