% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{addTrans}
\alias{addTrans}
\title{addTrans}
\usage{
addTrans(color, trans)
}
\arguments{
\item{color}{A vector of colors}

\item{trans}{A vector of integers between 0 and 255 where 0 is fully transparaent and 255 is fully visible.}
}
\description{
Add transparency to a color
}
