% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowtest_functions.R
\name{appendMeanInfo}
\alias{appendMeanInfo}
\title{(Internal) Append mean information to shadowtest}
\usage{
appendMeanInfo(shadow_test, v, mean_info_name)
}
\arguments{
\item{shadow_test}{a \code{\link{data.frame}} containing a shadowtest.}

\item{v}{the column name to use as grouping variable.}

\item{mean_info_name}{the column name to use for appending mean information.}
}
\value{
\code{\link{appendMeanInfo}} returns an updated shadowtest.
}
\description{
(Internal) Append mean information to shadowtest
}
\keyword{internal}
