% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_functions.R
\name{applyFading}
\alias{applyFading}
\title{(Internal) Apply fading to exposure record}
\usage{
applyFading(exposure_record, segments_to_apply, simulation_constants)
}
\arguments{
\item{exposure_record}{a named list containing exposure records.}

\item{segments_to_apply}{which segments to apply fading to.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{applyFading}} returns an updated exposure record.
}
\description{
\code{\link{applyFading}} is an internal function for applying fading to exposure record.
Specifically, the following exposure records are multiplied by \code{fading_factor}:
\itemize{
  \item{\code{n_k}: the number of examinees completed the test so far, at each segment \emph{k}.}
  \item{\code{f_k}: }
  \item{\code{a_ik}: the number of times each item \emph{i} was administered to examinees, at each segment \emph{k}}
  \item{\code{r_ik}: }
}
}
\keyword{internal}
