% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_functions.R
\name{applyOverlapConstraintsToInfo}
\alias{applyOverlapConstraintsToInfo}
\title{(Internal) Modify item information using overlap constraints}
\usage{
applyOverlapConstraintsToInfo(info, usage_flag, config, simulation_constants)
}
\arguments{
\item{info}{a length-\emph{ni} vector containing item information on each item, intended for shadow-test assembly.}

\item{usage_flag}{a length-\emph{ni} vector containing how many times each item has been administered previously to the examinee.}

\item{config}{a config object.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{applyEligibilityConstraintsToInfo}} returns an updated item information vector.
}
\description{
\code{\link{applyOverlapConstraintsToInfo}} is an internal function for modifying item information using eligibility constraints.
This is known as the big M method.
The function penalizes item information of items that have been administered previously (within examinees).
This leads to those items being deterred from selected in shadowtest assembly, unless necessary.
}
\keyword{internal}
