% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_functions.R
\name{applyShrinkageCorrection}
\alias{applyShrinkageCorrection}
\title{(Internal) Apply shrinkage correction to theta estimate}
\usage{
applyShrinkageCorrection(EAP, config_theta, j)
}
\arguments{
\item{EAP}{a named list containing an EAP theta estimate.}

\item{config_theta}{a list containing theta estimation configurations.}

\item{j}{the examinee index. Used to parse the prior SD from the config.}
}
\value{
\code{\link{applyShrinkageCorrection}} returns an updated list containing corrected EAP theta estimate.
}
\description{
\code{\link{applyShrinkageCorrection}} is an internal function for applying shrinkage correction to a theta estimate.
}
\keyword{internal}
