% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{applyThin}
\alias{applyThin}
\title{(Internal) Thin a MCMC chain}
\usage{
applyThin(posterior_sample, bayesian_constants)
}
\arguments{
\item{posterior_sample}{a vector containing values in a Markov chain.}

\item{bayesian_constants}{a named list containing Bayesian constants.}
}
\value{
\code{\link{applyThin}} returns an updated Markov chain.
}
\description{
\code{\link{applyThin}} is an internal function for
thinning a Markov chain. Burn-in is also applied.
}
\keyword{internal}
