% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_2pl}
\alias{array_info_2pl}
\title{Calculate the Fisher information for a vector of theta values according to the 2pl model}
\usage{
array_info_2pl(x, a, b)
}
\arguments{
\item{x}{a vector of theta values}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a length-one numeric vector for the difficulty parameter}
}
\description{
Calculate the Fisher information for a vector of theta values according to the 2pl model
}
\references{
{
  \insertRef{lord_theory_1952}{Shadow}

  \insertRef{birnbaum_efficient_1957}{Shadow}

  \insertRef{birnbaum_estimation_1958}{Shadow}

  \insertRef{birnbaum_further_1958}{Shadow}
}
}
