% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_3pl}
\alias{array_info_3pl}
\title{Calculate the Fisher information for a vector of theta values according to the 3pl model}
\usage{
array_info_3pl(x, a, b, c)
}
\arguments{
\item{x}{a vector of theta values}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a length-one numeric vector for the difficulty parameter}

\item{c}{a length-one numeric vector for the guessing parameter}
}
\description{
Calculate the Fisher information for a vector of theta values according to the 3pl model
}
\references{
{
  \insertRef{birnbaum_latent_1968}{Shadow}
}
}
