% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_gpc}
\alias{array_info_gpc}
\title{Calculate the Fisher information for a vector of theta values according to the generalized partial credit model}
\usage{
array_info_gpc(x, a, b)
}
\arguments{
\item{x}{a vector of theta values}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a numeric vector for the threshold parameters}
}
\description{
Calculate the Fisher information for a vector of theta values according to the generalized partial credit model
}
\references{
{
  \insertRef{muraki_generalized_1992}{Shadow}
}
}
