% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_gr}
\alias{array_info_gr}
\title{Calculate the Fisher information for a vector of theta values according to the graded response model}
\usage{
array_info_gr(x, a, b)
}
\arguments{
\item{x}{a vector of theta values}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a numeric vector for the category boundary parameters}
}
\description{
Calculate the Fisher information for a vector of theta values according to the graded response model
}
\references{
{
  \insertRef{samejima_estimation_1969}{Shadow}
}
}
