% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_p_1pl}
\alias{array_p_1pl}
\title{Calculate the probability of a correct response for a vector of theta values according to the 1pl model}
\usage{
array_p_1pl(x, b)
}
\arguments{
\item{x}{a numeric vector of theta values}

\item{b}{a length-one numeric vector for the difficulty parameter}
}
\description{
Calculate the probability of a correct response for a vector of theta values according to the 1pl model
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}
}
