% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_p_3pl}
\alias{array_p_3pl}
\title{Calculate the probability of a correct response for a vector of theta values according to the 3pl model}
\usage{
array_p_3pl(x, a, b, c)
}
\arguments{
\item{x}{a numeric vector of theta values}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a length-one numeric vector for the difficulty parameter}

\item{c}{a length-one numeric vector for the guessing parameter}
}
\description{
Calculate the probability of a correct response for a vector of theta values according to the 3pl model
}
\references{
{
  \insertRef{birnbaum_latent_1968}{Shadow}
}
}
