% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_p_m_gpc}
\alias{array_p_m_gpc}
\title{Calculate the probability of a correct response for a vector of theta values according to the 2pl model (multivariate)}
\usage{
array_p_m_gpc(x, a, b)
}
\arguments{
\item{x}{a numeric vector of theta values}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a numeric vector for the threshold parameters}
}
\description{
Calculate the probability of a correct response for a vector of theta values according to the 2pl model (multivariate)
}
\references{
{
  \insertRef{muraki_generalized_1992}{Shadow}
}
}
