% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_p_pc}
\alias{array_p_pc}
\title{Calculate probability at multiple thetas (PC)}
\usage{
array_p_pc(x, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the probability of correct response at theta values, under the partial credit model.
}
\references{
{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}
}
