% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_method.R
\docType{methods}
\name{ATA,ATA.config-method}
\alias{ATA,ATA.config-method}
\title{Run Automated Test Assembly}
\usage{
\S4method{ATA}{ATA.config}(config, Constraints, plot = FALSE,
  plotrange = c(-3, 3))
}
\arguments{
\item{config}{An \code{\linkS4class{ATA.config}} object containing configuration options.}

\item{Constraints}{A list representing optimization constraints. Use \code{\link{LoadConstraints}} for this.}

\item{plot}{Logical. If \code{TRUE}, draws Fisher information plot from the selected items.}

\item{plotrange}{Numeric. A vector of length 2 containing the lower and upper bounds of plot range. Default is \code{c(-3, 3)}.}
}
\value{
A list containing the following entries:
\itemize{
  \item{MIP} MIP model.
  \item{Selected} Solution returned.
  \item{solver} MIP solver.
  \item{obj.value} Objective value returned.
  \item{solve.time} Solver elapsed time.
}
}
\description{
Perform Automated Test Assembly from specified configurations.
}
