% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_functions.R
\docType{methods}
\name{ATA}
\alias{ATA}
\alias{ATA,ATA.config-method}
\title{Run Automated Test Assembly}
\usage{
ATA(config, Constraints, plot = FALSE, plotrange = c(-3, 3))

\S4method{ATA}{ATA.config}(config, Constraints, plot = FALSE,
  plotrange = c(-3, 3))
}
\arguments{
\item{config}{An \code{\linkS4class{ATA.config}} object containing configuration options. Use \code{\link{config.ATA}} for this.}

\item{Constraints}{A list representing optimization constraints. Use \code{\link{LoadConstraints}} for this.}

\item{plot}{Logical. Draws Fisher information plot from the selected items.}

\item{plotrange}{Numeric. A vector of length 2 containing the lower and upper bounds of plot range. Default is \code{c(-3, 3)}.}
}
\value{
A list containing the following entries:
\itemize{
  \item{\code{MIP}} MIP model.
  \item{\code{Selected}} Solution returned.
  \item{\code{solver}} MIP solver.
  \item{\code{obj.value}} Objective value returned.
  \item{\code{solve.time}} Solver elapsed time.
}
}
\description{
Perform Automated Test Assembly with specified configurations.
}
