% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{attrib_pool_C}
\alias{attrib_pool_C}
\title{attrib_pool_C}
\format{A neat format}
\description{
A pretty dataset you won't find anywhere else
}
\details{
\itemize{
  \item var. A nice variable
}
}
\keyword{datasets}
