% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_info_EB}
\alias{calc_info_EB}
\title{calc_info_EB}
\usage{
calc_info_EB(x, item_parm, ncat, model)
}
\arguments{
\item{x}{A nice parameter}

\item{item_parm}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}
}
\description{
calc_info_EB
}
