% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_info_matrix}
\alias{calc_info_matrix}
\title{Calculate the Fisher information matrix for a vector of theta values and a set of items, potentially with a mixture of different models}
\usage{
calc_info_matrix(x, item_parm, ncat, model)
}
\arguments{
\item{x}{a numeric vector of theta values}

\item{item_parm}{a matrix of item parameters}

\item{ncat}{a numeric vector of the number of response categories by item}

\item{model}{a numeric vector of the IRT model by item (1: 1pl, 2: 2pl, 3: 3pl, 4: pc, 5: gpc, 6: gr)}
}
\description{
Calculate the Fisher information matrix for a vector of theta values and a set of items, potentially with a mixture of different models
}
