% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_likelihood}
\alias{calc_likelihood}
\title{Calculate a likelihood value of theta}
\usage{
calc_likelihood(x, item_parm, resp, ncat, model)
}
\arguments{
\item{x}{a length-one numeric vector for a theta value}

\item{item_parm}{a numeric matrix of item parameters}

\item{resp}{a numeric rector of item responses}

\item{ncat}{a numeric vector of the number of response categories by item}

\item{model}{a numeric vector of the IRT model by item (1: 1pl, 2: 2pl, 3: 3pl, 4: pc, 5: gpc, 6: gr)}
}
\description{
Calculate a likelihood value of theta
}
