% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_MI_FB}
\alias{calc_MI_FB}
\title{Calculate the mutual information using fully Bayesian}
\usage{
calc_MI_FB(x, items_list, ncat, model)
}
\arguments{
\item{x}{a numeric vector of MCMC sampled theta values}

\item{items_list}{a list of item_parm matrices}

\item{ncat}{a numeric vector of the number of response categories by item}

\item{model}{a numeric vector of the IRT model by item (1: 1pl, 2: 2pl, 3: 3pl, 4: pc, 5: gpc, 6: gr)}
}
\description{
Calculate the mutual information using fully Bayesian
}
