% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_posterior}
\alias{calc_posterior}
\title{calc_posterior}
\usage{
calc_posterior(x, item_parm, resp, ncat, model, prior, prior_parm)
}
\arguments{
\item{x}{A nice parameter}

\item{item_parm}{A nice parameter}

\item{resp}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}

\item{prior}{A nice parameter}

\item{prior_parm}{A nice parameter}
}
\description{
calc_posterior
}
