% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cpp_calc_documents.r
\name{calc_thisdirinfo}
\alias{calc_thisdirinfo}
\title{(C++) For multiple items, calculate directional information}
\usage{
calc_thisdirinfo(x, item_parm, nd, ncat, model, alpha_vec)
}
\arguments{
\item{x}{the theta value. This must be a row vector in matrix form. Each column should represent a dimension.}

\item{item_parm}{a matrix containing item parameters. Each row should represent an item.}

\item{nd}{the number of dimensions. This must be a single value shared by all items.}

\item{ncat}{a vector containing the number of response categories of each item.}

\item{model}{a vector indicating item models of each item, using \itemize{
  \item{\code{102}}: M2PL model
  \item{\code{103}}: M3PL model
  \item{\code{105}}: MGPC model
  \item{\code{106}}: MGR model
}}

\item{alpha_vec}{the alpha angle vector.}
}
\description{
\code{calc_thisdirinfo()} is a function for calculating directional information using a specific angle.
These functions are designed for multiple items.
}
\details{
\code{calc_thisdirinfo()} accepts a single theta value.

This function is designed for multidimensional models.
}
\references{
Rasch, G. (1960). \emph{Probabilistic models for some intelligence and attainment tests.}
Copenhagen: Danish Institute for Educational Research.

Lord, F. M. (1952). A theory of test scores (Psychometric Monograph No. 7). Richmond, VA: Psychometric Corporation.

Birnbaum, A. (1957). \emph{Efficient design and use of tests of mental ability for various decision-making problems}
(Series Report No. 58-16. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). \emph{On the estimation of mental ability}
(Series Report No. 15. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). \emph{Further considerations of efficiency in tests of a mental ability}
(Series Report No. 17. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1968). Some latent trait models and their use in inferring an examinee's ability.
In Lord, F. M., Novick, M. R. (eds.), \emph{Statistical Theories of Mental Test Scores}, 397-479.
Reading, MA: Addison-Wesley.

Masters, G. N. (1982). A Rasch model for partial credit scoring.
\emph{Psychometrika, 47}(2), 149-174.

Andrich, D. (1978). A rating formulation for ordered response categories.
\emph{Psychometrika, 43}(4), 561-573.

Muraki, E. (1992). A generalized partial credit model: Application of an EM algorithm.
\emph{Applied Psychological Measurement, 16}(2), 159-176.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores.
\emph{Psychometrika Monograph, 17}.
}
