% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_functions.R
\docType{methods}
\name{calcDerivative}
\alias{calcDerivative}
\alias{calcDerivative,item.1pl,numeric-method}
\alias{calcDerivative,item.2pl,numeric-method}
\alias{calcDerivative,item.3pl,numeric-method}
\alias{calcDerivative,item.pc,numeric-method}
\alias{calcDerivative,item.gpc,numeric-method}
\alias{calcDerivative,item.gr,numeric-method}
\alias{calcDerivative,item.pool,numeric-method}
\alias{calcDerivative,pool.cluster,numeric-method}
\title{Calculate first derivative}
\usage{
calcDerivative(object, theta)

\S4method{calcDerivative}{item.1pl,numeric}(object, theta)

\S4method{calcDerivative}{item.2pl,numeric}(object, theta)

\S4method{calcDerivative}{item.3pl,numeric}(object, theta)

\S4method{calcDerivative}{item.pc,numeric}(object, theta)

\S4method{calcDerivative}{item.gpc,numeric}(object, theta)

\S4method{calcDerivative}{item.gr,numeric}(object, theta)

\S4method{calcDerivative}{item.pool,numeric}(object, theta)

\S4method{calcDerivative}{pool.cluster,numeric}(object, theta)
}
\arguments{
\item{object}{An instance of an item class.}

\item{theta}{A vector of theta values.}
}
\value{
First derivative values.
}
\description{
An S4 generic and its methods to calculate the first derivative of the probability function.
}
\examples{
item_1 = new("item.1pl", difficulty = 0.5)
d.item_1 = calcDerivative(item_1, seq(-3, 3, 1))
item_2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
d.item_2 = calcDerivative(item_2, seq(-3, 3, 1))
item_3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
d.item_3 = calcDerivative(item_3, seq(-3, 3, 1))
item_4 = new("item.pc", threshold = c(-1, 0, 1), ncat = 4)
d.item_4 = calcDerivative(item_4, seq(-3, 3, 1))
item_5 = new("item.gpc", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
d.item_5 = calcDerivative(item_5, seq(-3, 3, 1))
item_6 = new("item.gr", slope = 0.9, category = c(-1, 0, 1), ncat = 4)
d.item_6 = calcDerivative(item_6, seq(-3, 3, 1))
d_itempool <- calcDerivative(itempool_science, seq(-3, 3, 1))
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}

{
  \insertRef{lord_theory_1952}{TestDesign}

  \insertRef{birnbaum_efficient_1957}{TestDesign}

  \insertRef{birnbaum_estimation_1958}{TestDesign}

  \insertRef{birnbaum_further_1958}{TestDesign}
}

{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}

{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}

{
  \insertRef{muraki_generalized_1992}{TestDesign}
}

{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
