% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_method.R
\docType{methods}
\name{calcFisher,pool.cluster,numeric-method}
\alias{calcFisher,pool.cluster,numeric-method}
\title{calcFisher}
\usage{
\S4method{calcFisher}{pool.cluster,numeric}(object, theta)
}
\arguments{
\item{object}{An instance of an item class}

\item{theta}{A vector of theta values}
}
\description{
calcFisher
}
\examples{
\dontrun{
itemPool.1 = LoadItemPool("C:/item_par.csv")
#this needs to be updated for item.pool
info.itemPool.1 = calcFisher(itemPools, seq(-3, 3, 1))
}
}
