% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kl_functions.r
\docType{methods}
\name{calcKL}
\alias{calcKL}
\alias{calcKL,item_pool,numeric-method}
\alias{calcKL,item_pool,matrix-method}
\title{Calculate Kullback-Leibler information}
\usage{
calcKL(object, theta, theta_width, quadrature_unit)

\S4method{calcKL}{item_pool,numeric}(object, theta, theta_width, quadrature_unit)

\S4method{calcKL}{item_pool,matrix}(object, theta, theta_width, quadrature_unit)
}
\arguments{
\item{object}{an \code{\linkS4class{item_pool}} object.}

\item{theta}{the theta value to use. This must be a single value.}

\item{theta_width}{the theta width to use to define the range of integration. The range is \code{c(theta - theta_width, theta + theta_width)}.}

\item{quadrature_unit}{the quadrature spacing to use for integration. Smaller numbers yield more accurate values.}
}
\value{
\describe{
  \item{\code{\linkS4class{item_pool}} object:}{\code{\link{calcKL}} returns a length-\emph{ni} vector of information values.}
}
\describe{
  \item{\emph{notations}}{\itemize{
    \item{\emph{ni} denotes the number of items in the \code{\linkS4class{item_pool}} object.}
  }}
}
}
\description{
\code{\link{calcKL}} is a function to calculate Kullback-Leibler information.
}
\examples{
info_pool <- calcKL(itempool_science, 0, 1, 10)

}
\references{
Rasch, G. (1960). \emph{Probabilistic models for some intelligence and attainment tests.}
Copenhagen: Danish Institute for Educational Research.

Lord, F. M. (1952). A theory of test scores (Psychometric Monograph No. 7). Richmond, VA: Psychometric Corporation.

Birnbaum, A. (1957). \emph{Efficient design and use of tests of mental ability for various decision-making problems}
(Series Report No. 58-16. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). \emph{On the estimation of mental ability}
(Series Report No. 15. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). \emph{Further considerations of efficiency in tests of a mental ability}
(Series Report No. 17. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1968). Some latent trait models and their use in inferring an examinee's ability.
In Lord, F. M., Novick, M. R. (eds.), \emph{Statistical Theories of Mental Test Scores}, 397-479.
Reading, MA: Addison-Wesley.

Masters, G. N. (1982). A Rasch model for partial credit scoring.
\emph{Psychometrika, 47}(2), 149-174.

Andrich, D. (1978). A rating formulation for ordered response categories.
\emph{Psychometrika, 43}(4), 561-573.

Muraki, E. (1992). A generalized partial credit model: Application of an EM algorithm.
\emph{Applied Psychological Measurement, 16}(2), 159-176.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores.
\emph{Psychometrika Monograph, 17}.
}
