% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_functions.R
\docType{methods}
\name{calcLogLikelihood}
\alias{calcLogLikelihood}
\alias{calcLogLikelihood,item_pool,numeric,numeric-method}
\alias{calcLogLikelihood,item_pool,numeric,matrix-method}
\alias{calcLogLikelihood,item_pool,matrix,numeric-method}
\alias{calcLogLikelihood,item_pool,matrix,matrix-method}
\title{Calculate log-likelihood}
\usage{
calcLogLikelihood(object, theta, resp)

\S4method{calcLogLikelihood}{item_pool,numeric,numeric}(object, theta, resp)

\S4method{calcLogLikelihood}{item_pool,numeric,matrix}(object, theta, resp)

\S4method{calcLogLikelihood}{item_pool,matrix,numeric}(object, theta, resp)

\S4method{calcLogLikelihood}{item_pool,matrix,matrix}(object, theta, resp)
}
\arguments{
\item{object}{an \code{\linkS4class{item_pool}} object.}

\item{theta}{theta values to use.}

\item{resp}{the response data to use.}
}
\value{
\code{\link{calcLogLikelihood}} returns values of log-likelihoods.
}
\description{
\code{\link{calcLogLikelihood}} is a function to calculate log-likelihood values.
}
\examples{
j_pool   <- calcLogLikelihood(itempool_science, seq(-3, 3, 1), 0)

}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}

{
  \insertRef{lord_theory_1952}{TestDesign}
}

{
  \insertRef{birnbaum_efficient_1957}{TestDesign}
}

{
  \insertRef{birnbaum_estimation_1958}{TestDesign}
}

{
  \insertRef{birnbaum_further_1958}{TestDesign}
}

{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}

{
  \insertRef{masters_rasch_1982}{TestDesign}
}

{
  \insertRef{andrich_rating_1978}{TestDesign}
}

{
  \insertRef{muraki_generalized_1992}{TestDesign}
}

{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
