% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_method.R
\docType{methods}
\name{calcProb,pool.cluster,numeric-method}
\alias{calcProb,pool.cluster,numeric-method}
\title{calcProb}
\usage{
\S4method{calcProb}{pool.cluster,numeric}(object, theta)
}
\arguments{
\item{object}{An instance of an item class}

\item{theta}{A vector of theta values}
}
\description{
calcProb
}
\examples{
\dontrun{
itemPool.1 = LoadItemPool("C:/item_par_1.csv")
itemPool.2 = LoadItemPool("C:/item_par_2.csv")
itemPools = vector(mode = "list", 2)
itemPools@pools[[1]] = itemPool.1
itemPools@pools[[2]] = itemPool.2
prob.itemPools = calcProb(itemPools, seq(-3, 3, 1))
}
}
