% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{checkConstraints}
\alias{checkConstraints}
\title{checkConstraints}
\usage{
checkConstraints(constraints, usageMatrix, trueTheta = NULL)
}
\arguments{
\item{constraints}{A nice parameter}

\item{usageMatrix}{A nice parameter}

\item{trueTheta}{A nice parameter}
}
\description{
checkConstraints
}
