% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{checkExposureRate}
\alias{checkExposureRate}
\title{checkExposureRate}
\usage{
checkExposureRate(object, config)
}
\arguments{
\item{object}{A nice parameter}

\item{config}{A nice parameter}
}
\description{
checkExposureRate
}
