% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints_operators.R
\name{combineConstraintsData}
\alias{combineConstraintsData}
\title{(Internal) Combine constraints data}
\usage{
combineConstraintsData(raw1, raw2)
}
\arguments{
\item{raw1, raw2}{a \code{\link{data.frame}} containing constraints data.}
}
\value{
\code{\link{combineConstraintsData}} returns a \code{\link{data.frame}} containing the combined constraints data.
This can be supplied to \code{\link{loadConstraints}}.
}
\description{
\code{\link{combineConstraintsData}} is an internal function for combining constraints data.
Columns that exist on only one side are filled in with NA values.
}
\keyword{internal}
