% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_pool_operators.R
\name{combineItemPoolData}
\alias{combineItemPoolData}
\title{(Internal) Combine item pool data}
\usage{
combineItemPoolData(raw1, raw2, unique)
}
\arguments{
\item{raw1, raw2}{a \code{\link{data.frame}} containing item pool data.}

\item{unique}{if \code{TRUE}, rows that have duplicate IDs are removed.}
}
\value{
\code{\link{combineItemPoolData}} returns a \code{\link{data.frame}} containing the combined item pool data.
This can be supplied to \code{\link{loadItemPool}}.
}
\description{
\code{\link{combineItemPoolData}} is an internal function for combining item pool data.
Columns that exist on only one side are filled in with NA values.
}
\keyword{internal}
