% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdata_functions.R
\name{combineXdata}
\alias{combineXdata}
\title{(Internal) Combine two constraint matrix-data}
\usage{
combineXdata(x1, x2)
}
\arguments{
\item{x1, x2}{a named list containing constraint matrix-data}
}
\value{
\code{\link{combineXdata}} returns
the combined constraint matrix-data.
}
\description{
(Internal) Combine two constraint matrix-data
}
\keyword{internal}
