% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{computeInfoAtCurrentTheta}
\alias{computeInfoAtCurrentTheta}
\title{(Internal) Compute item information at current theta estimate}
\usage{
computeInfoAtCurrentTheta(
  item_selection,
  j,
  current_theta,
  item_pool,
  model_code,
  info_fixed_theta,
  info_grid,
  item_parameter_sample
)
}
\arguments{
\item{item_selection}{the \code{item_selection} slot of a \code{\linkS4class{config_Shadow}} object.}

\item{j}{the examinee index.}

\item{current_theta}{a named list containing current theta estimate.}

\item{item_pool}{an \code{\linkS4class{item_pool}} object.}

\item{model_code}{machine-sanitized item model codes for passing onto C++ functions.}

\item{info_fixed_theta}{a list containing item information at prespecified theta points.}

\item{info_grid}{a matrix containing item information across quadrature points.}

\item{item_parameter_sample}{a list containing item parameter samples for Bayesian estimation.}
}
\description{
\code{\link{computeInfoAtCurrentTheta}} is an internal function for
computing item information at current theta estimate.
}
\keyword{internal}
