% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{Shadow.config-class}
\alias{Shadow.config-class}
\alias{config.Shadow}
\title{config.Shadow}
\usage{
config.Shadow(itemSelection = NULL, contentBalancing = NULL,
  MIP = NULL, MCMC = NULL, refreshPolicy = NULL,
  exposureControl = NULL, stoppingCriterion = NULL,
  interimTheta = NULL, finalTheta = NULL, thetaGrid = seq(-4, 4,
  0.1), auditTrail = F)
}
\arguments{
\item{itemSelection}{A list containing item selection criteria.
\itemize{
  \item{\code{method}} The type of criteria. Accepts one of \code{MFI, MPWI, FB, EB}.
  \item{\code{infoType}} The type of information. Accepts \code{FISHER}.
  \item{\code{initialTheta}} 
  \item{\code{fixedTheta}} 
}}

\item{contentBalancing}{A list containing content balancing options.
\itemize{
  \item{\code{method}} The type of balancing method. Accepts one of \code{NONE, STA}.
}}

\item{MIP}{A list containing solver options.
\itemize{
  \item{\code{solver}} The type of solver. Accepts one of \code{SYMPHONY, GUROBI, GLPK, LPSOLVE}.
  \item{\code{verbosity}} Verbosity level.
  \item{\code{timeLimit}} Time limit to be passed onto solver. Used in solvers \code{SYMPHONY, GUROBI, GLPK}.
  \item{\code{gapLimit}} Gap limit to be passed onto solver. Used in solvers  \code{SYMPHONY, GUROBI}.
}}

\item{MCMC}{A list containing Markov-chain Monte Carlo configurations.
\itemize{
  \item{\code{burnIn}} Numeric. The number of chains from the start to discard.
  \item{\code{postBurnIn}}  Numeric. The number of chains to use after discarding the first \code{burnIn} chains.
  \item{\code{thin}} Thin.
  \item{\code{jumpfactor}} JumpFactor.
}}

\item{refreshPolicy}{A list containing refresh policy for obtaining a new shadow test.
\itemize{
  \item{\code{method}} The type of policy. Accepts one of \code{ALWAYS, POSITION, INTERVAL, THRESHOLD, INTERVAL-THRESHOLD, STIMULUS, SET, PASSAGE}.
  \item{\code{interval}} Integer. Set to 1 to refresh at each position, 2 to refresh at every two positions, and so on.
  \item{\code{threshold}} Numeric. The shadow test is refreshed when the absolute change in theta estimate is greater than this value.
  \item{\code{position}}
}}

\item{exposureControl}{A list containing exposure control settings.
\itemize{
  \item{\code{method}} Accepts one of \code{"NONE", "ELIGIBILITY", "BIGM", "BIGM-BAYESIAN"}.
  \item{\code{M}} 
  \item{\code{maxExposureRate}} 
  \item{\code{accelerationFactor}}
  \item{\code{nSegment}}
  \item{\code{firstSegment}}
  \item{\code{segmentCut}}
  \item{\code{initialEligibilityStats}}
  \item{\code{fadingFactor}}
  \item{\code{diagnosticStats}}
}}

\item{stoppingCriterion}{A list containing stopping criterion.
\itemize{
  \item{\code{method}} Accepts one of \code{"FIXED"}.
  \item{\code{testLength}} 
  \item{\code{minNI}}
  \item{\code{maxNI}}
  \item{\code{SeThreshold}}
}}

\item{interimTheta}{A list containing interim theta estimation options.
\itemize{
  \item{\code{method}} The type of estimation. Accepts one of \code{EAP, EB, FB}.
  \item{\code{shrinkageCorrection}}
  \item{\code{priorDist}} The type of prior distribution. Accepts one of \code{NORMAL, UNIF}.
  \item{\code{priorPar}} Distributional parameters for the prior.
  \item{\code{boundML}}
  \item{\code{maxIter}}
  \item{\code{crit}}
}}

\item{finalTheta}{A list containing final theta estimation options.
\itemize{
  \item{\code{method}} The type of estimation. Accepts one of \code{EAP, EB, FB}.
  \item{\code{shrinkageCorrection}}
  \item{\code{priorDist}} The type of prior distribution. Accepts one of \code{NORMAL, UNIF}.
  \item{\code{priorPar}} Distributional parameters for the prior.
  \item{\code{boundML}}
  \item{\code{maxIter}}
  \item{\code{crit}}
}}

\item{thetaGrid}{A numeric vector. Theta values to represent the continuum.}

\item{auditTrail}{Logical.}
}
\description{
Create a \code{\linkS4class{Shadow.config}} object for Shadow Test Assembly (STA).
}
