% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{constraint-class}
\alias{constraint-class}
\title{constraint}
\description{
constraint
}
\section{Slots}{

\describe{
\item{\code{CONSTRAINT}}{Character.}

\item{\code{mat}}{A matrix.}

\item{\code{dir}}{Character.}

\item{\code{rhs}}{Numeric.}

\item{\code{nc}}{Numeric.}
}}

