% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\docType{class}
\name{constraints-class}
\alias{constraints-class}
\title{An S4 class to represent a set of constraints}
\description{
An S4 class to represent a set of constraints.
}
\section{Slots}{

\describe{
\item{\code{slope}}{Numeric. A slope parameter value.}

\item{\code{difficulty}}{Numeric. A difficulty parameter value.}
}}

