% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{constraints_0}
\alias{constraints_0}
\title{constraints_0}
\format{A data.frame.}
\description{
A simple constraint set (1 constraint) to use with \code{item_params_A, item_params_B, item_params_C, item_attribs_A, item_attribs_B, item_attribs_C}.
}
\keyword{datasets}
