% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{constraints_1}
\alias{constraints_1}
\title{constraints_1}
\format{A data.frame.}
\description{
A more complex constraint set (14 constraints) to use with \code{item_params_A, item_params_B, item_params_C, item_attribs_A, item_attribs_B, item_attribs_C}.
}
\keyword{datasets}
