% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset.fatigue}
\alias{dataset.fatigue}
\alias{item.params.fatigue.raw}
\alias{item.attrib.fatigue.raw}
\alias{item.content.fatigue.raw}
\alias{constraints.fatigue.raw}
\alias{resp.fatigue.raw}
\alias{itempool.fatigue}
\alias{itemattrib.fatigue}
\alias{constraints.fatigue}
\title{Fatigue dataset}
\description{
Item-based example pool with item contents (95 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool.fatigue} An \code{\linkS4class{item.pool}} object.
  \item \code{itemattrib.fatigue} A data frame containing item attributes.
  \item \code{constraints.fatigue} A list containing 111 constraints.
}

Also, the following datasets are intended for illustrating expected data structures. See examples below.
\itemize{
  \item \code{item.params.fatigue.raw} Item parameters.
  \item \code{item.attrib.fatigue.raw} Item attributes.
  \item \code{constraints.fatigue.raw} Constraints.
  \item \code{item.content.fatigue.raw} Item contents.
  \item \code{resp.fatigue.raw} Raw response data.
}
}
\examples{
\dontrun{
write.csv(item.params.fatigue.raw, "item.params.fatigue.raw.csv", row.names = F)
itempool.fatigue = LoadItemPool("item.params.fatigue.raw.csv")

write.csv(item.attrib.fatigue, "item.attrib.fatigue.raw", row.names = F)
itemattrib.fatigue = LoadItemAttrib("item.attrib.fatigue.raw", itempool.fatigue)

write.csv(constraints.fatigue.raw, "constraints.fatigue.raw.csv", row.names = F)
constraints.fatigue = LoadConstraints("constraints.fatigue.raw.csv",
          itempool.fatigue, itemattrib.fatigue)

write.csv(item.content.fatigue.raw, "item.content.fatigue.raw.csv", row.names = F)
write.table(resp.fatigue.raw, "resp.fatigue.raw.csv", row.names = F, col.names = F, sep = ",")
}
}
\keyword{datasets}
