% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset.reading}
\alias{dataset.reading}
\alias{item.params.reading.raw}
\alias{item.attrib.reading.raw}
\alias{stim.attrib.reading.raw}
\alias{constraints.reading.raw}
\alias{itempool.reading}
\alias{itemattrib.reading}
\alias{stimattrib.reading}
\alias{constraints.reading}
\title{Reading dataset}
\description{
Stimulus-based example item pool (303 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool.reading} An \code{\linkS4class{item.pool}} object.
  \item \code{itemattrib.reading} A data frame containing item attributes.
  \item \code{stimattrib.reading} A data frame containing stimulus attributes.
  \item \code{constraints.reading} A list containing 18 constraints.
}

Also, the following datasets are intended for illustrating expected data structures. See examples below.
\itemize{
  \item \code{item.params.reading.raw} Item parameters.
  \item \code{item.attrib.reading.raw} Item attributes.
  \item \code{stim.attrib.reading.raw} Item attributes.
  \item \code{constraints.reading.raw} Constraints.
}
}
\examples{
\dontrun{
write.csv(item.params.reading.raw, "item.params.reading.raw.csv", row.names = F)
itempool.reading = LoadItemPool("item.params.reading.raw.csv")

write.csv(item.attrib.reading.raw, "item.attrib.reading.raw.csv", row.names = F)
itemattrib.reading = LoadItemAttrib("item.attrib.reading.raw.csv", itempool.reading)

write.csv(stim.attrib.reading.raw, "stim.attrib.reading.raw.csv", row.names = F)
stimattrib.reading = LoadStAttrib("stim.attrib.reading.raw.csv", itemattrib.reading)

write.csv(constraints.reading.raw, "constraints.reading.raw.csv", row.names = F)
constraints.reading = LoadConstraints("constraints.reading.raw.csv",
    itempool.reading, itemattrib.reading, stimattrib.reading)
}
}
\keyword{datasets}
