% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset.science}
\alias{dataset.science}
\alias{item.params.science.raw}
\alias{item.attrib.science.raw}
\alias{constraints.science.raw}
\alias{itempool.science}
\alias{itemattrib.science}
\alias{constraints.science}
\title{Science dataset}
\description{
Item-based example item pool (1000 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool.science} An \code{\linkS4class{item.pool}} object.
  \item \code{itemattrib.science} A data frame containing item attributes.
  \item \code{constraints.science} A list containing 36 constraints.
}

Also, the following datasets are intended for illustrating expected data structures. See examples below.
\itemize{
  \item \code{item.params.science.raw} Item parameters.
  \item \code{item.attrib.science.raw} Item attributes.
  \item \code{constraints.science.raw} Constraints.
}
}
\examples{
\dontrun{
write.csv(item.params.science.raw, "item.params.science.raw.csv", row.names = F)
itempool.science = LoadItemPool("item.params.science.raw.csv")

write.csv(item.attrib.science.raw, "item.attrib.science.raw.csv", row.names = F)
itemattrib.science = LoadItemAttrib("item.attrib.science.raw.csv", itempool.science)

write.csv(constraints.science.raw, "constraints.science.raw.csv", row.names = F)
constraints.science = LoadConstraints("constraints.science.raw.csv",
    itempool.science, itemattrib.science)
}
}
\keyword{datasets}
