% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{item_params_C}
\alias{item_params_C}
\alias{item_se_C}
\alias{item_attribs_C}
\title{item_params_C}
\description{
Item-based pool C (n = 300).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{item_params_C} Item parameters.
  \item \code{item_se_C} Standard errors of the item parameters.
  \item \code{item_attribs_C} Item attributes.
  \item \code{constraints_0} A simple constraint set (1 constraint).
  \item \code{constraints_1} A complex constraint set (14 constraints).
}
}
\keyword{datasets}
