% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{item_params_fatigue}
\alias{item_params_fatigue}
\alias{item_attribs_fatigue}
\alias{item_contents_fatigue}
\alias{constraints_fatigue}
\alias{raw_fatigue}
\title{item_params_fatigue}
\description{
Item-based pool with item contents.
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{item_params_fatigue} Item parameters.
  \item \code{item_attribs_fatigue} Item attributes.
  \item \code{item_contents_fatigue} Item contents.
  \item \code{constraints_fatigue} A constraint set.
  \item \code{raw_fatigue} Raw response data.
}
}
\keyword{datasets}
