% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_mat}
\alias{dataset_mat}
\alias{itempool_mat_data}
\alias{itempool_mat}
\title{MAT dataset}
\description{
Item-based example item pool (180 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool_mat} an \code{\linkS4class{item_pool}} object containing 180 items.
}

Also, the following objects are intended for illustrating expected data structures.
\itemize{
  \item \code{itempool_mat_data} a \code{\link{data.frame}} containing item parameters.
}
}
\examples{
itempool_mat   <- loadItemPool(itempool_mat_data)

}
\keyword{datasets}
