% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_reading}
\alias{dataset_reading}
\alias{itempool_reading_raw}
\alias{itemattrib_reading_raw}
\alias{stimattrib_reading_raw}
\alias{constraints_reading_raw}
\alias{itempool_reading}
\alias{itemattrib_reading}
\alias{stimattrib_reading}
\alias{constraints_reading}
\title{Reading dataset}
\description{
Stimulus-based example item pool (303 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool_reading} An \code{\linkS4class{item.pool}} object.
  \item \code{itemattrib_reading} A data frame containing item attributes.
  \item \code{stimattrib_reading} A data frame containing stimulus attributes.
  \item \code{constraints_reading} A list containing 18 constraints.
}

Also, the following datasets are intended for illustrating expected data structures. See examples below.
\itemize{
  \item \code{itempool_reading_raw} Item parameters.
  \item \code{itemattrib_reading_raw} Item attributes.
  \item \code{stimattrib_reading_raw} Item attributes.
  \item \code{constraints_reading_raw} Constraints.
}
}
\examples{
\dontrun{
write.csv(itempool_reading_raw, "itempool_reading.csv", row.names = FALSE)
itempool_reading <- loadItemPool("itempool_reading.csv")

write.csv(itemattrib_reading_raw, "itemattrib_reading.csv", row.names = FALSE)
itemattrib_reading <- loadItemAttrib("itemattrib_reading.csv", itempool_reading)

write.csv(stimattrib_reading_raw, "stimattrib_reading.csv", row.names = FALSE)
stimattrib_reading <- loadStAttrib("stimattrib_reading.csv", itemattrib_reading)

write.csv(constraints_reading_raw, "constraints_reading.csv", row.names = FALSE)
constraints_reading <- loadConstraints(
  "constraints_reading.csv",
  itempool_reading, itemattrib_reading, stimattrib_reading
)
}
}
\keyword{datasets}
