% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{item_params_stimbased}
\alias{item_params_stimbased}
\alias{item_attribs_stimbased}
\alias{stim_attribs_stimbased}
\alias{constraints_stimbased}
\title{item_params_stimbased}
\description{
Stimulus-based pool.
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{item_params_stimbased} Item parameters.
  \item \code{item_attribs_stimbased} Item attributes.
  \item \code{stim_attribs_stimbased} Stimulus attributes.
  \item \code{constraints_stimbased} A constraint set.
}
}
\keyword{datasets}
