% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning_functions.r
\name{dbind}
\alias{dbind}
\title{(Internal) Bind matrices diagonally}
\usage{
dbind(...)
}
\arguments{
\item{...}{input matrices.}
}
\value{
\code{\link{dbind}} returns a matrix.
}
\description{
\code{\link{dbind}} is an internal function for binding matrices diagonally.
This is used for constructing constraint matrix-data in \code{\link{Split}}.
}
\keyword{internal}
