% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_functions.R
\name{determineCurrentThetaSegment}
\alias{determineCurrentThetaSegment}
\title{(Internal) Determine the current theta segment}
\usage{
determineCurrentThetaSegment(
  current_theta,
  position,
  exposure_control,
  simulation_constants
)
}
\arguments{
\item{current_theta}{a list containing data on the current theta estimate.}

\item{position}{the position within the current administration (i.e., test progress)}

\item{exposure_control}{the \code{exposure_control} slot of \code{\linkS4class{config_Shadow}}.
Used to read custom first segments if available.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{determineCurrentThetaSegment}} returns a segment index.
}
\description{
\code{\link{determineCurrentThetaSegment}} is an internal function for
determining the current theta segment.
}
\keyword{internal}
