% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_control_functions.R
\name{doExposureControl}
\alias{doExposureControl}
\title{(Internal) Perform exposure control}
\usage{
doExposureControl(
  exposure_record,
  segment_record,
  o,
  j,
  current_theta,
  eligibility_flag,
  simulation_constants
)
}
\arguments{
\item{exposure_record}{a named list containing exposure records.}

\item{segment_record}{a named list containing counts on theta segments.}

\item{o}{an \code{\linkS4class{output_Shadow}} object, containing data for a single examinee.}

\item{j}{the numeric index of the examinee.}

\item{current_theta}{a named list containing data on current theta estimate.}

\item{eligibility_flag}{a named list containing eligibility flags.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{doExposureControl}} returns an updated list for exposure control records.
}
\description{
\code{\link{doExposureControl}} is an internal function for performing exposure control.
This is a wrapper function for calling other low-level functions for respective exposure control methods.
}
\keyword{internal}
