% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{.onAttach}
\alias{.onAttach}
\title{(Internal) Package startup functions}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname, pkgname}{not used; only defined for compatibility.}
}
\value{
\code{\link{.onAttach}} does not return anything.
}
\description{
\code{\link{.onAttach}} is an internal function called when the package is first loaded.
}
\keyword{internal}
