% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{eap}
\alias{eap}
\alias{eap,item.pool-method}
\title{eap}
\usage{
eap(object, theta, prior, resp, select = NULL)

\S4method{eap}{item.pool}(object, theta, prior, resp, select = NULL)
}
\arguments{
\item{object}{A nice parameter}

\item{theta}{A nice parameter}

\item{prior}{A nice parameter}

\item{resp}{A nice parameter}

\item{select}{A nice parameter}
}
\description{
Generate expected a posteriori estimates of theta
}
