% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\docType{methods}
\name{EAP}
\alias{EAP}
\alias{EAP,test-method}
\alias{EAP,test.cluster-method}
\title{Generate expected a posteriori estimates of theta}
\usage{
EAP(object, prior, select = NULL, reset.prior = FALSE)

\S4method{EAP}{test}(object, prior, select = NULL, reset.prior = FALSE)

\S4method{EAP}{test.cluster}(object, prior, select = NULL,
  reset.prior = FALSE)
}
\arguments{
\item{object}{A \code{\linkS4class{test}} or a \code{\linkS4class{test.cluster}} object.}

\item{prior}{A prior distribution, a numeric vector for a common prior or a matrix for individualized priors.}

\item{select}{A vector of indices identifying the items to subset.}

\item{reset.prior}{Set \code{TRUE} to reset the prior distribution for each test when object is of class \code{\linkS4class{test.cluster}}.}
}
\description{
Generate expected a posteriori estimates of theta.
}
