% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{==.item.pool}
\alias{==.item.pool}
\title{==.item.pool}
\usage{
\method{==}{item.pool}(pool1, pool2)
}
\arguments{
\item{pool1}{An \code{\linkS4class{item.pool}} object}

\item{pool2}{An \code{\linkS4class{item.pool}} object}
}
\description{
Test the equality of two item.pool objects
}
