% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{==.pool.cluster}
\alias{==.pool.cluster}
\title{==.pool.cluster}
\usage{
\method{==}{pool.cluster}(pool.cluster1, pool.cluster2)
}
\arguments{
\item{pool.cluster1}{A nice parameter}

\item{pool.cluster2}{A nice parameter}
}
\description{
==.pool.cluster
}
