% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cpp_theta_documents.r
\name{estimate_theta_map}
\alias{estimate_theta_map}
\title{(C++) Calculate a theta estimate using MAP (maximum a posteriori) method}
\usage{
estimate_theta_map(
  ipar,
  item_model,
  nd,
  n_pars,
  response,
  start_theta,
  prior_sigma,
  max_iteration = 30L,
  convergence_criterion = 0.001,
  use_Fisher = TRUE
)
}
\arguments{
\item{ipar}{a matrix containing item parameters.}

\item{item_model}{a vector indicating item models of each item, using \itemize{
  \item{\code{102}}: M2PL model
  \item{\code{103}}: M3PL model
  \item{\code{105}}: MGPC model
  \item{\code{106}}: MGR model
}}

\item{nd}{the number of dimensions for items.}

\item{n_pars}{a vector containing the number of item parameters for each item.}

\item{response}{responses on each item.}

\item{start_theta}{a vector containing the starting estimate.}

\item{prior_sigma}{a matrix for the prior distribution.}

\item{max_iteration}{the maximum number of iterations.}

\item{convergence_criterion}{the convergence criterion.}

\item{use_Fisher}{use Fisher method instead of Newton-Raphson method for optimization.}
}
\description{
\code{estimate_theta_map()} is a function for calculating a theta estimate using MAP (maximum a posteriori) method.
}
\details{
\code{estimate_theta_map()} is designed for multiple items.
}
