% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{evaluateSatisfaction}
\alias{evaluateSatisfaction}
\title{Evaluate a test for constraint satisfaction}
\usage{
evaluateSatisfaction(constraints, item_idx)
}
\arguments{
\item{constraints}{a \code{\linkS4class{constraints}} object.}

\item{item_idx}{item indices of a test.}
}
\value{
\code{\link{evaluateSatisfaction}} returns a vector of \code{TRUE} and \code{FALSE} values.
}
\description{
\code{\link{evaluateSatisfaction}} is a helper function for evaluating a test for its constraint satisfaction.
}
\examples{
set.seed(1)
item_idx <- sample(1:1000, 30)

evaluateSatisfaction(constraints_science, item_idx)

}
