% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\docType{methods}
\name{extract-methods}
\alias{extract-methods}
\alias{[,test,ANY,ANY,ANY-method}
\title{Extract}
\usage{
\S4method{[}{test,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{x}

\item{i}{i}

\item{j}{j}

\item{...}{...}

\item{drop}{drop}
}
\description{
Extract
}
