% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{generateDensityFromPriorPar}
\alias{generateDensityFromPriorPar}
\title{(Internal) Convert prior parameters to distribution densities}
\usage{
generateDensityFromPriorPar(config_theta, theta_q)
}
\arguments{
\item{config_theta}{a list containing theta estimation configurations.}

\item{theta_q}{a list containing all quadrature points.}
}
\value{
\code{\link{parsePriorParameters}} returns a named list containing:
\describe{
  \item{\code{prior_dist}}{the type of distribution. One of \code{NORMAL, UNIFORM, RAW}.}
  \item{\code{prior_par}}{an examinee-wise list containing prior parameters.}
}
}
\description{
\code{\link{generateDensityFromPriorPar}} is an internal function for converting prior parameters to distribution densities.
}
\keyword{internal}
