% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{generateItemParameterSample}
\alias{generateItemParameterSample}
\title{(Internal) Generate item parameter samples for Bayesian purposes}
\usage{
generateItemParameterSample(config, item_pool, bayesian_constants)
}
\arguments{
\item{config}{a \code{\linkS4class{config_Shadow}} object. Use \code{\link{createShadowTestConfig}} for this.}

\item{item_pool}{an \code{\linkS4class{item_pool}} object. Use \code{\link{loadItemPool}} for this.}

\item{bayesian_constants}{a named list containing Bayesian constants.}
}
\value{
If either the interim or the final theta estimation method is \code{FB},
\code{\link{generateItemParameterSample}} returns a length-\emph{ni} list of item parameter matrices, with each matrix having \code{n_sample} rows.
Otherwise, the function returns NULL.
}
\description{
\code{\link{generateItemParameterSample}} is a function for generating item parameter samples.
Used for the FB (full-Bayesian) estimation method.
}
\keyword{internal}
