% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{generateSampleFromPriorPar}
\alias{generateSampleFromPriorPar}
\title{(Internal) Generate random theta samples from prior distribution}
\usage{
generateSampleFromPriorPar(config_theta, j, bayesian_constants)
}
\arguments{
\item{config_theta}{a named list containing theta estimation configurations.}

\item{j}{the examinee index to perform the sample generation.}

\item{bayesian_constants}{a named list containing Bayesian constants.}
}
\value{
\code{\link{generateSampleFromPriorPar}} returns \code{n_sample} theta samples.
}
\description{
\code{\link{generateSampleFromPriorPar}} is an internal function for
generating random theta samples from prior distribution.
}
\keyword{internal}
