% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{getConstants}
\alias{getConstants}
\title{(Internal) Parse constants for adaptive test assembly simulation}
\usage{
getConstants(constraints, config, arg_data, true_theta, max_info)
}
\arguments{
\item{constraints}{a \code{\linkS4class{constraints}} object representing test specifications. Use \code{\link{loadConstraints}} for this.}

\item{config}{a \code{\linkS4class{config_Shadow}} object. Use \code{\link{createShadowTestConfig}} for this.}

\item{arg_data}{the \code{data} argument from \code{\link{Shadow}}.}

\item{true_theta}{the \code{true_theta} argument from \code{\link{Shadow}}.}

\item{max_info}{the maximum possible information attainable by any single item in the pool.
This is used to determine the M value if M is not supplied.}
}
\value{
\code{\link{getConstants}} returns a named list containing constants.
}
\description{
\code{\link{getConstants}} is an internal function for
parsing constants for adaptive test assembly simulation.
}
\keyword{internal}
