% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning_functions.r
\name{getDecisionVariablesOfItemForMultipool}
\alias{getDecisionVariablesOfItemForMultipool}
\title{(Internal) Make decision variables for selecting an item}
\usage{
getDecisionVariablesOfItemForMultipool(item_idx, nv_per_bin, n_bins)
}
\arguments{
\item{item_idx}{the target item index.}

\item{nv_per_bin}{the number of decision variables in the pool if this was a regular ATA.}

\item{n_bins}{the number of pools being modeled.}
}
\value{
\code{\link{getDecisionVariablesOfItemForMultipool}} returns a vector of indices.
}
\description{
\code{\link{getDecisionVariablesOfItemForMultipool}} is an internal function for
making decision variables for selecting a single item, for the purpose of multiple-pool assembly.
}
\keyword{internal}
