% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning_functions.r
\name{getDecisionVariablesOfPoolForMultipool}
\alias{getDecisionVariablesOfPoolForMultipool}
\title{(Internal) Make decision variables for selecting a pool}
\usage{
getDecisionVariablesOfPoolForMultipool(pool_idx, ni_per_bin, nv_per_bin)
}
\arguments{
\item{pool_idx}{the target pool index.}

\item{ni_per_bin}{the number of items in the pool.}

\item{nv_per_bin}{the number of decision variables in the pool if this was a regular ATA.}
}
\value{
\code{\link{getDecisionVariablesOfPoolForMultipool}} returns a vector of indices.
}
\description{
\code{\link{getDecisionVariablesOfPoolForMultipool}} is an internal function for
making decision variables for selecting a single pool, for the purpose of multiple-pool assembly.
}
\keyword{internal}
