% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdata_functions.R
\name{getIndexOfExcludedEntry}
\alias{getIndexOfExcludedEntry}
\title{(Internal) Convert item IDs to item indices for exclusion}
\usage{
getIndexOfExcludedEntry(exclude, constraints)
}
\arguments{
\item{exclude}{a length-\emph{nj} list, where each entry is a named list containing:
\itemize{
  \item{\code{i} item IDs to exclude for this examinee}
  \item{\code{s} item set IDs to exclude for this examinee}
}}

\item{constraints}{a \code{\linkS4class{constraints}} object representing test specifications. Use \code{\link{loadConstraints}} for this.}
}
\value{
\code{\link{getIndexOfExcludedEntry}} returns a length-\emph{nj} list in the same structure,
with its contents converted to item indices.
}
\description{
\code{\link{getIndexOfExcludedEntry}} is an internal function for
converting item IDs to item indices for exclusion.
}
\details{
Notation:
\itemize{
  \item{\emph{nj} the number of examinees}
}
}
\keyword{internal}
