% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{getInfoFixedTheta}
\alias{getInfoFixedTheta}
\title{(Internal) Precalculate item information for fixed-theta item selection methods}
\usage{
getInfoFixedTheta(item_selection, simulation_constants, item_pool, model)
}
\arguments{
\item{item_selection}{a list containing item selection config.}

\item{simulation_constants}{a named list containing simulation constants.}

\item{item_pool}{an item pool object.}

\item{model}{sanitized item model codes to be fed to C++ functions.}
}
\value{
\code{\link{getInfoFixedTheta}} returns a named list containing item information values
at designated thetas for each simulee.
}
\description{
\code{\link{getInfoFixedTheta}} is an internal function for
calculating item information for fixed-theta item selection methods.
This is done once at the start of the simulation and cached for speed gain.
}
\keyword{internal}
