% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdata_functions.R
\name{getInfoOfExcludedEntry}
\alias{getInfoOfExcludedEntry}
\title{(Internal) Apply information penalty on items to be excluded}
\usage{
getInfoOfExcludedEntry(info, exclude_index, simulation_constants)
}
\arguments{
\item{info}{a one-row matrix containing information values on each item.}

\item{exclude_index}{a named list containing item/set indices that need to be excluded.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{getInfoOfExcludedEntry}} returns
an updated one-row matrix containing information values.
}
\description{
(Internal) Apply information penalty on items to be excluded
}
\keyword{internal}
