% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{getLBUBInConstraintData}
\alias{getLBUBInConstraintData}
\title{(Internal) Parse item/stimulus lower/upper bounds from constraints data}
\usage{
getLBUBInConstraintData(
  constants,
  constraints,
  item_constraints,
  stim_constraints
)
}
\arguments{
\item{constants}{a names list containing constants.}

\item{constraints}{a \code{\link{data.frame}} containing constraints data.}

\item{item_constraints}{row numbers indicating which are item-level constraints.}

\item{stim_constraints}{row numbers indicating which are stimulus-level constraints.}
}
\value{
\code{\link{getLBUBInConstraintData}} returns an updated list of constants.
}
\description{
\code{\link{getLBUBInConstraintData}} is an internal function for
parsing maximum item/stimulus lower/upper bounds from constraints data.
}
\keyword{internal}
