% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_control_functions.R
\name{getSegmentOf}
\alias{getSegmentOf}
\title{(Internal) Calculate theta segment of a given examinee}
\usage{
getSegmentOf(x, simulation_constants)
}
\arguments{
\item{x}{an \code{\linkS4class{output_Shadow}} object.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{getSegmentOf}} returns a list containing:
\itemize{
  \item{\code{final_theta_est} the theta segment based on final theta estimate.}
  \item{\code{true_theta} the theta segment based on true theta.}
  \item{\code{visited} segments the examinee visited.}
}
}
\description{
\code{\link{getSegmentOf}} is an internal function for
calculating which theta segment an examinee belongs to.
}
\keyword{internal}
