% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{getSolution}
\alias{getSolution}
\alias{getSolution,list-method}
\alias{getSolution,output_Static-method}
\title{Print solution items}
\usage{
getSolution(object, examinee = NA, position = NA, index_only = TRUE)

\S4method{getSolution}{list}(object, examinee = NA, position = NA, index_only = TRUE)

\S4method{getSolution}{output_Static}(object, examinee = NA, position = NA, index_only = TRUE)
}
\arguments{
\item{object}{an \code{\linkS4class{output_Static}} object or an \code{\linkS4class{output_Shadow}} object.}

\item{examinee}{(optional) the examinee index to display the solution. Used when the 'object' argument is an \code{\linkS4class{output_Shadow}} object.}

\item{position}{(optional) if supplied, display the item attributes of the assembled test at that item position. If not supplied, display the item attributes of the administered items. Used when the 'object' argument is an \code{\linkS4class{output_Shadow}} object.}

\item{index_only}{if \code{TRUE}, only print item indices. if \code{FALSE}, print all item attributes. (default = {TRUE})}
}
\value{
Item attributes of solution items.
}
\description{
Print solution items
}
