% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_functions.R
\name{getThetaQuadrature}
\alias{getThetaQuadrature}
\title{Get theta quadrature points}
\usage{
getThetaQuadrature(
  theta_center,
  theta_width,
  quadrature_unit,
  use_ellipse = FALSE
)
}
\arguments{
\item{theta_center}{the theta value to use as center.}

\item{theta_width}{the width for each dimension.}

\item{quadrature_unit}{the interval between quadrature points in each dimension.}

\item{use_ellipse}{if \code{TRUE}, return an ellipse area. If \code{FALSE}, return a rectangle area. (default = \code{FALSE})}
}
\description{
\code{getThetaQuadrature()} is a helper function for getting quadrature points for numerical integration.
}
\examples{
theta_q <- getThetaQuadrature(0, 3, .1)
plot(theta_q)

theta_q <- getThetaQuadrature(c(1, 2), 1, .1)
plot(theta_q)

theta_q <- getThetaQuadrature(c(1, 2), c(1, 2), .1)
plot(theta_q)

theta_q <- getThetaQuadrature(c(1, 2), 1, .1, TRUE)
plot(theta_q)

theta_q <- getThetaQuadrature(c(1, 2), c(1, 2), .1, TRUE)
plot(theta_q)

}
