% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_functions.R
\name{incrementAlpha}
\alias{incrementAlpha}
\title{(Internal) Increment exposure record variable: alpha}
\usage{
incrementAlpha(
  exposure_record,
  segments_to_apply,
  segment_prob,
  x,
  simulation_constants
)
}
\arguments{
\item{exposure_record}{a named list containing exposure records.}

\item{segments_to_apply}{which segments to apply the incrementing.}

\item{segment_prob}{the amount of increment.}

\item{x}{an \code{\linkS4class{output_Shadow}} object, containing data for a single examinee.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{incrementAlpha}} returns an updated exposure record.
}
\description{
\code{\link{incrementAlpha}} is an internal function for incrementing an exposure record variable.
Specifically, the \code{a_ijk} is incremented by one for administered items/sets.
}
\keyword{internal}
