% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_functions.R
\name{incrementN}
\alias{incrementN}
\title{(Internal) Increment exposure record variable: n}
\usage{
incrementN(
  exposure_record,
  segments_to_apply,
  segment_prob,
  simulation_constants
)
}
\arguments{
\item{exposure_record}{a named list containing exposure records.}

\item{segments_to_apply}{which segments to apply the incrementing.}

\item{segment_prob}{the amount of increment.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{incrementN}} returns an updated exposure record.
}
\description{
\code{\link{incrementPhi}} is an internal function for incrementing an exposure record variable.
Specifically, the \code{n_k} is incremented by one.
}
\keyword{internal}
