% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_3pl}
\alias{info_3pl}
\title{Calculate the Fisher information according to the 3pl model}
\usage{
info_3pl(x, a, b, c)
}
\arguments{
\item{x}{a length-one numeric vector for a theta value}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a length-one numeric vector for the difficulty parameter}

\item{c}{a length-one numeric vector for the guessing parameter}
}
\description{
Calculate the Fisher information according to the 3pl model
}
\references{
{
  \insertRef{birnbaum_latent_1968}{Shadow}
}
}
