% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_gpc}
\alias{info_gpc}
\title{Calculate the Fisher information according to the generalized partial credit model}
\usage{
info_gpc(x, a, b)
}
\arguments{
\item{x}{a length-one numeric vector for a theta value}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a numeric vector for the threshold parameters}
}
\description{
Calculate the Fisher information according to the generalized partial credit model
}
\references{
{
  \insertRef{muraki_generalized_1992}{Shadow}
}
}
