% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cpp_documents.R
\name{info_1pl}
\alias{info_1pl}
\alias{info_2pl}
\alias{info_3pl}
\alias{info_pc}
\alias{info_gpc}
\alias{info_gr}
\alias{array_info_1pl}
\alias{array_info_2pl}
\alias{array_info_3pl}
\alias{array_info_pc}
\alias{array_info_gpc}
\alias{array_info_gr}
\alias{info_item}
\title{Calculate Fisher information (single item)}
\usage{
info_1pl(x, b)

info_2pl(x, a, b)

info_3pl(x, a, b, c)

info_pc(x, b)

info_gpc(x, a, b)

info_gr(x, a, b)

array_info_1pl(x, b)

array_info_2pl(x, a, b)

array_info_3pl(x, a, b, c)

array_info_pc(x, b)

array_info_gpc(x, a, b)

array_info_gr(x, a, b)
}
\arguments{
\item{x}{the theta value. This must be a column vector in matrix form for \code{array_info_*} functions.}

\item{b}{the *b*-parameter.}

\item{a}{the *a*-parameter.}

\item{c}{the *c*-parameter.}
}
\description{
\code{info_*} and \code{array_info_*} are functions to calculate Fisher information.
}
\details{
\code{info_*} functions accept a single theta value, and \code{array_info_*} functions accept multiple theta values.

Currently supports unidimensional models.

\itemize{
  \item{\code{info_1pl}, \code{array_info_1pl}}: 1PL models
  \item{\code{info_2pl}, \code{array_info_2pl}}: 2PL models
  \item{\code{info_3pl}, \code{array_info_3pl}}: 3PL models
  \item{\code{info_pc}, \code{array_info_pc}}: PC (partial credit) models
  \item{\code{info_gpc}, \code{array_info_gpc}}: GPC (generalized partial credit) models
  \item{\code{info_gr}, \code{array_info_gr}}: GR (graded response) models
}
}
\examples{
x <- 0.5

info_1pl(x, 1)
info_2pl(x, 1, 2)
info_3pl(x, 1, 2, 0.25)
info_pc(x, c(0, 1))
info_gpc(x, 2, c(0, 1))
info_gr(x, 2, c(0, 2))

x <- matrix(seq(0.1, 0.5, 0.1)) # column vector in matrix form

array_info_1pl(x, 1)
array_info_2pl(x, 1, 2)
array_info_3pl(x, 1, 2, 0.25)
array_info_pc(x, c(0, 1))
array_info_gpc(x, 2, c(0, 1))
array_info_gr(x, 2, c(0, 2))

}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}

{
  \insertRef{lord_theory_1952}{TestDesign}
}

{
  \insertRef{birnbaum_efficient_1957}{TestDesign}
}

{
  \insertRef{birnbaum_estimation_1958}{TestDesign}
}

{
  \insertRef{birnbaum_further_1958}{TestDesign}
}

{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}

{
  \insertRef{masters_rasch_1982}{TestDesign}
}

{
  \insertRef{andrich_rating_1978}{TestDesign}
}

{
  \insertRef{muraki_generalized_1992}{TestDesign}
}

{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
