% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_pc}
\alias{info_pc}
\title{Calculate Fisher information at a single theta (PC)}
\usage{
info_pc(x, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the Fisher information at a theta value according to the partial credit model.
}
\references{
{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}
}
