% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{initializeCompletedGroupingsRecord}
\alias{initializeCompletedGroupingsRecord}
\title{(Internal) Initialize groupings record}
\usage{
initializeCompletedGroupingsRecord()
}
\value{
\code{\link{initializeCompletedGroupingsRecord}} returns a groupings record.
}
\description{
\code{\link{initializeCompletedGroupingsRecord}} is an internal function for
creating a new groupings record. Used for keeping track of completed groups such as item sets.
}
\keyword{internal}
