% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_control_functions.R
\name{initializeDiagnosticExposureRecord}
\alias{initializeDiagnosticExposureRecord}
\title{(Internal) Initialize diagnostic exposure record}
\usage{
initializeDiagnosticExposureRecord(simulation_constants)
}
\arguments{
\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{initializeDiagnosticExposureRecord}} returns a list containing diagnostic exposure record.
}
\description{
\code{\link{initializeDiagnosticExposureRecord}} is an internal function
for creating a new diagnostic exposure record.
}
\keyword{internal}
