% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_control_functions.R
\name{initializeExposureRecord}
\alias{initializeExposureRecord}
\title{(Internal) Initialize exposure record}
\usage{
initializeExposureRecord(exposure_control, simulation_constants)
}
\arguments{
\item{exposure_control}{the \code{exposure_control} slot from a \code{\linkS4class{config_Shadow}} object.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{initializeExposureRecord}} returns a list containing exposure record.
}
\description{
\code{\link{initializeExposureRecord}} is an internal function
for creating a new exposure record.
}
\keyword{internal}
