% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_control_functions.R
\name{initializeUsageMatrix}
\alias{initializeUsageMatrix}
\title{(Internal) Initialize item usage matrix}
\usage{
initializeUsageMatrix(simulation_constants)
}
\arguments{
\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{initializeUsageMatrix}} returns a new item usage matrix.
}
\description{
\code{\link{initializeUsageMatrix}} is an internal function for
creating a new item usage matrix. An item usage matrix is a
(\emph{nj}, \emph{nv}) matrix with \code{TRUE} or \code{FALSE} values.
The rows represent examinees, and the columns represent decision variables.
}
\keyword{internal}
