% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{iparPosteriorSample_foreach}
\alias{iparPosteriorSample_foreach}
\title{iparPosteriorSample_foreach}
\usage{
iparPosteriorSample_foreach(pool, nSample = 500)
}
\arguments{
\item{pool}{A nice parameter}

\item{nSample}{A nice parameter}
}
\description{
iparPosteriorSample_foreach
}
