% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{iparPosteriorSample_lapply}
\alias{iparPosteriorSample_lapply}
\title{iparPosteriorSample_lapply}
\usage{
iparPosteriorSample_lapply(pool, nSample = 500)
}
\arguments{
\item{pool}{A nice parameter}

\item{nSample}{A nice parameter}
}
\description{
iparPosteriorSample_lapply
}
