% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{iparPosteriorSample_which}
\alias{iparPosteriorSample_which}
\title{iparPosteriorSample_which}
\usage{
iparPosteriorSample_which(pool, nSample = 500)
}
\arguments{
\item{pool}{A nice parameter}

\item{nSample}{A nice parameter}
}
\description{
iparPosteriorSample_which
}
