% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_functions.R
\name{isCustomFirstSegmentAvailable}
\alias{isCustomFirstSegmentAvailable}
\title{(Internal) Check if customized first segments are available}
\usage{
isCustomFirstSegmentAvailable(available, n_values, position)
}
\arguments{
\item{available}{\code{TRUE} or \code{FALSE}.
This indicates whether segment values have been supplied in the config.}

\item{n_values}{the number of supplied values.}

\item{position}{the position within the current administration (i.e., test progress)}
}
\value{
\code{\link{isCustomFirstSegmentAvailable}} returns \code{TRUE} or \code{FALSE}.
}
\description{
\code{\link{isCustomFirstSegmentAvailable}} is an internal function for
checking if customized first segments are available.
}
\keyword{internal}
