% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_functions.R
\name{isSolutionOptimal}
\alias{isSolutionOptimal}
\title{(Internal) Check whether solution is optimal}
\usage{
isSolutionOptimal(status, solver)
}
\arguments{
\item{status}{status code returned by the solver function.}

\item{solver}{solver name.}
}
\value{
\code{\link{isSolutionOptimal}} returns \code{TRUE} or \code{FALSE}.
}
\description{
\code{\link{isSolutionOptimal}} is an internal function for
checking whether a solution is optimal.
}
\keyword{internal}
