% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\docType{class}
\name{item-classes}
\alias{item-classes}
\alias{item}
\alias{item_1PL-class}
\alias{item_2PL-class}
\alias{item_3PL-class}
\alias{item_PC-class}
\alias{item_GPC-class}
\alias{item_GR-class}
\title{Item classes}
\description{
\itemize{
  \item{\code{\linkS4class{item_1PL}} class represents a 1PL item.}
  \item{\code{\linkS4class{item_2PL}} class represents a 2PL item.}
  \item{\code{\linkS4class{item_3PL}} class represents a 3PL item.}
  \item{\code{\linkS4class{item_PC}} class represents a partial credit item.}
  \item{\code{\linkS4class{item_GPC}} class represents a generalized partial credit item.}
  \item{\code{\linkS4class{item_GR}} class represents a graded response item.}
}
}
\section{Slots}{

\describe{
\item{\code{slope}}{a slope parameter value}

\item{\code{difficulty}}{a difficulty parameter value}

\item{\code{guessing}}{a guessing parameter value}

\item{\code{threshold}}{a vector of threshold parameter values}

\item{\code{category}}{a vector of category boundary values}

\item{\code{ncat}}{the number of response categories}
}}

\examples{
item_1 <- new("item_1PL", difficulty = 0.5)
item_2 <- new("item_2PL", slope = 1.0, difficulty = 0.5)
item_3 <- new("item_3PL", slope = 1.0, difficulty = 0.5, guessing = 0.2)
item_4 <- new("item_PC", threshold = c(-0.5, 0.5), ncat = 3)
item_5 <- new("item_GPC", slope = 1.0, threshold = c(-0.5, 0.0, 0.5), ncat = 4)
item_6 <- new("item_GR", slope = 1.0, category = c(-2.0, -1.0, 0, 1.0, 2.0), ncat = 6)

}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}

{
  \insertRef{lord_theory_1952}{TestDesign}

  \insertRef{birnbaum_efficient_1957}{TestDesign}

  \insertRef{birnbaum_estimation_1958}{TestDesign}

  \insertRef{birnbaum_further_1958}{TestDesign}
}

{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}

{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}

{
  \insertRef{muraki_generalized_1992}{TestDesign}
}

{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
