% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.1pl-class}
\alias{item.1pl-class}
\title{An S4 class to represent an item of the 1pl model}
\description{
An S4 class to represent an item of the 1pl model
}
\section{Slots}{

\describe{
\item{\code{difficulty}}{A numeric vector of length one to contain the difficulty parameter of a 1pl item}
}}

\examples{
item.1 = new("item.1pl", difficulty = 0.5)
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}
}
