% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.2pl-class}
\alias{item.2pl-class}
\title{An S4 class to represent an item of the 2pl model}
\description{
An S4 class to represent an item of the 2pl model
}
\section{Slots}{

\describe{
\item{\code{slope}}{A numeric vector of length one to contain the slope parameter of a 2pl item}

\item{\code{difficulty}}{A numeric vector of length one to contain the difficulty parameter of a 2pl item}
}}

\examples{
item.2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
}
\references{
{
  \insertRef{lord_theory_1952}{Shadow}
  
  \insertRef{birnbaum_efficient_1957}{Shadow}
  
  \insertRef{birnbaum_estimation_1958}{Shadow}
  
  \insertRef{birnbaum_further_1958}{Shadow}
}
}
