% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.3pl-class}
\alias{item.3pl-class}
\title{An S4 class to represent an item of the 3pl model}
\description{
An S4 class to represent an item of the 3pl model
}
\section{Slots}{

\describe{
\item{\code{slope}}{A numeric vector of length one to contain the slope parameter of a 3pl item}

\item{\code{difficulty}}{A numeric vector of length one to contain the difficulty parameter of a 3pl item}

\item{\code{guessing}}{A numeric vector of length one to contain the guessing parameter of a 3pl item}
}}

\examples{
item.3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
}
\references{
{
  \insertRef{birnbaum_latent_1968}{Shadow}
}
}
