% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.gpc-class}
\alias{item.gpc-class}
\title{An S4 class to represent an item of the generalized partial credit (gpc) model}
\description{
An S4 class to represent an item of the generalized partial credit (gpc) model
}
\section{Slots}{

\describe{
\item{\code{slope}}{A numeric vector of length one to contain the slope parameter of a gpc item}

\item{\code{threshold}}{A numeric vector to contain the threshold parameters of a gpc item}

\item{\code{ncat}}{A numeric vector of length one to indicate the number of response categories for a gpc item}
}}

\examples{
item.5 = new("item.gpc", slope = 1.0, threshold = c(-0.5, 0.0, 0.5), ncat = 4)
}
\references{
{
  \insertRef{muraki_generalized_1992}{Shadow}
}
}
