% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.gr-class}
\alias{item.gr-class}
\title{An S4 class to represent an item of the graded response (gr) model}
\description{
An S4 class to represent an item of the graded response (gr) model
}
\section{Slots}{

\describe{
\item{\code{slope}}{A numeric vector of length one to contain the slope parameter of a gr item}

\item{\code{category}}{A numeric vector to contain the threshold parameters of a gr item}

\item{\code{ncat}}{A numeric vector of length one to indicate the number of response categories for a gr item}
}}

\examples{
item.6 = new("item.gr", slope = 1.0, category = c(-2.0, -1.0, 0, 1.0, 2.0), ncat = 6)
}
\references{
{
  \insertRef{samejima_estimation_1969}{Shadow}
}
}
