% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.pc-class}
\alias{item.pc-class}
\title{An S4 class to represent an item of the partial credit (pc) model}
\description{
An S4 class to represent an item of the partial credit (pc) model
}
\section{Slots}{

\describe{
\item{\code{threshold}}{A numeric vector to contain the threshold parameters of a pc item}

\item{\code{ncat}}{A numeric vector of length one to indicate the number of response categories for a pc item}
}}

\examples{
item.4 = new("item.pc", threshold = c(-0.5, 0.5), ncat = 3)
}
\references{
{
  \insertRef{masters_rasch_1982}{Shadow}

  \insertRef{andrich_rating_1978}{Shadow}
}
}
