% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.pool-class}
\alias{item.pool-class}
\title{An S4 class to represent an item pool}
\description{
An S4 class to represent an item pool.
}
\section{Slots}{

\describe{
\item{\code{ni}}{Numeric. The number of items in the item pool.}

\item{\code{maxCat}}{Numeric. The maximum number of response categories across all items.}

\item{\code{index}}{Numeric. A vector of item indices.}

\item{\code{ID}}{Character. A vector of item IDs.}

\item{\code{model}}{Numeric. A vector of item model codes (1: item.1pl, 2: item.2pl, 3: item.3pl, 4: item.pc, 5: item.gpc, 6: item.gr).}

\item{\code{NCAT}}{Numeric. A vector of the number of response categories for each item.}

\item{\code{parms}}{A list of item parameters in the pool.}

\item{\code{ipar}}{A matrix of item parameters in the pool.}
}}

