% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{+.item.pool}
\alias{+.item.pool}
\alias{-.item.pool}
\alias{==.item.pool}
\title{Item pool operators}
\usage{
\method{+}{item.pool}(pool1, pool2)

\method{-}{item.pool}(pool1, pool2)

\method{==}{item.pool}(pool1, pool2)
}
\arguments{
\item{pool1}{An \code{\linkS4class{item.pool}} object.}

\item{pool2}{An \code{\linkS4class{item.pool}} object.}
}
\description{
\code{pool1 + pool2} combines two \code{\linkS4class{item.pool}} objects.

\code{pool1 - pool2} combines the second from the first. Thw two \code{\linkS4class{item.pool}} objects must overlap for this to be performed.

\code{pool1 == pool2} tests equality of the two item.pool objects.
}
