% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\docType{class}
\name{item_1PL-class}
\alias{item_1PL-class}
\title{An S4 class to represent a 1PL item}
\description{
An S4 class to represent a 1PL item.
}
\section{Slots}{

\describe{
\item{\code{difficulty}}{Numeric. A difficulty parameter value.}
}}

\examples{
item_1 <- new("item_1PL", difficulty = 0.5)
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}
}
