% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\docType{class}
\name{item_2PL-class}
\alias{item_2PL-class}
\title{An S4 class to represent a 2PL item}
\description{
An S4 class to represent a 2PL item.
}
\section{Slots}{

\describe{
\item{\code{slope}}{Numeric. A slope parameter value.}

\item{\code{difficulty}}{Numeric. A difficulty parameter value.}
}}

\examples{
item_2 <- new("item_2PL", slope = 1.0, difficulty = 0.5)
}
\references{
{
  \insertRef{lord_theory_1952}{TestDesign}

  \insertRef{birnbaum_efficient_1957}{TestDesign}

  \insertRef{birnbaum_estimation_1958}{TestDesign}

  \insertRef{birnbaum_further_1958}{TestDesign}
}
}
