% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\docType{class}
\name{item_3PL-class}
\alias{item_3PL-class}
\title{An S4 class to represent a 3PL item}
\description{
An S4 class to represent a 3PL item.
}
\section{Slots}{

\describe{
\item{\code{slope}}{Numeric. A slope parameter value.}

\item{\code{difficulty}}{Numeric. A difficulty parameter value.}

\item{\code{guessing}}{Numeric. A guessing parameter value.}
}}

\examples{
item_3 <- new("item_3PL", slope = 1.0, difficulty = 0.5, guessing = 0.2)
}
\references{
{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}
}
