% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_attrib_operators.R
\docType{methods}
\name{item_attrib-operators}
\alias{item_attrib-operators}
\alias{[,item_attrib,numeric-method}
\alias{[,item_attrib,numeric,ANY,ANY-method}
\alias{dim,item_attrib-method}
\alias{colnames,item_attrib-method}
\alias{rownames,item_attrib-method}
\alias{names,item_attrib-method}
\alias{as.data.frame,item_attrib-method}
\title{Basic functions for item attribute objects}
\usage{
\S4method{[}{item_attrib,numeric}(x, i, j, ..., drop = TRUE)

\S4method{dim}{item_attrib}(x)

\S4method{colnames}{item_attrib}(x)

\S4method{rownames}{item_attrib}(x)

\S4method{names}{item_attrib}(x)

\S4method{as.data.frame}{item_attrib}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an \code{\linkS4class{item_attrib}} object.}

\item{i, j}{indices to use in subsetting.}

\item{...}{not used, exists for compatibility.}

\item{drop}{not used, exists for compatibility.}

\item{row.names}{not used, exists for compatibility.}

\item{optional}{not used, exists for compatibility.}
}
\description{
Basic functions for item attribute objects
}
\examples{
x <- itemattrib_science
x[1:10]
dim(x)
ncol(x)
nrow(x)
colnames(x)
rownames(x)
names(x)
as.data.frame(x)

}
