% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\docType{class}
\name{item_GPC-class}
\alias{item_GPC-class}
\title{An S4 class to represent a generalized partial credit item}
\description{
An S4 class to represent a generalized partial credit item.
}
\section{Slots}{

\describe{
\item{\code{slope}}{Numeric. A slope parameter value.}

\item{\code{threshold}}{Numeric. A vector of threshold parameter values.}

\item{\code{ncat}}{Numeric. The number of response categories.}
}}

\examples{
item_5 <- new("item_GPC", slope = 1.0, threshold = c(-0.5, 0.0, 0.5), ncat = 4)
}
\references{
{
  \insertRef{muraki_generalized_1992}{TestDesign}
}
}
