% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\name{validateDifficulty}
\alias{validateDifficulty}
\alias{validateSlope}
\alias{validateGuessing}
\alias{validateNcat}
\alias{validateNthr}
\alias{validateCategory}
\alias{validateOrder}
\title{(Internal) Validate item objects}
\usage{
validateDifficulty(object)

validateSlope(object)

validateGuessing(object)

validateNcat(object)

validateNthr(object)

validateCategory(object)

validateOrder(object)
}
\arguments{
\item{object}{an item object.}
}
\value{
These functions return an error message string if the input object is not valid.
These do not return anything for valid objects.
}
\description{
These are internal functions for validating item objects for their item parameters.
}
\keyword{internal}
