% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\docType{class}
\name{item_pool-class}
\alias{item_pool-class}
\title{An S4 class to represent an item pool}
\description{
An S4 class to represent an item pool.
}
\section{Slots}{

\describe{
\item{\code{ni}}{Numeric. The number of items in the item pool.}

\item{\code{max_cat}}{Numeric. The maximum number of response categories across all items.}

\item{\code{index}}{Numeric. A vector of item indices.}

\item{\code{id}}{Character. A vector of item ids.}

\item{\code{model}}{Numeric. A vector of item model codes (1: item.1pl, 2: item_2PL, 3: item_3PL, 4: item_PC, 5: item_GPC, 6: item_GR).}

\item{\code{NCAT}}{Numeric. A vector of the number of response categories for each item.}

\item{\code{parms}}{A list of item parameters in the pool.}

\item{\code{ipar}}{A matrix of item parameters in the pool.}

\item{\code{se}}{A matrix representing standard errors of the item parameters.}

\item{\code{raw}}{A data.frame containing raw input data.}
}}

