% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\docType{class}
\name{item_pool_cluster-class}
\alias{item_pool_cluster-class}
\title{Class 'item_pool_cluster': an item pool}
\description{
\code{\linkS4class{item_pool_cluster}} is an S4 class to represent a group of item pools.
}
\section{Slots}{

\describe{
\item{\code{np}}{the number of item pools.}

\item{\code{pools}}{a list of \code{\linkS4class{item_pool}} objects.}

\item{\code{names}}{a vector containing item pool names.}
}}

