% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{loadConstraints}
\alias{loadConstraints}
\title{Load constraints}
\usage{
loadConstraints(file, pool, item_attrib, st_attrib = NULL)
}
\arguments{
\item{file}{Character. The name of the file containing specifications for constraints.}

\item{pool}{An \code{item_pool} object.}

\item{item_attrib}{A \code{data.frame} containing item attributes. Use \code{\link{loadItemAttrib}} for this.}

\item{st_attrib}{(Optional) A \code{data.frame} containing stimulus attributes. Use \code{\link{loadStAttrib}} for this.}
}
\value{
A list containing the parsed constraints, to be used in \code{\link{ATA}} and \code{\link{Shadow}}.
}
\description{
Read constraints from specified file.
}
\details{
Use \code{vignette("constraints")} for instructions on how to create a constraint set object.
}
\examples{
\donttest{
itempool_science <- loadItemPool("itempool_science.csv")
itemattrib_science <- loadItemAttrib("itemattrib_science.csv", itempool_science)
constraints_science <- loadConstraints("constraints_science.csv",
  itempool_science, itemattrib_science)
}
}
\seealso{
\link{dataset_science} for example usage.
}
