% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{loadItemPool}
\alias{loadItemPool}
\title{Load item paramaters}
\usage{
loadItemPool(ipar, ipar_se = NULL, file = NULL, se_file = NULL)
}
\arguments{
\item{ipar}{Item parameters. Can be a data.frame or the file path of a .csv file. The content should at least include columns 'ID' and 'MODEL'.}

\item{ipar_se}{(Optional) Standard errors. Can be a data.frame or file path of a .csv file.}

\item{file}{(Deprecated) Use 'ipar' above.}

\item{se_file}{(Deprecated) Use 'ipar_se' above.}
}
\value{
An \code{\linkS4class{item_pool}} object.
}
\description{
\code{\link{loadItemPool}} is a data loading function to create an \linkS4class{item_pool} class.
\code{\link{loadItemPool}} can read item parameters and standard errors from a data.frame or a .csv file.
}
\examples{
## Read from data.frame:
itempool_science <- loadItemPool(itempool_science_data)

## Read from file: write to tempdir() for illustration and clean afterwards
f <- file.path(tempdir(), "itempool_science.csv")
write.csv(itempool_science_data, f, row.names = FALSE)
itempool_science <- loadItemPool(f)
file.remove(f)

## TestDesign 1.1.0 - Deprecated arguments
\dontrun{
loadItemPool(ipar = "ipar.csv", ipar_se = "se.csv") # is equivalent to
loadItemPool(file = "ipar.csv", se_file = "se.csv") # pre 1.1.0
}

}
\seealso{
\link{dataset_science} for example usage.
}
