% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{loadStAttrib}
\alias{loadStAttrib}
\title{Load set/stimulus/passage attributes}
\usage{
loadStAttrib(file, item_attrib)
}
\arguments{
\item{file}{Character. The name of the file containing item attributes.}

\item{item_attrib}{A \code{data.frame} containing item attributes. Use \code{\link{loadItemAttrib}} for this.}
}
\value{
A \code{data.frame} containing stimulus attributes.
}
\description{
Read set, stimulus, or passage attributes from specified file.
}
\examples{
\donttest{
itempool_reading <- loadItemPool("itempool_reading.csv")
itemattrib_reading <- loadItemAttrib("itemattrib_reading.csv", itempool_reading)
stimattrib_reading <- loadStAttrib("stimattrib_reading.csv", itemattrib_reading)
constraints_reading <- loadConstraints("constraints_reading.csv",
  itempool_reading, itemattrib_reading, stimattrib_reading)
}
}
\seealso{
\link{dataset_reading} for example usage.
}
