% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{logitHyperPars}
\alias{logitHyperPars}
\title{logitHyperPars}
\usage{
logitHyperPars(mean, sd)
}
\arguments{
\item{mean}{A nice parameter}

\item{sd}{A nice parameter}
}
\description{
logitHyperPars
}
