% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning_functions.r
\name{makeConstraintsByEachPartition}
\alias{makeConstraintsByEachPartition}
\title{make constraints objects from Split() solution indices}
\usage{
makeConstraintsByEachPartition(constraints, solution_per_bin)
}
\arguments{
\item{constraints}{a \code{\linkS4class{constraints}} object representing test specifications. Use \code{\link{loadConstraints}} for this.}

\item{solution_per_bin}{a list containing item/stimulus indices for each partition.
This accepts a list stored in the \code{output} slot of an \code{\linkS4class{output_Split}} object.}
}
\value{
\code{\link{makeConstraintsByEachPartition}} returns a list of \code{\linkS4class{constraints}} objects.
}
\description{
\code{\link{makeConstraintsByEachPartition}} is a helper function for making
\code{\linkS4class{constraints}} objects from \code{\link{Split}} solution indices.
}
