% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_functions.R
\name{makeDiagnosticExposureRecord}
\alias{makeDiagnosticExposureRecord}
\title{(Internal) Collect diagnostic exposure record}
\usage{
makeDiagnosticExposureRecord(
  true_theta,
  segment_record,
  diagnostic_exposure_record,
  config,
  simulation_constants
)
}
\arguments{
\item{true_theta}{examinee's true theta. Used to determine the segment the true theta belongs to.}

\item{segment_record}{a named list containing counts on theta segments.}

\item{diagnostic_exposure_record}{a named list containing the diagnostic exposure record.}

\item{config}{a \code{\linkS4class{config_Shadow}} object. Use \code{\link{createShadowTestConfig}} for this.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{makeDiagnosticExposureRecord}} returns an updated exposure record.
}
\description{
\code{\link{makeDiagnosticExposureRecord}} is an internal function for
collecting diagnostic exposure record.
}
\keyword{internal}
