% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{MakeItemPoolCluster}
\alias{MakeItemPoolCluster}
\title{MakeItemPoolCluster}
\usage{
MakeItemPoolCluster(pools, names = NULL)
}
\arguments{
\item{pools}{A nice parameter}

\item{names}{A nice parameter}
}
\description{
Create a \code{\linkS4class{pool.cluster}} object
}
