% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning_functions.r
\name{makeSetStructureConstraints}
\alias{makeSetStructureConstraints}
\title{(Internal) Make set-based strucutre constraints}
\usage{
makeSetStructureConstraints(constraints)
}
\arguments{
\item{constraints}{a \code{\linkS4class{constraints}} object representing test specifications. Use \code{\link{loadConstraints}} for this.}
}
\value{
\code{\link{makeSetStructureConstraints}} returns a matrix.
}
\description{
\code{\link{makeSetStructureConstraints}} is an internal function for
making the left-hand side matrix of set-based structure constraints.
}
\keyword{internal}
