% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{MakeTest}
\alias{MakeTest}
\alias{MakeTest,item.pool-method}
\title{An S4 Generic}
\usage{
MakeTest(object, theta, infoType = "FISHER", trueTheta = NULL)

\S4method{MakeTest}{item.pool}(object, theta, infoType = "FISHER",
  trueTheta = NULL)
}
\arguments{
\item{object}{A nice parameter}

\item{theta}{A nice parameter}

\item{infoType}{A nice parameter}

\item{trueTheta}{A nice parameter}
}
\description{
MakeTest
}
