% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{makeTestCluster}
\alias{makeTestCluster}
\alias{makeTestCluster,pool.cluster,numeric,numeric-method}
\alias{makeTestCluster,pool.cluster,numeric,list-method}
\title{Generate a test cluster object}
\usage{
makeTestCluster(object, theta, trueTheta)

\S4method{makeTestCluster}{pool.cluster,numeric,numeric}(object, theta,
  trueTheta)

\S4method{makeTestCluster}{pool.cluster,numeric,list}(object, theta,
  trueTheta)
}
\arguments{
\item{object}{An \code{\linkS4class{pool.cluster}} object}

\item{theta}{A grid of theta values}

\item{trueTheta}{An optional vector of true theta values to simulate response data}
}
\description{
Generate a \code{\linkS4class{test.cluster}} object
}
