% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{maxinfo_plot}
\alias{maxinfo_plot}
\title{maxinfo_plot}
\usage{
maxinfo_plot(pool, const)
}
\arguments{
\item{pool}{An \code{\linkS4class{item.pool}} object}

\item{const}{A nice parameter}
}
\description{
maxinfo_plot
}
