% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{maxinfoplot}
\alias{maxinfoplot}
\title{maxinfoplot}
\usage{
maxinfoplot(pool, const)
}
\arguments{
\item{pool}{An \code{\linkS4class{item.pool}} object}

\item{const}{A nice parameter}
}
\description{
maxinfoplot
}
