% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{mle}
\alias{mle}
\alias{mle,item.pool-method}
\title{Generate maximum likelihood estimates of theta}
\usage{
mle(object, resp, startTheta = NULL, maxIter = 100, crit = 0.001,
  select = NULL, thetaRange = c(-4, 4), truncate = FALSE,
  maxChange = 1, FisherScoring = TRUE)

\S4method{mle}{item.pool}(object, resp, startTheta = NULL,
  maxIter = 50, crit = 0.005, select = NULL, thetaRange = c(-4, 4),
  truncate = FALSE, maxChange = 1, FisherScoring = TRUE)
}
\arguments{
\item{object}{A \code{\linkS4class{test}} object.}

\item{resp}{A vector (or matrix) of item responses.}

\item{startTheta}{An optional vector of start theta values.}

\item{maxIter}{Maximum number of iterations.}

\item{crit}{Convergence criterion.}

\item{select}{A vector of indices identifying the items to subset.}

\item{thetaRange}{A range of theta values.}

\item{truncate}{Set \code{TRUE} to bound MLE to thetaRange: c(minTheta, maxTheta).}

\item{maxChange}{Maximum change between iterations.}

\item{FisherScoring}{\code{TRUE} to use Fisher's method of scoring.}
}
\description{
Generate maximum likelihood estimates of theta.
}
