% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{MLE}
\alias{MLE}
\alias{MLE,test-method}
\alias{MLE,test.cluster-method}
\title{MLE}
\usage{
MLE(object, startTheta = NULL, maxIter = 100, crit = 0.001,
  select = NULL, thetaRange = c(-4, 4), truncate = FALSE,
  maxChange = 1, FisherScoring = TRUE)

\S4method{MLE}{test}(object, startTheta = NULL, maxIter = 100,
  crit = 0.001, select = NULL, thetaRange = c(-4, 4),
  truncate = FALSE, maxChange = 1, FisherScoring = TRUE)

\S4method{MLE}{test.cluster}(object, startTheta = NULL, maxIter = 100,
  crit = 0.001, select = NULL)
}
\arguments{
\item{object}{A \code{\linkS4class{test}} object}

\item{startTheta}{An optional vector of start theta values}

\item{maxIter}{Maximum number of iterations}

\item{crit}{Convergence criterion}

\item{select}{A vector of indices identifying the items to subset}

\item{thetaRange}{A range of theta values, c(minTheta, maxTheta)}

\item{truncate}{TRUE to bound MLE to thetaRange}

\item{maxChange}{Maximum change between iterations}

\item{FisherScoring}{TRUE to use Fisher's method of scoring}
}
\description{
Generate maximum likelihood estimates of theta

Genereate maximum likelihood estimates of theta
}
