% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_class.R
\docType{class}
\name{output_Shadow_all-class}
\alias{output_Shadow_all-class}
\title{An S4 class to represent the output from Shadow()}
\description{
An S4 class to represent the output from Shadow()
}
\section{Slots}{

\describe{
\item{\code{output}}{List of \code{\linkS4class{output_Shadow}} objects. The list of assembled test for each examinee.}

\item{\code{pool}}{A \code{\linkS4class{item_pool}} object. The item pool used in assembly.}

\item{\code{config}}{A \code{\linkS4class{config_Shadow}} object. The config used in assembly.}

\item{\code{true_theta}}{True theta values if were supplied.}

\item{\code{constraints}}{A \code{\linkS4class{constraints}} object. The constraints used in assembly.}

\item{\code{prior}}{Prior}

\item{\code{prior_par}}{foo}

\item{\code{data}}{The response data used in assembly.}

\item{\code{final_theta_est}}{Final (i.e. not interim) estimates of theta for each examinee.}

\item{\code{final_se_est}}{Final (i.e. not interim) standard error estimates of theta for each examinee.}

\item{\code{exposure_rate}}{Exposure rate of each item in the pool.}

\item{\code{usage_matrix}}{The matrix representing which items were used in each item position.}

\item{\code{true_segment_count}}{foo}

\item{\code{est_segment_count}}{foo}

\item{\code{eligibility_stats}}{foo}

\item{\code{check_eligibility_stats}}{foo}

\item{\code{no_fading_eligibility_stats}}{foo}

\item{\code{freq_infeasible}}{foo}
}}

