% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_1pl}
\alias{p_1pl}
\title{Calculate the probability of a correct response according to the 1pl model}
\usage{
p_1pl(x, b)
}
\arguments{
\item{x}{a length-one numeric vector for a theta value}

\item{b}{a length-one numeric vector for the difficulty parameter}
}
\description{
Calculate the probability of a correct response according to the 1pl model
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}
}
