% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_2pl}
\alias{p_2pl}
\title{Calculate the probability of a correct response according to the 2PL model}
\usage{
p_2pl(x, a, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A difficulty parameter value.}
}
\description{
Calculate the probability of a correct response according to the 2PL model.
}
\references{
{
  \insertRef{lord_theory_1952}{Shadow}

  \insertRef{birnbaum_efficient_1957}{Shadow}

  \insertRef{birnbaum_estimation_1958}{Shadow}

  \insertRef{birnbaum_further_1958}{Shadow}
}
}
