% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gr}
\alias{p_gr}
\title{Calculate the response probabilities according to the graded response model}
\usage{
p_gr(x, a, b)
}
\arguments{
\item{x}{a length-one numeric vector for a theta value}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a numeric vector for the category boundary parameters}
}
\description{
Calculate the response probabilities according to the graded response model
}
\references{
{
  \insertRef{samejima_estimation_1969}{Shadow}
}
}
