% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cpp_documents.R
\name{p_1pl}
\alias{p_1pl}
\alias{p_2pl}
\alias{p_3pl}
\alias{p_pc}
\alias{p_gpc}
\alias{p_gr}
\alias{array_p_1pl}
\alias{array_p_2pl}
\alias{array_p_3pl}
\alias{array_p_pc}
\alias{array_p_gpc}
\alias{array_p_gr}
\alias{p_item}
\title{Calculate item response probability}
\usage{
p_1pl(x, b)

p_2pl(x, a, b)

p_3pl(x, a, b, c)

p_pc(x, b)

p_gpc(x, a, b)

p_gr(x, a, b)

array_p_1pl(x, b)

array_p_2pl(x, a, b)

array_p_3pl(x, a, b, c)

array_p_pc(x, b)

array_p_gpc(x, a, b)

array_p_gr(x, a, b)
}
\arguments{
\item{x}{the theta value. This must be a column vector in matrix form for \code{array_p_*} functions.}

\item{b}{the *b*-parameter.}

\item{a}{the *a*-parameter.}

\item{c}{the *c*-parameter.}
}
\description{
\code{p_*} and \code{array_p_*} are functions to calculate item response probability.
}
\details{
\code{p_*} functions accept a single theta value, and \code{array_p_*} functions accept multiple theta values.

Currently supports unidimensional models.

\itemize{
  \item{\code{p_1pl}, \code{array_p_1pl}}: 1PL models
  \item{\code{p_2pl}, \code{array_p_2pl}}: 2PL models
  \item{\code{p_3pl}, \code{array_p_3pl}}: 3PL models
  \item{\code{p_pc}, \code{array_p_pc}}: PC (partial credit) models
  \item{\code{p_gpc}, \code{array_p_gpc}}: GPC (generalized partial credit) models
  \item{\code{p_gr}, \code{array_p_gr}}: GR (graded response) models
}
}
\examples{
x <- 0.5

p_1pl(x, 1)
p_2pl(x, 1, 2)
p_3pl(x, 1, 2, 0.25)
p_pc(x, c(0, 1))
p_gpc(x, 2, c(0, 1))
p_gr(x, 2, c(0, 2))

x <- matrix(seq(0.1, 0.5, 0.1)) # column vector in matrix form

array_p_1pl(x, 1)
array_p_2pl(x, 1, 2)
array_p_3pl(x, 1, 2, 0.25)
array_p_pc(x, c(0, 1))
array_p_gpc(x, 2, c(0, 1))
array_p_gr(x, 2, c(0, 2))

}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}

{
  \insertRef{lord_theory_1952}{TestDesign}
}

{
  \insertRef{birnbaum_efficient_1957}{TestDesign}
}

{
  \insertRef{birnbaum_estimation_1958}{TestDesign}
}

{
  \insertRef{birnbaum_further_1958}{TestDesign}
}

{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}

{
  \insertRef{masters_rasch_1982}{TestDesign}
}

{
  \insertRef{andrich_rating_1978}{TestDesign}
}

{
  \insertRef{muraki_generalized_1992}{TestDesign}
}

{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
