% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_m_3pl}
\alias{p_m_3pl}
\title{Calculate the probability of a correct response according to the 2pl model (multivariate)}
\usage{
p_m_3pl(x, a, b, c)
}
\arguments{
\item{x}{a length-one numeric vector for a theta value}

\item{b}{a length-one numeric vector for the difficulty parameter}
}
\description{
Calculate the probability of a correct response according to the 2pl model (multivariate)
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{IRTclass}
}
}
