% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_m_gpc}
\alias{p_m_gpc}
\title{Calculate the response probabilities according to the generalizaed partial credit model (multivariate)}
\usage{
p_m_gpc(x, a, b)
}
\arguments{
\item{x}{a length-one numeric vector for a theta value}

\item{a}{a length-one numeric vector for the slope parameter}

\item{b}{a numeric vector for the threshold parameters}
}
\description{
Calculate the response probabilities according to the generalizaed partial credit model (multivariate)
}
\references{
{
  \insertRef{muraki_generalized_1992}{Shadow}
}
}
