% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{parseConstraintData}
\alias{parseConstraintData}
\title{(Internal) Parse a constraint data into an object}
\usage{
parseConstraintData(x, attrib, constants)
}
\arguments{
\item{x}{a \code{\link{data.frame}} row containing a single constraint data.}

\item{attrib}{an \code{\linkS4class{item_attrib}} object or a \code{\linkS4class{st_attrib}} object.}

\item{constants}{a named list containing constants.}
}
\value{
\code{\link{parseConstraintData}} returns a \code{\linkS4class{constraint}} object.
}
\description{
\code{\link{parseConstraintData}} is an internal function for
parsing data for a single constraint.
}
\keyword{internal}
