% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowtest_functions.R
\name{parseShadowTestRefreshSchedule}
\alias{parseShadowTestRefreshSchedule}
\title{(Internal) Parse shadowtest refresh schedule}
\usage{
parseShadowTestRefreshSchedule(simulation_constants, refresh_policy)
}
\arguments{
\item{simulation_constants}{a named list containing simulation constants.}

\item{refresh_policy}{the \code{refresh_policy} slot from a \code{\linkS4class{config_Shadow}} object.}
}
\value{
\code{\link{parseShadowTestRefreshSchedule}} returns a named list.
}
\description{
\code{\link{parseShadowTestRefreshSchedule}} is an internal function for
parsing shadowtest refresh schedule from supplied config.
}
\keyword{internal}
