% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/plot_functions.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,item_pool-method}
\alias{plot,output_Static-method}
\title{Extension of plot() for objects in TestDesign package}
\usage{
\S4method{plot}{item_pool}(
  x,
  y,
  theta = seq(-3, 3, 0.1),
  type = "info",
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "blue",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(1, 1),
  ...
)

\S4method{plot}{output_Static}(
  x,
  y,
  theta = seq(-3, 3, 0.1),
  type = NULL,
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "blue",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(2, 4),
  ...
)
}
\arguments{
\item{x}{An \code{\linkS4class{item_pool}} object to draw pool-level or item-level information, or a list from \code{\link{Static}} to draw test-level information.}

\item{y}{Unused argument, exists for compatibility with \code{\link{plot}} in the base R package.}

\item{theta}{Theta values for drawing the curve. Default is \code{seq(-3, 3, .1)}.}

\item{type}{Type of the plot. \code{info} to plot information, \code{score} to plot expected scores, \code{shadow} to plot shadow test chart, \code{audit} to plot audit trail, and \code{exposure} to plot exposure rates.}

\item{info_type}{Type of information. Currently only accepts \code{FISHER} (default).}

\item{plot_sum}{When 'object' is an \code{\linkS4class{item_pool}} object, if \code{TRUE} then draw pool-level information, and if \code{FALSE} draw item-level information for every item in the pool.}

\item{select}{A vector of indices identifying the items to subset, for when 'object' is an \code{\linkS4class{item_pool}} object.}

\item{color}{The color of the curve.}

\item{file_pdf}{If supplied a filename, save as a PDF file.}

\item{width}{Width of graphics device.}

\item{height}{Height of graphics device.}

\item{mfrow}{Multipanel configurations as c(nrow, ncol).}

\item{...}{Arguments to pass onto \code{\link{plot}}.}
}
\description{
Extension of plot() for objects in TestDesign package

Plot from objects in the TestDesign package.
}
\examples{
## Plot item information of a pool
subitempool <- subsetItemPool(itempool_science, 1:8)
plot(subitempool)
plot(itempool_science, select = 1:8)

## Plot expected score of a pool
plot(subitempool, type = "score")
plot(itempool_science, type = "score", select = 1:8)

## Plot assembly results from Static()
config <- createStaticTestConfig()
solution <- Static(config, constraints_science)
plot(solution)

## Plot attainable information range from constraints
plot(constraints_science)

}
