% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\docType{methods}
\name{plot,output_Shadow_all-method}
\alias{plot,output_Shadow_all-method}
\alias{plot,output_Shadow-method}
\title{Draw an audit trail plot}
\usage{
\S4method{plot}{output_Shadow_all}(
  x,
  y,
  examinee_id = 1,
  type = "audit",
  min_theta = -5,
  max_theta = 5,
  min_score = 0,
  max_score = 1,
  z_ci = 1.96,
  simple = FALSE,
  sort_by_difficulty = FALSE,
  theta_segment = "Estimated",
  color = "blue",
  color_final = "blue",
  file_pdf = NULL,
  ...
)

\S4method{plot}{output_Shadow}(
  x,
  y,
  examinee_id = 1,
  type = "audit",
  min_theta = -5,
  max_theta = 5,
  min_score = 0,
  max_score = 1,
  z_ci = 1.96,
  ...
)
}
\arguments{
\item{x}{An output object generated by \code{\link{Shadow}}.}

\item{y}{Unused argument, exists for compatibility with \code{\link{plot}} in the base R package.}

\item{examinee_id}{Numeric ID of the examinee to draw the plot.}

\item{type}{If 'audit', draw an audit trail plot. If 'shadow', draw a shadow chart. If 'exposure', draw exposure rate plots.}

\item{min_theta}{For type 'audit', the lower bound of theta range to plot.}

\item{max_theta}{For type 'audit', the upper bound of theta range to plot.}

\item{min_score}{For type 'audit', the minimum item score.}

\item{max_score}{For type 'audit', the maximum item score.}

\item{z_ci}{For type 'audit', the quantile of the normal distribution for confidence intervals.}

\item{simple}{For type 'shadow', if \code{TRUE}, simplity the chart by hiding unused items.}

\item{sort_by_difficulty}{For type 'shadow', sort the items by difficulty. (not implemented)}

\item{theta_segment}{For type 'exposure', True or Estimated theta used to create segments ("Estimated" or "True").}

\item{color}{For type 'exposure', color of item-wise exposure rates.}

\item{color_final}{For type 'exposure', color of item-wise exposure rates, only counting the items while in the final theta segment as exposed.}

\item{file_pdf}{If supplied a filename, save as a PDF file.}

\item{...}{Additional options to be passed on to \code{pdf()}.}
}
\description{
Draw an audit trail plot.
}
\examples{
config <- createShadowTestConfig()
true_theta <- rnorm(1)
solution <- Shadow(config, constraints_science, true_theta)
plot(solution, type = 'audit' , examinee_id = 1)
plot(solution, type = 'shadow', examinee_id = 1)
plot(solution, type = 'exposure')

}
