% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotAvgSegmentVisit}
\alias{plotAvgSegmentVisit}
\title{plotAvgSegmentVisit}
\usage{
plotAvgSegmentVisit(object, pdfFile = NULL, width = 7, height = 6)
}
\arguments{
\item{object}{A nice parameter}

\item{pdfFile}{A nice parameter}

\item{width}{A nice parameter}

\item{height}{A nice parameter}
}
\description{
plotAvgSegmentVisit
}
