% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{plotCAT}
\alias{plotCAT}
\alias{plotCAT,Shadow.output-method}
\alias{plotCAT,list-method}
\title{plotCAT}
\usage{
plotCAT(object, minTheta = -5, maxTheta = 5, minScore = 0,
  maxScore = 1, zCI = 1.96, PDF = NULL)

\S4method{plotCAT}{Shadow.output}(object, minTheta = -5, maxTheta = 5,
  minScore = 0, maxScore = 1, zCI = 1.96, PDF = NULL)

\S4method{plotCAT}{list}(object, minTheta = -5, maxTheta = 5,
  minScore = 0, maxScore = 1, zCI = 1.96, PDF = NULL)
}
\arguments{
\item{object}{A nice parameter}

\item{minTheta}{A nice parameter}

\item{maxTheta}{A nice parameter}

\item{minScore}{A nice parameter}

\item{maxScore}{A nice parameter}

\item{zCI}{A nice parameter}

\item{PDF}{A nice parameter}
}
\description{
plotCAT

plotCAT
}
