% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotEligibilityStats}
\alias{plotEligibilityStats}
\title{plotEligibilityStats}
\usage{
plotEligibilityStats(config, object = NULL, objectNoFading = NULL,
  file = NULL, fileNoFading = NULL, segment = 1, items = c(1),
  PDF = NULL, maxRate = 0.25, discardFirst = NULL)
}
\arguments{
\item{config}{A \code{\linkS4class{Shadow.config}} object.}

\item{object}{A nice parameter}

\item{objectNoFading}{A nice parameter}

\item{file}{A nice parameter}

\item{fileNoFading}{A nice parameter}

\item{segment}{A nice parameter}

\item{items}{A nice parameter}

\item{PDF}{A nice parameter}

\item{maxRate}{A nice parameter}

\item{discardFirst}{A nice parameter}
}
\description{
plotEligibilityStats
}
