% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotExposureRateBySegment}
\alias{plotExposureRateBySegment}
\title{plotExposureRateBySegment}
\usage{
plotExposureRateBySegment(object, config, maxRate = 0.25,
  pdfFile = NULL, width = 7, height = 6, mfrow = c(2, 4))
}
\arguments{
\item{object}{A nice parameter}

\item{config}{A nice parameter}

\item{maxRate}{A nice parameter}

\item{pdfFile}{A nice parameter}

\item{width}{A nice parameter}

\item{height}{A nice parameter}

\item{mfrow}{A nice parameter}
}
\description{
plotExposureRateBySegment
}
