% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotExposureRateFinal}
\alias{plotExposureRateFinal}
\title{plotExposureRateFinal}
\usage{
plotExposureRateFinal(object, config = NULL, maxRate = 0.25,
  theta = "Estimated", segmentCut = NULL, color = "red",
  pdfFile = NULL, width = 7, height = 6, mfrow = c(2, 4),
  burnIn = 0, retain = NULL)
}
\arguments{
\item{object}{A nice parameter}

\item{config}{A nice parameter}

\item{maxRate}{A nice parameter}

\item{theta}{A nice parameter}

\item{segmentCut}{A nice parameter}

\item{color}{A nice parameter}

\item{pdfFile}{A nice parameter}

\item{width}{A nice parameter}

\item{height}{A nice parameter}

\item{mfrow}{A nice parameter}

\item{burnIn}{A nice parameter}

\item{retain}{A nice parameter}
}
\description{
plotExposureRateFinal
}
