% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotExposureRateFinalFlag}
\alias{plotExposureRateFinalFlag}
\title{plotExposureRateFinalFlag}
\usage{
plotExposureRateFinalFlag(object, pool, theta = seq(-3, 3, 0.1),
  flagCriterior = 0.4, pdfFile = NULL, width = 7, height = 6,
  color = "red", mfrow = c(2, 4))
}
\arguments{
\item{object}{A nice parameter}

\item{pool}{A nice parameter}

\item{theta}{A nice parameter}

\item{flagCriterior}{A nice parameter}

\item{pdfFile}{A nice parameter}

\item{width}{A nice parameter}

\item{height}{A nice parameter}

\item{color}{A nice parameter}

\item{mfrow}{A nice parameter}
}
\description{
plotExposureRateFinalFlag
}
