% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotInfoOverlay}
\alias{plotInfoOverlay}
\title{plotInfoOverlay}
\usage{
plotInfoOverlay(object, theta, infoType = "FISHER", select = NULL,
  pdfFile = NULL, color = "red", width = 7, height = 6)
}
\arguments{
\item{object}{A nice parameter}

\item{theta}{A nice parameter}

\item{infoType}{A nice parameter}

\item{select}{A nice parameter}

\item{pdfFile}{A nice parameter}

\item{color}{A nice parameter}

\item{width}{A nice parameter}

\item{height}{A nice parameter}
}
\description{
plotInfoOverlay
}
