% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{plotMaxInfo}
\alias{plotMaxInfo}
\title{Draw a plot of maximum attainable information given the imposed constraints}
\usage{
plotMaxInfo(pool, constraints, theta = seq(-3, 3, 0.5))
}
\arguments{
\item{pool}{An \code{\linkS4class{item_pool}} object.}

\item{constraints}{A list constraints generated by \code{\link{loadConstraints}}.}

\item{theta}{A theta grid.}
}
\description{
Draw a plot of maximum attainable information given the imposed constraints.
}
\examples{
p <- plotMaxInfo(itempool_science, constraints_science)

}
