% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotRMSE}
\alias{plotRMSE}
\title{plotRMSE}
\usage{
plotRMSE(..., title = NULL, legendTitle = NULL, legendLabels = NULL,
  ltySet = NULL, colSet = NULL, theta = seq(-2, 2, 1))
}
\arguments{
\item{...}{Some dots}

\item{title}{A nice parameter}

\item{legendTitle}{A nice parameter}

\item{legendLabels}{A nice parameter}

\item{ltySet}{A nice parameter}

\item{colSet}{A nice parameter}

\item{theta}{A nice parameter}
}
\description{
plotRMSE
}
