% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{plotShadow}
\alias{plotShadow}
\alias{plotShadow,Shadow.output-method}
\alias{plotShadow,list-method}
\title{An S4 generic}
\usage{
plotShadow(object, Constraints, sortByDifficulty = FALSE, PDF = NULL)

\S4method{plotShadow}{Shadow.output}(object, Constraints,
  sortByDifficulty = FALSE, PDF = NULL)

\S4method{plotShadow}{list}(object, Constraints,
  sortByDifficulty = FALSE, PDF = NULL)
}
\arguments{
\item{object}{A nice parameter}

\item{Constraints}{A nice parameter}

\item{sortByDifficulty}{A nice parameter}

\item{PDF}{A nice parameter}
}
\description{
plotShadow

plotShadow
}
