% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotShadowAudit}
\alias{plotShadowAudit}
\title{(Internal) Plot audit trail of a single examinee}
\usage{
plotShadowAudit(x, theta_range, z_ci, use_par, ...)
}
\arguments{
\item{x}{an \code{\linkS4class{output_Shadow}} object.}

\item{theta_range}{the theta range to use on the y-axis.}

\item{z_ci}{the confindence interval value to use in the \emph{z} metric. \code{z_ci = 1.96} is equivalent to plotting 95\% error bars.}

\item{use_par}{\code{TRUE} allows this function to set graphic margins through \code{\link{par}}.}

\item{...}{not used.}
}
\description{
\code{\link{plotShadowAudit}} is an internal function for
plotting audit trail of a single examinee,
showing interim theta estimates over the course of an adaptive test.
}
\keyword{internal}
