% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotShadowChart}
\alias{plotShadowChart}
\title{(Internal) Plot shadow chart of a single examinee}
\usage{
plotShadowChart(x, simple, use_par, ...)
}
\arguments{
\item{x}{an \code{\linkS4class{output_Shadow}} object.}

\item{simple}{\code{TRUE} simplifies the chart by hiding items that were never used in any shadowtest.}

\item{use_par}{\code{TRUE} allows this function to set graphic margins through \code{\link{par}}.}

\item{...}{not used.}
}
\description{
\code{\link{plotShadowAudit}} is an internal function for
plotting shadow chart of a single examinee,
showing all shadowtests over the course of an adaptive test.
}
\keyword{internal}
