% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotShadowExposure}
\alias{plotShadowExposure}
\title{(Internal) Plot exposure rates from a simulation}
\usage{
plotShadowExposure(
  x,
  theta_type,
  segment,
  use_segment_label,
  color,
  color_final,
  rmse,
  use_par,
  ...
)
}
\arguments{
\item{x}{an \code{\linkS4class{output_Shadow_all}} object.}

\item{theta_type}{\code{true} or \code{estimated}. The type of theta to define theta segments for plotting purposes.}

\item{segment}{the segment index to do the plotting. \code{NULL} plots all segments.}

\item{use_segment_label}{\code{TRUE} labels each segment using the theta range of the segment.}

\item{color}{the color for exposure rate bars.}

\item{color_final}{the color for exposure rate bars, treating items administered in non-true segments as not exposed.}

\item{rmse}{\code{TRUE} shows the theta estimation RMSE in each segment.}

\item{use_par}{\code{TRUE} allows this function to set graphic margins through \code{\link{par}}.}

\item{...}{not used.}
}
\description{
\code{\link{plotShadowExposure}} is an internal function for
plotting exposure rates from a simulation.
}
\keyword{internal}
