% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotShadowInfo}
\alias{plotShadowInfo}
\title{(Internal) Plot test information of a single shadow test}
\usage{
plotShadowInfo(
  x,
  examinee_id,
  position,
  info_type,
  ylim = NULL,
  theta,
  color,
  use_par,
  ...
)
}
\arguments{
\item{x}{an \code{\linkS4class{output_Shadow_all}} object.}

\item{examinee_id}{the numeric index of an examinee.}

\item{position}{the position within the test administration (i.e., test progress), ranging from 1 to test length.}

\item{info_type}{the information type. Currently only accepts \code{FISHER}.}

\item{ylim}{(optional) the y-axis range. Defaults to 0 to maximum information.}

\item{theta}{theta quadrature points to compute information values at.}

\item{color}{the color for test information curve.}

\item{use_par}{not used.}

\item{...}{not used.}
}
\description{
\code{\link{plotShadowInfo}} is an internal function for
plotting the test information curve of a single shadow test.
}
\keyword{internal}
