% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{+.item.pool}
\alias{+.item.pool}
\title{+.item.pool}
\usage{
\method{+}{item.pool}(pool1, pool2)
}
\arguments{
\item{pool1}{A nice parameter}

\item{pool2}{A nice parameter}
}
\description{
+.item.pool
}
