% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{pool.cluster-class}
\alias{pool.cluster-class}
\title{An S4 class to represent a cluster of item pools}
\description{
An S4 class to represent a cluster of item pools.
}
\section{Slots}{

\describe{
\item{\code{np}}{A scalar to indicate the number of item pools in the cluster.}

\item{\code{pools}}{A list of \code{item.pool} objects.}

\item{\code{names}}{A character vector of item pool names of length np.}
}}

