% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\docType{methods}
\name{print}
\alias{print}
\title{Extension of print() for objects in TestDesign package}
\arguments{
\item{x}{an object to print.}
}
\description{
Extension of print() for objects in TestDesign package
}
