% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\name{returnErrors}
\alias{returnErrors}
\title{(Internal) Return S4 object validation error messages}
\usage{
returnErrors(errors)
}
\arguments{
\item{errors}{a vector containing error messages.}
}
\value{
\code{\link{returnErrors}} returns \code{TRUE} if the input is empty.
otherwise it returns the error messages as a pass-thru.
}
\description{
\code{\link{returnErrors}} is an internal function for
S4 object validation error messages.
}
\keyword{internal}
