% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(x, y, conditional = TRUE)
}
\arguments{
\item{x}{A nice parameter}

\item{y}{A nice parameter}

\item{conditional}{A nice parameter}
}
\description{
RMSE
}
