% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{runShadow}
\alias{runShadow}
\title{runShadow}
\usage{
runShadow(pool, trueTheta, constraints, exposureControl = "ELIGIBILITY",
  maxExposureRate = 0.2, nSegment = 1, segmentCut = c(-Inf, Inf),
  itemSelection = "MPWI", refreshPolicy = "ALWAYS", position = 1:12,
  threshold = 0.1, config = NULL, Data = NULL, fixedTheta = NULL)
}
\arguments{
\item{pool}{A nice parameter}

\item{trueTheta}{A nice parameter}

\item{constraints}{A nice parameter}

\item{exposureControl}{A nice parameter}

\item{maxExposureRate}{A nice parameter}

\item{nSegment}{A nice parameter}

\item{segmentCut}{A nice parameter}

\item{itemSelection}{A nice parameter}

\item{refreshPolicy}{A nice parameter}

\item{position}{A nice parameter}

\item{threshold}{A nice parameter}

\item{config}{A nice parameter}

\item{Data}{A nice parameter}

\item{fixedTheta}{A nice parameter}
}
\description{
runShadow
}
