% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{sanitizeConstraintsData}
\alias{sanitizeConstraintsData}
\title{(Internal) Sanitize constraints data}
\usage{
sanitizeConstraintsData(x)
}
\arguments{
\item{x}{a \code{\link{data.frame}} containing constraints data.}
}
\value{
\code{\link{sanitizeConstraintsData}} returns sanitized constraints data.
}
\description{
\code{\link{sanitizeConstraintsData}} is an internal function for
sanitizing constraints data.
}
\keyword{internal}
