% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{sanitizeModel}
\alias{sanitizeModel}
\title{(Internal) Sanitize item model names for C++ functions}
\usage{
sanitizeModel(model)
}
\arguments{
\item{model}{a vector containing item model names.}
}
\value{
\code{\link{sanitizeModel}} returns a numeric vector, with \itemize{
  \item{\code{item_1PL} converted to \code{1}}
  \item{\code{item_2PL} converted to \code{2}}
  \item{\code{item_3PL} converted to \code{3}}
  \item{\code{item_PC} converted to \code{4}}
  \item{\code{item_GPC} converted to \code{5}}
  \item{\code{item_GR} converted to \code{6}}
}
}
\description{
\code{\link{sanitizeModel}} is an internal function for
sanitizing item model names to be passed onto C++ functions.
}
\keyword{internal}
