% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{saveOutput}
\alias{saveOutput}
\title{saveOutput}
\usage{
saveOutput(objectList, file = NULL)
}
\arguments{
\item{objectList}{A nice parameter}

\item{file}{A nice parameter}
}
\value{
An output
}
\description{
Do fancy stuff.
}
