% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowtest_functions.R
\name{selectItemFromShadowTest}
\alias{selectItemFromShadowTest}
\title{(Internal) Select item from a shadowtest}
\usage{
selectItemFromShadowTest(
  shadow_test,
  position,
  simulation_constants,
  x,
  previous_selection
)
}
\arguments{
\item{shadow_test}{a \code{\link{data.frame}} containing a shadowtest.}

\item{position}{the item position, ranging from 1 to test length.}

\item{simulation_constants}{a named list containing simulation constants.}

\item{x}{a \code{\linkS4class{output_Shadow}} object.}

\item{previous_selection}{a named list containing item selection from previous position.}
}
\value{
a named list containing data on the selected item.
}
\description{
(Internal) Select item from a shadowtest
}
\keyword{internal}
