% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{Shadow.config-class}
\alias{Shadow.config-class}
\title{Shadow.config}
\description{
Shadow.config
}
\section{Slots}{

\describe{
\item{\code{itemSelection}}{A list.}

\item{\code{contentBalancing}}{A list.}

\item{\code{MIP}}{A list.}

\item{\code{MCMC}}{A list.}

\item{\code{refreshPolicy}}{A list.}

\item{\code{exposureControl}}{A list.}

\item{\code{stoppingCriterion}}{A list.}

\item{\code{interimTheta}}{A list.}

\item{\code{finalTheta}}{A list.}

\item{\code{thetaGrid}}{Numeric.}

\item{\code{auditTrail}}{Logical.}
}}

