% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{Shadow.output-class}
\alias{Shadow.output-class}
\title{Shadow.output}
\description{
Shadow.output
}
\section{Slots}{

\describe{
\item{\code{simuleeIndex}}{Numeric.}

\item{\code{trueTheta}}{Numeric or NULL.}

\item{\code{trueThetaSegment}}{Numeric or NULL.}

\item{\code{finalThetaEst}}{Numeric.}

\item{\code{finalSeEst}}{Numeric.}

\item{\code{administeredItemIndex}}{Numeric.}

\item{\code{administeredItemResp}}{Numeric.}

\item{\code{administeredStimulusIndex}}{Numeric.}

\item{\code{shadowTestRefreshed}}{Logical.}

\item{\code{shadowTestFeasible}}{Logical.}

\item{\code{solveTime}}{Numeric.}

\item{\code{interimThetaEst}}{Numeric.}

\item{\code{interimSeEst}}{Numeric.}

\item{\code{thetaSegmentIndex}}{Numeric.}

\item{\code{prior}}{Numeric.}

\item{\code{priorPar}}{Numeric.}

\item{\code{posterior}}{Numeric.}

\item{\code{posteriorSample}}{Numeric.}

\item{\code{likelihood}}{Numeric.}

\item{\code{shadowTest}}{A list.}
}}

