% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowtest_functions.R
\name{shouldShadowTestBeRefreshed}
\alias{shouldShadowTestBeRefreshed}
\title{(Internal) Determine whether shadowtest should be refreshed}
\usage{
shouldShadowTestBeRefreshed(
  shadowtest_refresh_schedule,
  position,
  theta_change,
  previous_selection
)
}
\arguments{
\item{shadowtest_refresh_schedule}{refresh schedule.}

\item{position}{the position within the current administration (i.e., test progress)}

\item{theta_change}{the change in interim theta.}

\item{previous_selection}{a named list containing data on what item was selected previously.}
}
\value{
\code{\link{shouldShadowTestBeRefreshed}} returns a \code{TRUE} or \code{FALSE} value.
}
\description{
\code{\link{shouldShadowTestBeRefreshed}} is an internal function for
deciding whether a new shadowtest should be assembled.
}
\keyword{internal}
