% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\docType{methods}
\name{show}
\alias{show}
\title{Extension of show() for objects in TestDesign package}
\arguments{
\item{object}{an object to display.}
}
\description{
Extension of show() for objects in TestDesign package
}
