% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_functions.R, R/shadow_functions.R
\docType{methods}
\name{simResp}
\alias{simResp}
\alias{simResp,item.1pl,numeric-method}
\alias{simResp,item.2pl,numeric-method}
\alias{simResp,item.3pl,numeric-method}
\alias{simResp,item.pc,numeric-method}
\alias{simResp,item.gpc,numeric-method}
\alias{simResp,item.gr,numeric-method}
\alias{simResp,item.pool,numeric-method}
\alias{simResp,pool.cluster,numeric-method}
\alias{simResp,pool.cluster,list-method}
\title{Simulate item responses}
\usage{
simResp(object, theta)

\S4method{simResp}{item.1pl,numeric}(object, theta)

\S4method{simResp}{item.2pl,numeric}(object, theta)

\S4method{simResp}{item.3pl,numeric}(object, theta)

\S4method{simResp}{item.pc,numeric}(object, theta)

\S4method{simResp}{item.gpc,numeric}(object, theta)

\S4method{simResp}{item.gr,numeric}(object, theta)

\S4method{simResp}{item.pool,numeric}(object, theta)

\S4method{simResp}{pool.cluster,numeric}(object, theta)

\S4method{simResp}{pool.cluster,list}(object, theta)

\S4method{simResp}{pool.cluster,list}(object, theta)
}
\arguments{
\item{object}{An instance of an item class.}

\item{theta}{A vector of theta values.}
}
\value{
Simulated responses.
}
\description{
An S4 generic and its methods to simulate responses.
}
\examples{
item_1 = new("item.1pl", difficulty = 0.5)
sim_item_1 = simResp(item_1, seq(-3, 3, 1))
item_2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
sim_item_2 = simResp(item_2, seq(-3, 3, 1))
item_3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
sim_item_3 = simResp(item_3, seq(-3, 3, 1))
item_4 = new("item.pc", threshold = c(-1, 0, 1), ncat = 4)
sim_item_4 = simResp(item_4, seq(-3, 3, 1))
item_5 = new("item.gpc", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
sim_item_5 = simResp(item_5, seq(-3, 3, 1))
item_6 = new("item.gr", slope = 0.9, category = c(-1, 0, 1), ncat = 4)
sim_item_6 = simResp(item_6, seq(-3, 3, 1))
sim_itempool <- simResp(itempool_science, seq(-3, 3, 1))
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}

{
  \insertRef{lord_theory_1952}{TestDesign}

  \insertRef{birnbaum_efficient_1957}{TestDesign}

  \insertRef{birnbaum_estimation_1958}{TestDesign}

  \insertRef{birnbaum_further_1958}{TestDesign}
}

{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}

{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}

{
  \insertRef{muraki_generalized_1992}{TestDesign}
}

{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
