% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_cache_class.r
\docType{class}
\name{simulation_data_cache-class}
\alias{simulation_data_cache-class}
\title{Class 'simulation_data_cache': data cache for Shadow()}
\description{
\code{\linkS4class{simulation_data_cache}} is an S4 class to represent data cache for Shadow().
}
\section{Slots}{

\describe{
\item{\code{item_pool}}{the \code{\linkS4class{item_pool}} object.}

\item{\code{theta_grid}}{the theta grid to use as quadrature points.}

\item{\code{prob_grid}}{the list containing item response probabilities at theta quadratures.}

\item{\code{info_grid}}{the matrix containing item information values at theta quadratures.}

\item{\code{max_info}}{the maximum value of \code{info_grid}.}

\item{\code{true_theta}}{(optional) the true theta values.}

\item{\code{response_data}}{(optional) the matrix containing item responses.}
}}

