% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioning_functions.r
\name{splitSolutionToBins}
\alias{splitSolutionToBins}
\title{(Internal) Convert a partitioning problem solution to indices}
\usage{
splitSolutionToBins(solution, n_bins, ni_per_bin, nv_per_bin)
}
\arguments{
\item{solution}{the solution vector from the solver.}

\item{n_bins}{the number of bins (i.e., partitions)}

\item{ni_per_bin}{the number of decision variables for items in each bin.
This is the number of items in the pool.}

\item{nv_per_bin}{the number of decision variables for items+sets in each bin.
This is the number of items in the pool plus the number of sets.
The deviation variable is not counted here.}
}
\value{
\code{\link{splitSolutionToBins}} returns a partition-wise list containing item/set indices.
}
\description{
\code{\link{splitSolutionToBins}} is an internal function for converting a paritioning problem solution to indices.
}
\keyword{internal}
