% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_attrib_operators.R
\docType{methods}
\name{st_attrib-operators}
\alias{st_attrib-operators}
\alias{[,st_attrib,numeric-method}
\alias{[,st_attrib,numeric,ANY,ANY-method}
\alias{dim,st_attrib-method}
\alias{colnames,st_attrib-method}
\alias{rownames,st_attrib-method}
\alias{names,st_attrib-method}
\alias{as.data.frame,st_attrib-method}
\title{Basic functions for stimulus attribute objects}
\usage{
\S4method{[}{st_attrib,numeric}(x, i, j, ..., drop = TRUE)

\S4method{dim}{st_attrib}(x)

\S4method{colnames}{st_attrib}(x)

\S4method{rownames}{st_attrib}(x)

\S4method{names}{st_attrib}(x)

\S4method{as.data.frame}{st_attrib}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{a \code{\linkS4class{st_attrib}} object.}

\item{i, j}{indices to use in subsetting.}

\item{...}{not used, exists for compatibility.}

\item{drop}{not used, exists for compatibility.}

\item{row.names}{not used, exists for compatibility.}

\item{optional}{not used, exists for compatibility.}
}
\description{
Basic functions for stimulus attribute objects
}
\examples{
x <- stimattrib_reading
x[1:10]
dim(x)
ncol(x)
nrow(x)
colnames(x)
rownames(x)
names(x)
as.data.frame(x)

}
