% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{STA}
\alias{STA}
\title{STA}
\usage{
STA(Constraints, objective, solver = "Symphony", xmat = NULL,
  xdir = NULL, xrhs = NULL, maximize = TRUE, mps = FALSE,
  lp = FALSE, verbosity = -2, time_limit = 5, gap_limit = -1, ...)
}
\arguments{
\item{Constraints}{A list representing optimization constraints. Use \code{\link{LoadConstraints}} for this.}

\item{objective}{A vector of objective values.}

\item{solver}{The type of solver. Accepts \code{SYMPHONY, GUROBI, GLPK, LPSOLVE}.}

\item{xmat}{A nice parameter}

\item{xdir}{A nice parameter}

\item{xrhs}{A vector of right-side values.}

\item{maximize}{If \code{TRUE}, treat as a maximization problem. Otherwise treat as a minimization problem.}

\item{mps}{Only used when \code{solver} is \code{SYMPHONY}. If \code{TRUE}, print an MPS representation of the problem for debugging purposes.}

\item{lp}{Only used when \code{solver} is \code{SYMPHONY}. If \code{TRUE}, print an LP representation of the problem for debugging purposes.}

\item{verbosity}{Verbosity level.}

\item{time_limit}{Time limit passed onto the solver.}

\item{gap_limit}{Gap limit passed onto the solver.}

\item{...}{Only used when \code{solver} is \code{SYMPHONY}. Additional parameters to be passed onto the solver.}
}
\value{
An output
}
\description{
Shadow Test Approach to CAT.
}
