% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_functions.R
\docType{methods}
\name{Static}
\alias{Static}
\alias{Static,config_Static-method}
\title{Run fixed-form test assembly}
\usage{
Static(config, constraints)

\S4method{Static}{config_Static}(config, constraints)
}
\arguments{
\item{config}{a \code{\linkS4class{config_Static}} object. Use \code{\link{createStaticTestConfig}} for this.}

\item{constraints}{a \code{\linkS4class{constraints}} object representing test specifications. Use \code{\link{loadConstraints}} for this.}
}
\value{
\code{\link{Static}} returns a \code{\linkS4class{output_Static}} object containing the selected items.
}
\description{
\code{\link{Static}} is a test assembly function to perform fixed-form test assembly based on the generalized shadow-test framework.
}
\examples{
config_science <- createStaticTestConfig(
  list(
    method = "MAXINFO",
    target_location = c(-1, 1)
  )
)
solution <- Static(config_science, constraints_science)

}
\references{
\insertRef{van_der_linden_linear_2005}{TestDesign}
}
