% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratification_functions.R
\name{stratifyItemPool}
\alias{stratifyItemPool}
\title{Stratify item pool}
\usage{
stratifyItemPool(cfg, item_pool, target_a, target_b)
}
\arguments{
\item{cfg}{a \code{\linkS4class{config_Static}} object. Used for solver configuration.}

\item{item_pool}{the item pool to be stratified.}

\item{target_a}{target a-parameter values to use in stratification. This determines the number of produced strata.}

\item{target_b}{target b-parameter values to use in stratification. This affects the evenness of b-distributions between strata. This does not affect the number of produced strata.}
}
\value{
\code{\link{stratifyItemPool}} returns a vector containing stratum indices for each item.
}
\description{
\code{\link{stratifyItemPool}} is function to perform item pool stratification, using the linear programming method described in Chang & van der Linden (2003).
}
\examples{

}
\references{
Chang, H.-H., van der Linden, W. J. (2003).
Optimal stratification of item pools in alpha-stratified computerized adaptive testing.
\emph{Applied Psychological Measurement, 27}(4), 262-274.
}
