% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{subsetItemPool}
\alias{subsetItemPool}
\title{Create a subset of an item pool}
\usage{
subsetItemPool(pool, select = NULL)
}
\arguments{
\item{pool}{An \code{\linkS4class{item.pool}} object.}

\item{select}{A vector of indices identifying the items to subset.}
}
\description{
Create a subset of an item pool.
}
