% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{subsetTest}
\alias{subsetTest}
\title{subsetTest}
\usage{
subsetTest(test, select = NULL)
}
\arguments{
\item{test}{An \code{\linkS4class{test}} object}

\item{select}{A vector of item indices to subset}
}
\description{
Create a subset of a test object
}
