% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/summary_class.R
\docType{class}
\name{summary-classes}
\alias{summary-classes}
\alias{summary_item_pool-class}
\alias{summary_item_attrib-class}
\alias{summary_constraints-class}
\alias{summary_output_Static-class}
\alias{summary_output_Shadow_all-class}
\title{Summary classes}
\description{
Summary classes
}
