% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\docType{methods}
\name{summary-method}
\alias{summary-method}
\alias{summary,item_pool-method}
\alias{summary,item_attrib-method}
\title{summary}
\usage{
\S4method{summary}{item_pool}(object)

\S4method{summary}{item_attrib}(object)
}
\arguments{
\item{object}{An \linkS4class{item_pool} object.}
}
\description{
summary

summary
}
