% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\docType{methods}
\name{summary}
\alias{summary}
\title{Extension of summary() for objects in TestDesign package}
\arguments{
\item{object}{an object to summarize.}
}
\description{
Extension of summary() for objects in TestDesign package
}
