% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{test-class}
\alias{test-class}
\title{test}
\description{
test
}
\section{Slots}{

\describe{
\item{\code{pool}}{An item.pool object.}

\item{\code{theta}}{Numeric.}

\item{\code{Prob}}{A list.}

\item{\code{Info}}{A matrix.}

\item{\code{trueTheta}}{Numeric or Null.}

\item{\code{Data}}{A matrix or Null.}
}}

