% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{test.cluster-class}
\alias{test.cluster-class}
\title{test.cluster}
\description{
Create a \code{\linkS4class{test.cluster}} object from a list of \code{\linkS4class{test}} objects.
}
\section{Slots}{

\describe{
\item{\code{nt}}{Numeric. A scalar to indicate the number of \code{\linkS4class{test}} objects to be clustered.}

\item{\code{tests}}{A list \code{\linkS4class{test}} objects.}

\item{\code{names}}{Character. A vector of names correspondign to the \code{\linkS4class{test}} objects.}
}}

