% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_class.R
\docType{class}
\name{test_cluster-class}
\alias{test_cluster-class}
\title{An S4 class to represent a test cluster}
\description{
An S4 class to represent a test cluster from a list of \code{\linkS4class{test}} objects.
}
\section{Slots}{

\describe{
\item{\code{nt}}{Numeric. A scalar to indicate the number of \code{\linkS4class{test}} objects to be clustered.}

\item{\code{tests}}{A list \code{\linkS4class{test}} objects.}

\item{\code{names}}{Character. A vector of names corresponding to the \code{\linkS4class{test}} objects.}
}}

