% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_operators.R
\docType{methods}
\name{test_operators}
\alias{test_operators}
\alias{subsetTest}
\alias{[,test,ANY-method}
\alias{[,test,numeric,ANY,ANY-method}
\title{Basic operators for test objects}
\usage{
subsetTest(x, i = NULL)

\S4method{[}{test,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{\linkS4class{test}} object.}

\item{i}{item indices to use in subsetting.}

\item{j, drop, ...}{not used, exists for compatibility.}
}
\description{
Create a subset of a \code{\linkS4class{test}} object.
}
