% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{TestDesign}
\alias{TestDesign}
\title{Open TestDesign app}
\usage{
TestDesign()
}
\description{
\code{\link{TestDesign}} is a caller function to open the Shiny interface of TestDesign package.
}
\examples{

\dontrun{
if (interactive()) {
  TestDesign()
}
}

}
