% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runshiny.R
\name{app}
\alias{app}
\alias{OAT}
\title{Open TestDesign app}
\usage{
app()

OAT()
}
\description{
\code{\link{app}} and \code{\link{OAT}} are aliases of \code{\link{TestDesign}}.
}
\details{
\code{\link{TestDesign}} is a caller function to open the Shiny interface of TestDesign package.
}
\examples{

\dontrun{
if (interactive()) {
  TestDesign()
}
}

}
