% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_functions.R
\docType{methods}
\name{testSolver}
\alias{testSolver}
\title{Test solver}
\usage{
testSolver(solver)
}
\arguments{
\item{solver}{a solver package name. Accepts \code{lpSolve, Rsymphony, gurobi, Rglpk}.}
}
\value{
empty string \code{""} if solver works. A string containing error messages otherwise.
}
\description{
Test solver
}
