% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_EAP}
\alias{theta_EAP}
\title{theta_EAP}
\usage{
theta_EAP(theta_grid, item_parm, resp, ncat, model, prior, prior_parm)
}
\arguments{
\item{theta_grid}{A nice parameter}

\item{item_parm}{A nice parameter}

\item{resp}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}

\item{prior}{A nice parameter}

\item{prior_parm}{A nice parameter}
}
\description{
theta_EAP
}
