% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_EB_single}
\alias{theta_EB_single}
\title{theta_EB_single}
\usage{
theta_EB_single(nx, theta_init, theta_prop, item_parm, resp, ncat, model,
  prior, prior_parm)
}
\arguments{
\item{nx}{A nice parameter}

\item{theta_init}{A nice parameter}

\item{theta_prop}{A nice parameter}

\item{item_parm}{A nice parameter}

\item{resp}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}

\item{prior}{A nice parameter}

\item{prior_parm}{A nice parameter}
}
\description{
theta_EB_single
}
