% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_FB}
\alias{theta_FB}
\title{theta_FB}
\usage{
theta_FB(nx, theta_init, theta_prop, items_list, item_init, resp, ncat,
  model, prior, prior_parm)
}
\arguments{
\item{nx}{A nice parameter}

\item{theta_init}{A nice parameter}

\item{theta_prop}{A nice parameter}

\item{items_list}{A nice parameter}

\item{item_init}{A nice parameter}

\item{resp}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}

\item{prior}{A nice parameter}

\item{prior_parm}{A nice parameter}
}
\description{
theta_FB
}
