% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowtest_functions.R
\name{updateCompletedGroupingsRecordForStimulus}
\alias{updateCompletedGroupingsRecordForStimulus}
\title{(Internal) Update groupings record for stimulus}
\usage{
updateCompletedGroupingsRecordForStimulus(
  groupings_record,
  selection,
  o,
  position
)
}
\arguments{
\item{groupings_record}{a list containing data on completed item sets.}

\item{selection}{a named list containing data on selected item.}

\item{o}{a \code{\linkS4class{output_Shadow}} object.}

\item{position}{the item position, ranging from 1 to test length.}
}
\value{
\code{\link{updateCompletedGroupingsRecordForStimulus}} returns
an updated groupings record.
}
\description{
(Internal) Update groupings record for stimulus
}
\keyword{internal}
