% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_functions.R
\name{updateEligibilityRates}
\alias{updateEligibilityRates}
\title{(Internal) Update eligibility rates based on exposure rates}
\usage{
updateEligibilityRates(exposure_record, simulation_constants)
}
\arguments{
\item{exposure_record}{a named list containing exposure records.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{updateEligibilityRates}} returns an updated exposure record.
}
\description{
\code{\link{updateEligibilityRates}} is an internal function for
updating eligibility rates based on exposure rates.
}
\keyword{internal}
