% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_control_functions.R
\name{updateSegmentRecord}
\alias{updateSegmentRecord}
\title{(Internal) Update segment record}
\usage{
updateSegmentRecord(segment_record, segment_of, j)
}
\arguments{
\item{segment_record}{a named list containing counts on theta segments.}

\item{segment_of}{a named list containing theta segments of this examinee.}

\item{j}{examinee index.}
}
\value{
\code{\link{updateSegmentRecord}} returns an updated segment record.
}
\description{
\code{\link{updateSegmentRecord}} is an internal function for
updating segment record.
}
\keyword{internal}
