% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_functions.R
\name{updateThetaPosterior}
\alias{updateThetaPosterior}
\title{(Internal) Update posterior densities}
\usage{
updateThetaPosterior(o, prob_resp)
}
\arguments{
\item{o}{a named list containing posterior densities and likelihoods.}

\item{prob_resp}{a vector containing response probability on a single response category
over a theta grid.}
}
\value{
\code{\link{updateThetaPosterior}} returns an updated list.
}
\description{
\code{\link{updateThetaPosterior}} is an internal function
for updating posterior densities using a response probabilty function.
}
\keyword{internal}
