% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_control_functions.R
\name{updateUsageMatrix}
\alias{updateUsageMatrix}
\title{(Internal) Update item usage matrix}
\usage{
updateUsageMatrix(usage_matrix, j, x, simulation_constants)
}
\arguments{
\item{usage_matrix}{a (\emph{nj}, \emph{nv}) matrix containing item usage.}

\item{j}{the examinee index.}

\item{x}{an \code{\linkS4class{output_Shadow}} object containing the examinee's simulation data.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{updateUsageMatrix}} returns an updated item usage matrix.
}
\description{
\code{\link{updateUsageMatrix}} is an internal function for
updating an item usage matrix. An item usage matrix is a
(\emph{nj}, \emph{nv}) matrix with \code{TRUE} or \code{FALSE} values.
The rows represent examinees, and the columns represent decision variables.
}
\keyword{internal}
