% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{validateConstraintData}
\alias{validateConstraintData}
\title{(Internal) Validate constraint (wrapper for other validators)}
\usage{
validateConstraintData(x, attrib)
}
\arguments{
\item{x}{a \code{\link{data.frame}} row containing a single constraint data.}

\item{attrib}{an \code{\linkS4class{item_attrib}} object or a \code{\linkS4class{st_attrib}} object.}
}
\value{
\code{\link{validateConstraintData}} does not return any values;
it \code{\link{stop}}s if the input constraint is not valid.
}
\description{
\code{\link{validateConstraintData}} is an internal function for
validating a constraint. This is a wrapper function that calls other validator functions
depending on the constraint.
}
\keyword{internal}
