% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{validateExpression}
\alias{validateExpression}
\title{(Internal) Validate constraint condition expression}
\usage{
validateExpression(x, attrib, unit_name, use_lt)
}
\arguments{
\item{x}{a \code{\link{data.frame}} row containing a single constraint data.}

\item{attrib}{an \code{\linkS4class{item_attrib}} object or a \code{\linkS4class{st_attrib}} object.}

\item{unit_name}{\code{items} or \code{stimuli}.}

\item{use_lt}{if \code{TRUE}, will raise an error when number of matching items/stimuli is less than 2.
if \code{FALSE}, will raise an error when number of matching items/stimuli is 0.}
}
\value{
\code{\link{validateExpression}} does not return any values;
it \code{\link{stop}}s if the input constraint is not valid.
}
\description{
\code{\link{validateExpression}} is an internal function for
validating a constraint's condition expression.
}
\keyword{internal}
