% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{validateFullColumn}
\alias{validateFullColumn}
\title{(Internal) Validate constraint for completeness of its required attribute column}
\usage{
validateFullColumn(x, attrib, class_name)
}
\arguments{
\item{x}{a \code{\link{data.frame}} row containing a single constraint data.}

\item{attrib}{an \code{\linkS4class{item_attrib}} object or a \code{\linkS4class{st_attrib}} object.}

\item{class_name}{\code{item_attrib} or \code{st_attrib}.}
}
\value{
\code{\link{validateFullColumn}} does not return any values;
it \code{\link{stop}}s if the input constraint is not valid.
}
\description{
\code{\link{validateFullColumn}} is an internal function for
validating a constraint for whether its required attribute column is complete (i.e., does not have NA values).
}
\keyword{internal}
