% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_functions.R
\name{validateLBUB}
\alias{validateLBUB}
\title{(Internal) Validate constraint lower/upper bounds}
\usage{
validateLBUB(x, allow_range = TRUE)
}
\arguments{
\item{x}{a \code{\link{data.frame}} row containing a single constraint data.}

\item{allow_range}{whether to allow unequal LB and UB values as valid. (default = \code{TRUE})}
}
\value{
\code{\link{validateLBUB}} does not return any values;
it \code{\link{stop}}s if the input constraint is not valid.
}
\description{
\code{\link{validateLBUB}} is an internal function for
validating a constraint's lower/upper bounds.
}
\keyword{internal}
