% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_functions.R
\name{validateSolver}
\alias{validateSolver}
\title{(Internal) Validate solver for interactive use}
\usage{
validateSolver(config, constraints, purpose = NULL)
}
\arguments{
\item{config}{a \code{\linkS4class{config_Shadow}} or a \code{\linkS4class{config_Static}} object.}

\item{constraints}{a \code{\linkS4class{constraints}} object representing test specifications. Use \code{\link{loadConstraints}} for this.}

\item{purpose}{the purpose of the task. If \code{SPLIT} then extra checks are performed.}
}
\value{
\code{\link{validateSolver}} returns \code{TRUE} or \code{FALSE}.
}
\description{
\code{\link{validateSolver}} is an internal function for
validating whether the solver is appropriate for the task.
}
\keyword{internal}
