% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{addTrans}
\alias{addTrans}
\title{Add transparancy to color}
\usage{
addTrans(color, alpha)
}
\arguments{
\item{color}{A vector of color names or RGB color codes.}

\item{alpha}{A vector of integers between 0 and 255 (0 = fully transparent, 255 = fully visible).}
}
\description{
Add transparancy to color.
}
