% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_control_functions.R
\name{aggregateUsageMatrix}
\alias{aggregateUsageMatrix}
\title{(Internal) Aggregate item usage matrix into exposure rate table}
\usage{
aggregateUsageMatrix(usage_matrix, simulation_constants, constraints)
}
\arguments{
\item{usage_matrix}{a (\emph{nj}, \emph{nv}) matrix containing item usage.}

\item{simulation_constants}{a named list containing simulation constants.}

\item{constraints}{a \code{\linkS4class{constraints}} object.}
}
\value{
\code{\link{aggregateUsageMatrix}} returns a \code{\link{data.frame}} containing exposure rates.
}
\description{
\code{\link{aggregateUsageMatrix}} is an internal function for
aggregating item usage matrix into an exposure rate table.
}
\keyword{internal}
