% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eligibility_functions.R
\name{applyEligibilityConstraintsToXdata}
\alias{applyEligibilityConstraintsToXdata}
\title{(Internal) Augment constraint matrix-data with eligibility constraints}
\usage{
applyEligibilityConstraintsToXdata(
  xdata,
  eligibility_flag_in_current_theta_segment,
  simulation_constants,
  constraints
)
}
\arguments{
\item{xdata}{the constriant matrix-data.}

\item{eligibility_flag_in_current_theta_segment}{a list containing eligibility flags for the current theta segment.}

\item{simulation_constants}{a named list containing simulation constants.}

\item{constraints}{a \code{\linkS4class{constraints}} object representing test specifications. Use \code{\link{loadConstraints}} for this.}
}
\value{
\code{\link{applyEligibilityConstraintsToXdata}} returns an updated constriant matrix-data.
}
\description{
\code{\link{applyEligibilityConstraintsToXdata}} is an internal function for augmenting constraint matrix-data with eligibility constraints.
The function marks items/sets marked as ineligibile to be formally excluded in the constraint matrix-data.
}
\keyword{internal}
