% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_1pl}
\alias{array_info_1pl}
\title{Calculate the Fisher information for a vector of theta values according to the 1pl model}
\usage{
array_info_1pl(x, b)
}
\arguments{
\item{x}{a vector of theta value}

\item{b}{a length-one numeric vector for the difficulty parameter}
}
\description{
Calculate the Fisher information for a vector of theta values according to the 1pl model
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}
}
