% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_1pl}
\alias{array_info_1pl}
\title{Calculate Fisher information at multiple thetas (1PL)}
\usage{
array_info_1pl(x, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{b}{Numeric. A difficulty parameter value.}
}
\description{
Calculate the Fisher information at theta values according to the 1PL model.
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}
}
