% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_3pl}
\alias{array_info_3pl}
\title{Calculate Fisher information at multiple thetas (3PL)}
\usage{
array_info_3pl(x, a, b, c)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A difficulty parameter value.}

\item{c}{Numeric. A guessing parameter value.}
}
\description{
Calculate the Fisher information at theta values according to the 3PL model.
}
\references{
{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}
}
