% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_gpc}
\alias{array_info_gpc}
\title{Calculate Fisher information at multiple thetas (GPC)}
\usage{
array_info_gpc(x, a, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the Fisher information at theta values according to the generalized partial credit model.
}
\references{
{
  \insertRef{muraki_generalized_1992}{TestDesign}
}
}
