% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_gr}
\alias{array_info_gr}
\title{Calculate Fisher information at multiple thetas (GR)}
\usage{
array_info_gr(x, a, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of category boundary parameter values.}
}
\description{
Calculate the Fisher information at theta values according to the graded response model.
}
\references{
{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
