% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_info_pc}
\alias{array_info_pc}
\title{Calculate Fisher information at multiple thetas (PC)}
\usage{
array_info_pc(x, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the Fisher information at theta values according to the partial credit model.
}
\references{
{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}
}
