% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_p_1pl}
\alias{array_p_1pl}
\title{Calculate the probability of a correct response according to 1PL model}
\usage{
array_p_1pl(x, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{b}{Numeric. A difficulty parameter value.}
}
\description{
Calculate the probability of a correct response according to 1PL model.
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}
}
