% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_p_2pl}
\alias{array_p_2pl}
\title{Calculate probability at multiple thetas (2PL)}
\usage{
array_p_2pl(x, a, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A difficulty parameter value.}
}
\description{
Calculate the probability of correct response at theta values, under the 2PL model.
}
\references{
{
  \insertRef{lord_theory_1952}{TestDesign}

  \insertRef{birnbaum_efficient_1957}{TestDesign}

  \insertRef{birnbaum_estimation_1958}{TestDesign}

  \insertRef{birnbaum_further_1958}{TestDesign}
}
}
