% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_p_gpc}
\alias{array_p_gpc}
\title{Calculate probability at multiple thetas (GPC)}
\usage{
array_p_gpc(x, a, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the probability of correct response at theta values, under the generalized partial credit model.
}
\references{
{
  \insertRef{muraki_generalized_1992}{TestDesign}
}
}
