% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{array_p_gr}
\alias{array_p_gr}
\title{Calculate probability at multiple thetas (GR)}
\usage{
array_p_gr(x, a, b)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of category boundary parameter values.}
}
\description{
Calculate the probability of correct response at theta values, under the graded response model.
}
\references{
{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
