% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_method.R
\docType{methods}
\name{ATA}
\alias{ATA}
\title{An S4 generic and its methods for Automated Test Assembly (ATA).}
\usage{
ATA(config, Constraints, plot, plotrange = c(-3, 3))
}
\arguments{
\item{config}{An \code{\linkS4class{ATA.config}} object containing configuration options.}

\item{Constraints}{A list representing optimization constraints. Use \code{\link{LoadConstraints}} for this.}

\item{plot}{Logical. Draws Fisher information plot from the selected items.}

\item{plotrange}{Numeric. A vector of length 2 containing the lower and upper bounds of plot range. Default is \code{c(-3, 3)}.}
}
\description{
An S4 generic and its methods for Automated Test Assembly (ATA).
}
