% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_functions.R
\docType{methods}
\name{ATA}
\alias{ATA}
\alias{ATA,ATA.config-method}
\title{Run Automated Test Assembly}
\usage{
ATA(config, constraints, plot = FALSE, plot.range = c(-3, 3))

\S4method{ATA}{ATA.config}(config, constraints, plot = FALSE,
  plot.range = c(-3, 3))
}
\arguments{
\item{config}{An \code{\linkS4class{ATA.config}} object containing configuration options. Use \code{\link{createStaticTestConfig}} for this.}

\item{constraints}{A list representing optimization constraints. Use \code{\link{LoadConstraints}} for this.}

\item{plot}{Logical. If \code{TRUE}, draw Fisher information plot from the selected items.}

\item{plot.range}{Numeric. A vector of length 2 containing the lower and upper bounds of plot range. The default is \code{c(-3, 3)}.}
}
\value{
A list containing the following entries:
\itemize{
  \item{\code{MIP}} A list containing the result from MIP solver.
  \itemize{
    \item{\code{solution}} Solution vector. Each value represents an item. A value of 1 indicates the item was selected.
    \item{\code{objval}} Objective value of the solution.
    \item{\code{status}} Status value indicating whether an optimal solution was found.
  }
  \item{\code{Selected}} The attributes of the selected items.
  \item{\code{solver}} The name of the MIP solver used in the assembly.
  \item{\code{obj.value}} Objective value of the solution. Identical to the one above.
  \item{\code{solve.time}} The elapsed time in running the solver.
}
}
\description{
Perform Automated Test Assembly with specified configurations.
}
