% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loading_functions.R
\name{buildConstraints}
\alias{buildConstraints}
\title{Build constraints}
\usage{
buildConstraints(pool, file.constraints, file.item.attrib,
  file.st.attrib = NULL)
}
\arguments{
\item{pool}{An \code{item.pool} object. Use \code{\link{loadItemPool}} for this.}

\item{file.constraints}{Character. The name of the file containing constraint specifications.}

\item{file.item.attrib}{Character. The name of the file containing item attributes.}

\item{file.st.attrib}{(Optional) Character. The name of the file containing set attributes.}
}
\value{
A list containing the parsed constraints, to be used in \code{\link{ATA}} and \code{\link{Shadow}}.
}
\description{
Read constraints from specified files.
}
