% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loading_functions.R
\name{BuildConstraints}
\alias{BuildConstraints}
\title{Build constraints}
\usage{
BuildConstraints(pool, file.Constraints, file.ItemAttrib,
  file.StAttrib = NULL)
}
\arguments{
\item{pool}{An \code{item.pool} object. Use \code{\link{LoadItemPool}} for this.}

\item{file.Constraints}{Character. The name of the file containing constraint specifications.}

\item{file.ItemAttrib}{Character. The name of the file containing item attributes.}

\item{file.StAttrib}{(Optional) Character. The name of the file containing set attributes.}
}
\value{
A list containing the parsed constraints, to be used in \code{\link{ATA}} and \code{\link{Shadow}}.
}
\description{
Read constraints from specified files.
}
