% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_functions.R
\name{BuildConstraints}
\alias{BuildConstraints}
\title{BuildConstraints}
\usage{
BuildConstraints(pool, file.Constraints, file.ItemAttrib,
  file.StAttrib = NULL)
}
\arguments{
\item{pool}{An \code{item.pool} object.}

\item{file.Constraints}{Character. The name of the file containing specifications for constraints.}

\item{file.ItemAttrib}{Character. The name of the file containing item attributes.}

\item{file.StAttrib}{Character. The name of the file containing set attributes.}
}
\value{
A list containing the parsed constraints. See th output section of \link{LoadConstraints} for details.
}
\description{
Read constraints from specified files.
}
