% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_info_EB}
\alias{calc_info_EB}
\title{Calculate the Fisher information using empirical Bayes}
\usage{
calc_info_EB(x, item_parm, ncat, model)
}
\arguments{
\item{x}{A numeric vector of MCMC sampled theta values.}

\item{item_parm}{A numeric matrix of item parameters.}

\item{ncat}{a numeric vector specifying the number of response categories in each item.}

\item{model}{a numeric vector indicating the IRT models of each item (1: 1PL, 2: 2PL, 3: 3PL, 4: PC, 5: GPC, 6: GR).}
}
\description{
Calculate the Fisher information using empirical Bayes.
}
