% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_info_FB}
\alias{calc_info_FB}
\title{calc_info_FB}
\usage{
calc_info_FB(x, items_list, ncat, model, useEAP = FALSE)
}
\arguments{
\item{x}{A nice parameter}

\item{items_list}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}

\item{useEAP}{A nice parameter}
}
\description{
calc_info_FB
}
