% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_info_FB}
\alias{calc_info_FB}
\title{Calculate the Fisher information using full Bayesian}
\usage{
calc_info_FB(x, items_list, ncat, model, useEAP = FALSE)
}
\arguments{
\item{x}{A numeric vector of MCMC sampled theta values.}

\item{items_list}{A list of item parameter matrices.}

\item{ncat}{a numeric vector specifying the number of response categories in each item.}

\item{model}{a numeric vector indicating the IRT models of each item (1: 1PL, 2: 2PL, 3: 3PL, 4: PC, 5: GPC, 6: GR).}

\item{useEAP}{\code{TRUE} to use the mean of MCMC theta draws.}
}
\description{
Calculate the Fisher information using full Bayesian.
}
