% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_info_matrix}
\alias{calc_info_matrix}
\title{Calculate the Fisher information matrix for a vector of theta values and a set of items, potentially with a mixture of different models}
\usage{
calc_info_matrix(x, item_parm, ncat, model)
}
\arguments{
\item{x}{Numeric. A vector of theta values.}

\item{item_parm}{A matrix of item parameters.}

\item{ncat}{A numeric vector of the number of response categories by item.}

\item{model}{A numeric vector indicating the IRT models of each item (1: 1PL, 2: 2PL, 3: 3PL, 4: PC, 5: GPC, 6: GR).}
}
\description{
Calculate the Fisher information matrix for a vector of theta values and a set of items, potentially with a mixture of different models
}
