% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_likelihood}
\alias{calc_likelihood}
\title{calc_likelihood}
\usage{
calc_likelihood(x, item_parm, resp, ncat, model)
}
\arguments{
\item{x}{A nice parameter}

\item{item_parm}{A nice parameter}

\item{resp}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}
}
\description{
calc_likelihood
}
