% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_likelihood_function}
\alias{calc_likelihood_function}
\title{calc_likelihood_function}
\usage{
calc_likelihood_function(theta_grid, item_parm, resp, ncat, model)
}
\arguments{
\item{theta_grid}{A nice parameter}

\item{item_parm}{A nice parameter}

\item{resp}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}
}
\description{
calc_likelihood_function
}
