% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_likelihood_function}
\alias{calc_likelihood_function}
\title{Calculate a likelihood function of theta}
\usage{
calc_likelihood_function(theta_grid, item_parm, resp, ncat, model)
}
\arguments{
\item{theta_grid}{An equi-spaced grid of theta values.}

\item{item_parm}{A numeric matrix of item parameters.}

\item{resp}{A numeric vector of item responses.}

\item{ncat}{A numeric vector of the number of response categories by item.}

\item{model}{A numeric vector indicating the IRT models of each item (1: 1PL, 2: 2PL, 3: 3PL, 4: PC, 5: GPC, 6: GR).}
}
\description{
Calculate a likelihood function of theta.
}
