% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_log_likelihood}
\alias{calc_log_likelihood}
\title{Calculate a log-likelihood value of theta}
\usage{
calc_log_likelihood(x, item_parm, resp, ncat, model, prior, prior_parm)
}
\arguments{
\item{x}{a length-one numeric vector for a theta value}

\item{item_parm}{a numeric matrix of item parameters}

\item{resp}{a numeric vector of item responses}

\item{ncat}{a numeric vector of the number of response categories by item}

\item{model}{a numeric vector of the IRT model by item (1: 1pl, 2: 2pl, 3: 3pl, 4: pc, 5: gpc, 6: gr)}

\item{prior}{prior distribution (1: normal, 2: uniform)}

\item{prior_parm}{a numeric vector of hyper parameters for the prior distribution, c(mu, sigma) or c(ll, ul)}
}
\description{
Calculate a log-likelihood value of theta
}
