% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_MI_FB}
\alias{calc_MI_FB}
\title{Calculate the mutual information using full Bayesian}
\usage{
calc_MI_FB(x, items_list, ncat, model)
}
\arguments{
\item{x}{A numeric vector of MCMC sampled theta values.}

\item{items_list}{A list of item parameter matrices.}

\item{ncat}{a numeric vector specifying the number of response categories in each item.}

\item{model}{a numeric vector indicating the IRT models of each item (1: 1PL, 2: 2PL, 3: 3PL, 4: PC, 5: GPC, 6: GR).}
}
\description{
Calculate the mutual information using full Bayesian.
}
