% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_method.R
\docType{methods}
\name{calcDerivative2}
\alias{calcDerivative2}
\alias{calcDerivative2,item.1pl,numeric-method}
\alias{calcDerivative2,item.2pl,numeric-method}
\alias{calcDerivative2,item.3pl,numeric-method}
\alias{calcDerivative2,item.pc,numeric-method}
\alias{calcDerivative2,item.gpc,numeric-method}
\alias{calcDerivative2,item.gr,numeric-method}
\alias{calcDerivative2,item.pool,numeric-method}
\alias{calcDerivative2,pool.cluster,numeric-method}
\title{An S4 generic and its methods to calculate the second derivative of the probability function}
\usage{
calcDerivative2(object, theta)

\S4method{calcDerivative2}{item.1pl,numeric}(object, theta)

\S4method{calcDerivative2}{item.2pl,numeric}(object, theta)

\S4method{calcDerivative2}{item.3pl,numeric}(object, theta)

\S4method{calcDerivative2}{item.pc,numeric}(object, theta)

\S4method{calcDerivative2}{item.gpc,numeric}(object, theta)

\S4method{calcDerivative2}{item.gr,numeric}(object, theta)

\S4method{calcDerivative2}{item.pool,numeric}(object, theta)

\S4method{calcDerivative2}{pool.cluster,numeric}(object, theta)
}
\arguments{
\item{object}{An instance of an item class}

\item{theta}{A vector of theta values}
}
\value{
Second derivative values
}
\description{
An S4 generic and its methods to calculate the second derivative of the probability function
}
\examples{
item.1 = new("item.1pl", difficulty = 0.5)
dd.item.1 = calcDerivative2(item.1, seq(-3, 3, 1))
item.2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
dd.item.2 = calcDerivative2(item.2, seq(-3, 3, 1))
item.3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
dd.item.3 = calcDerivative2(item.3, seq(-3, 3, 1))
item.4 = new("item.pc", threshold = c(-1, 0, 1), ncat = 4)
dd.item.4 = calcDerivative2(item.4, seq(-3, 3, 1))
item.5 = new("item.gpc", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
dd.item.5 = calcDerivative2(item.5, seq(-3, 3, 1))
item.6 = new("item.gr", slope = 0.9, category = c(-1, 0 , 1), ncat = 4)
dd.item.6 = calcDerivative2(item.6, seq(-3, 3, 1))
\dontrun{
itemPool.1 = LoadItemPool("C:/item_par.csv")
dd.itemPool.1 = calcDerivative2(itemPool.1, seq(-3, 3, 1))
}
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{IRTclass}
}

{
  \insertRef{lord_theory_1952}{IRTclass}

  \insertRef{birnbaum_efficient_1957}{IRTclass}

  \insertRef{birnbaum_estimation_1958}{IRTclass}

  \insertRef{birnbaum_further_1958}{IRTclass}
}

{
  \insertRef{birnbaum_latent_1968}{IRTclass}
}

{
  \insertRef{masters_rasch_1982}{IRTclass}

  \insertRef{andrich_rating_1978}{IRTclass}
}

{
  \insertRef{muraki_generalized_1992}{IRTclass}
}

{
  \insertRef{samejima_estimation_1969}{IRTclass}
}
}
