% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_functions.R
\docType{methods}
\name{calcEscore}
\alias{calcEscore}
\alias{calcEscore,item.1pl,numeric-method}
\alias{calcEscore,item.2pl,numeric-method}
\alias{calcEscore,item.3pl,numeric-method}
\alias{calcEscore,item.pc,numeric-method}
\alias{calcEscore,item.gpc,numeric-method}
\alias{calcEscore,item.gr,numeric-method}
\alias{calcEscore,item.pool,numeric-method}
\alias{calcEscore,pool.cluster,numeric-method}
\title{Calculate expected scores}
\usage{
calcEscore(object, theta)

\S4method{calcEscore}{item.1pl,numeric}(object, theta)

\S4method{calcEscore}{item.2pl,numeric}(object, theta)

\S4method{calcEscore}{item.3pl,numeric}(object, theta)

\S4method{calcEscore}{item.pc,numeric}(object, theta)

\S4method{calcEscore}{item.gpc,numeric}(object, theta)

\S4method{calcEscore}{item.gr,numeric}(object, theta)

\S4method{calcEscore}{item.pool,numeric}(object, theta)

\S4method{calcEscore}{pool.cluster,numeric}(object, theta)
}
\arguments{
\item{object}{An instance of an item class.}

\item{theta}{A vector of theta values.}
}
\value{
A vector of expected scores of length nq (the number of values on theta grid).
}
\description{
An S4 generic and its methods to calculate expected scores given a vector of thetas for different item classes.
}
\examples{
item.1 = new("item.1pl", difficulty = 0.5)
ICC.item.1 = calcEscore(item.1, seq(-3, 3, 1))
item.2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
ICC.item.2 = calcEscore(item.2, seq(-3, 3, 1))
item.3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
ICC.item.3 = calcEscore(item.3, seq(-3, 3, 1))
item.4 = new("item.pc", threshold = c(-1, 0, 1), ncat = 4)
ICC.item.4 = calcEscore(item.4, seq(-3, 3, 1))
item.5 = new("item.gpc", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
ICC.item.5 = calcEscore(item.5, seq(-3, 3, 1))
item.6 = new("item.gr", slope = 0.9, category = c(-1, 0 , 1), ncat = 4)
ICC.item.6 = calcEscore(item.6, seq(-3, 3, 1))
TCC.itempool = calcEscore(itempool.science, seq(-3, 3, 1))
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}

{
  \insertRef{lord_theory_1952}{TestDesign}

  \insertRef{birnbaum_efficient_1957}{TestDesign}

  \insertRef{birnbaum_estimation_1958}{TestDesign}

  \insertRef{birnbaum_further_1958}{TestDesign}
}

{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}

{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}

{
  \insertRef{muraki_generalized_1992}{TestDesign}
}

{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
