% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_method.R
\docType{methods}
\name{calcEscore,pool.cluster,numeric-method}
\alias{calcEscore,pool.cluster,numeric-method}
\title{calcEscore}
\usage{
\S4method{calcEscore}{pool.cluster,numeric}(object, theta)
}
\arguments{
\item{object}{An instance of an item class}

\item{theta}{A vector of theta values}
}
\description{
calcEscore
}
\examples{
\dontrun{
itemPool.1 = LoadItemPool("C:/item_par.csv")
#this needs to be updated for pool.cluster
TCC.itemPool.1 = calcEscore(itemPools, seq(-3, 3, 1))
}
}
