% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_functions.R
\docType{methods}
\name{calcHessian}
\alias{calcHessian}
\alias{calcHessian,item.1pl,numeric,numeric-method}
\alias{calcHessian,item.1pl,numeric-method}
\alias{calcHessian,item.2pl,numeric,numeric-method}
\alias{calcHessian,item.2pl,numeric-method}
\alias{calcHessian,item.3pl,numeric,numeric-method}
\alias{calcHessian,item.3pl,numeric-method}
\alias{calcHessian,item.pc,numeric,numeric-method}
\alias{calcHessian,item.pc,numeric-method}
\alias{calcHessian,item.gpc,numeric,numeric-method}
\alias{calcHessian,item.gpc,numeric-method}
\alias{calcHessian,item.gr,numeric,numeric-method}
\alias{calcHessian,item.gr,numeric-method}
\alias{calcHessian,item.pool,numeric,numeric-method}
\alias{calcHessian,item.pool,numeric-method}
\alias{calcHessian,pool.cluster,numeric,list-method}
\alias{calcHessian,pool.cluster,numeric-method}
\title{Calculate second derivative of log-likelihood}
\usage{
calcHessian(object, theta, resp)

\S4method{calcHessian}{item.1pl,numeric,numeric}(object, theta, resp)

\S4method{calcHessian}{item.2pl,numeric,numeric}(object, theta, resp)

\S4method{calcHessian}{item.3pl,numeric,numeric}(object, theta, resp)

\S4method{calcHessian}{item.pc,numeric,numeric}(object, theta, resp)

\S4method{calcHessian}{item.gpc,numeric,numeric}(object, theta, resp)

\S4method{calcHessian}{item.gr,numeric,numeric}(object, theta, resp)

\S4method{calcHessian}{item.pool,numeric,numeric}(object, theta, resp)

\S4method{calcHessian}{pool.cluster,numeric,list}(object, theta, resp)
}
\arguments{
\item{object}{An instance of an item class.}

\item{theta}{A vector of theta values.}

\item{resp}{Response data.}
}
\value{
Second derivative values of log-likelihoods.
}
\description{
An S4 generic and its methods to calculate the second derivative of the log-likelihood function.
}
\examples{
\dontrun{
item.1 = new("item.1pl", difficulty = 0.5)
h.item.1 = calcHessian(item.1, seq(-3, 3, 1), 0)
}
\dontrun{
item.2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
h.item.2 = calcHessian(item.2, seq(-3, 3, 1), 0)
}
\dontrun{
item.3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
h.item.3 = calcHessian(item.3, seq(-3, 3, 1), 0)
}
\dontrun{
item.4 = new("item.pc", threshold = c(-1, 0, 1), ncat = 4)
h.item.4 = calcHessian(item.4, seq(-3, 3, 1), 0)
}
\dontrun{
item.5 = new("item.gpc", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
h.item.5 = calcHessian(item.5, seq(-3, 3, 1), 0)
}
\dontrun{
item.6 = new("item.gr", slope = 0.9, category = c(-1, 0 , 1), ncat = 4)
h.item.6 = calcHessian(item.6, seq(-3, 3, 1), 0)
}
\dontrun{
itemPool.1 = LoadItemPool("C:/item_par.csv")
h.itemPool.1 = calcHessian(itemPool.1, seq(-3, 3, 1), 0)
}
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}

{
  \insertRef{lord_theory_1952}{Shadow}

  \insertRef{birnbaum_efficient_1957}{Shadow}

  \insertRef{birnbaum_estimation_1958}{Shadow}

  \insertRef{birnbaum_further_1958}{Shadow}
}

{
  \insertRef{birnbaum_latent_1968}{Shadow}
}

{
  \insertRef{masters_rasch_1982}{Shadow}

  \insertRef{andrich_rating_1978}{Shadow}
}

{
  \insertRef{muraki_generalized_1992}{Shadow}
}

{
  \insertRef{samejima_estimation_1969}{Shadow}
}
}
