% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_functions.R
\docType{methods}
\name{calcProb}
\alias{calcProb}
\alias{calcProb,item.1pl,numeric-method}
\alias{calcProb,item.2pl,numeric-method}
\alias{calcProb,item.3pl,numeric-method}
\alias{calcProb,item.pc,numeric-method}
\alias{calcProb,item.gpc,numeric-method}
\alias{calcProb,item.gr,numeric-method}
\alias{calcProb,item.pool,numeric-method}
\alias{calcProb,pool.cluster,numeric-method}
\title{Calculate item response probabilities}
\usage{
calcProb(object, theta)

\S4method{calcProb}{item.1pl,numeric}(object, theta)

\S4method{calcProb}{item.2pl,numeric}(object, theta)

\S4method{calcProb}{item.3pl,numeric}(object, theta)

\S4method{calcProb}{item.pc,numeric}(object, theta)

\S4method{calcProb}{item.gpc,numeric}(object, theta)

\S4method{calcProb}{item.gr,numeric}(object, theta)

\S4method{calcProb}{item.pool,numeric}(object, theta)

\S4method{calcProb}{pool.cluster,numeric}(object, theta)
}
\arguments{
\item{object}{An instance of an item class.}

\item{theta}{A vector of theta values.}
}
\value{
A matrix of probability values with a dimension (nq, ncat) for a single item or a list of matrices for an instance of "item.pool".
}
\description{
An S4 generic and its methods to calculate item response probabilities for different item classes
}
\examples{
item_1 = new("item.1pl", difficulty = 0.5)
prob_item_1 = calcProb(item_1, seq(-3, 3, 1))
item_2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
prob_item_2 = calcProb(item_2, seq(-3, 3, 1))
item_3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
prob_item_3 = calcProb(item_3, seq(-3, 3, 1))
item_4 = new("item.pc", threshold = c(-1, 0, 1), ncat = 4)
prob_item_4 = calcProb(item_4, seq(-3, 3, 1))
item_5 = new("item.gpc", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
prob_item_5 = calcProb(item_5, seq(-3, 3, 1))
item_6 = new("item.gr", slope = 0.9, category = c(-1, 0, 1), ncat = 4)
prob_item_6 = calcProb(item_6, seq(-3, 3, 1))
prob_itempool <- calcProb(itempool_science, seq(-3, 3, 1))
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}

{
  \insertRef{lord_theory_1952}{TestDesign}

  \insertRef{birnbaum_efficient_1957}{TestDesign}

  \insertRef{birnbaum_estimation_1958}{TestDesign}

  \insertRef{birnbaum_further_1958}{TestDesign}
}

{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}

{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}

{
  \insertRef{muraki_generalized_1992}{TestDesign}
}

{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
