% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_method.R
\docType{methods}
\name{calcProb}
\alias{calcProb}
\alias{calcProb,item.1pl,numeric-method}
\alias{calcProb,item.2pl,numeric-method}
\alias{calcProb,item.3pl,numeric-method}
\alias{calcProb,item.pc,numeric-method}
\alias{calcProb,item.gpc,numeric-method}
\alias{calcProb,item.gr,numeric-method}
\alias{calcProb,item.pool,numeric-method}
\alias{calcProb,pool.cluster,numeric-method}
\title{An S4 generic and its methods to calculate item response probabilities for different item classes}
\usage{
calcProb(object, theta)

\S4method{calcProb}{item.1pl,numeric}(object, theta)

\S4method{calcProb}{item.2pl,numeric}(object, theta)

\S4method{calcProb}{item.3pl,numeric}(object, theta)

\S4method{calcProb}{item.pc,numeric}(object, theta)

\S4method{calcProb}{item.gpc,numeric}(object, theta)

\S4method{calcProb}{item.gr,numeric}(object, theta)

\S4method{calcProb}{item.pool,numeric}(object, theta)

\S4method{calcProb}{pool.cluster,numeric}(object, theta)
}
\arguments{
\item{object}{An instance of an item class}

\item{theta}{A vector of theta values}
}
\value{
A matrix of probability values with a dimension (nq, ncat) for a single item or a list of matrices for an instance of "item.pool"
}
\description{
An S4 generic and its methods to calculate item response probabilities for different item classes
}
\examples{
item.1 = new("item.1pl", difficulty = 0.5)
prob.item.1 = calcProb(item.1, seq(-3, 3, 1))
item.2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
prob.item.2 = calcProb(item.2, seq(-3, 3, 1))
item.3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
prob.item.3 = calcProb(item.3, seq(-3, 3, 1))
item.4 = new("item.pc", threshold = c(-1, 0, 1), ncat = 4)
prob.item.4 = calcProb(item.4, seq(-3, 3, 1))
item.5 = new("item.gpc", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
prob.item.5 = calcProb(item.5, seq(-3, 3, 1))
item.6 = new("item.gr", slope = 0.9, category = c(-1, 0 , 1), ncat = 4)
prob.item.6 = calcProb(item.6, seq(-3, 3, 1))
\dontrun{
itemPool.1 = LoadItemPool("C:/item_par.csv")
prob.itemPool.1 = calcProb(itemPool.1, seq(-3, 3, 1))
}
\dontrun{
itemPool.1 = LoadItemPool("C:/item_par_1.csv")
itemPool.2 = LoadItemPool("C:/item_par_2.csv")
itemPools = vector(mode = "list", 2)
itemPools@pools[[1]] = itemPool.1
itemPools@pools[[2]] = itemPool.2
prob.itemPools = calcProb(itemPools, seq(-3, 3, 1))
}
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}

{
  \insertRef{lord_theory_1952}{Shadow}
  
  \insertRef{birnbaum_efficient_1957}{Shadow}
  
  \insertRef{birnbaum_estimation_1958}{Shadow}
  
  \insertRef{birnbaum_further_1958}{Shadow}
}

{
  \insertRef{birnbaum_latent_1968}{Shadow}
}

{
  \insertRef{masters_rasch_1982}{Shadow}
  
  \insertRef{andrich_rating_1978}{Shadow}
}

{
  \insertRef{muraki_generalized_1992}{Shadow}
}

{
  \insertRef{samejima_estimation_1969}{Shadow}
}
}
