% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{calcRP}
\alias{calcRP}
\title{calcRP}
\usage{
calcRP(object, rp = 0.5, maxIter = 100, conv = 1e-04,
  startTheta = 0)
}
\arguments{
\item{object}{A nice parameter}

\item{rp}{A nice parameter}

\item{maxIter}{A nice parameter}

\item{conv}{A nice parameter}

\item{startTheta}{A nice parameter}
}
\description{
calcRP
}
