% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{calcRP}
\alias{calcRP}
\title{Find matching theta to supplied probability}
\usage{
calcRP(object, rp = 0.5, max_iter = 100, conv = 1e-04, start_theta = 0)
}
\arguments{
\item{object}{An \code{\linkS4class{item_pool}} object.}

\item{rp}{A response probability value.}

\item{max_iter}{A maximum number of iterations.}

\item{conv}{A convergence criterion.}

\item{start_theta}{A starting theta value.}
}
\description{
Find theta corresponding to a response probability value for each item.
}
