% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_adaptivity_measures.r
\name{calculateAdaptivityMeasures}
\alias{calculateAdaptivityMeasures}
\title{Calculate Adaptivity Measures}
\usage{
calculateAdaptivityMeasures(x)
}
\arguments{
\item{x}{an \code{\linkS4class{output_Shadow_all}} object.}
}
\value{
\code{\link{calculateAdaptivityMeasures}} returns a named list:
\itemize{
  \item{\code{corr} the correlation between final theta estimates and average test locations.}
  \item{\code{ratio} the ratio of (1) standard deviation of average test locations, versus (2) standard deviation of final theta estimates.}
  \item{\code{PRV} the proportion of variance reduced, from (1) the variance of item locations of all items in the pool, by (2) the average of test location variances.}
  \item{\code{info} (1) average information of a test at final theta estimate, relative to (2) best average obtainable from item pool using same test length, adjusting for (3) average information from item pool using random selection.}
}
}
\description{
\code{\link{calculateAdaptivityMeasures}} is a function for
calculating commonly used adaptivity measures.
}
