% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{checkConstraints}
\alias{checkConstraints}
\title{Check the consistency of constraints and item usage}
\usage{
checkConstraints(constraints, usage.matrix, true.theta = NULL)
}
\arguments{
\item{constraints}{A list constraints generated by \code{\link{loadConstraints}}.}

\item{usage.matrix}{A matrix of item usage data from \code{\link{Shadow}}.}

\item{true.theta}{A vector of true theta values.}
}
\description{
Check the consistency of constraints and item usage.
}
