% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{checkConstraints}
\alias{checkConstraints}
\title{Check the consistency of constraints and item usage}
\usage{
checkConstraints(constraints, usageMatrix, trueTheta = NULL)
}
\arguments{
\item{constraints}{A list constraints generated by \code{\link{LoadConstraints}}.}

\item{usageMatrix}{A matrix of item usage data from \code{\link{Shadow}}.}

\item{trueTheta}{A vector of true theta values.}
}
\description{
Check the consistency of constraints and item usage.
}
