% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{checkConstraints}
\alias{checkConstraints}
\title{Check the consistency of constraints and item usage}
\usage{
checkConstraints(constraints, usage_matrix, true_theta = NULL)
}
\arguments{
\item{constraints}{A \code{\linkS4class{constraints}} object generated by \code{\link{loadConstraints}}.}

\item{usage_matrix}{A matrix of item usage data from \code{\link{Shadow}}.}

\item{true_theta}{A vector of true theta values.}
}
\description{
Check the consistency of constraints and item usage.
}
