% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\docType{class}
\name{constraint-class}
\alias{constraint-class}
\title{Class 'constraints': a single constraint}
\description{
\code{\linkS4class{constraint}} is an S4 class to represent a single constraint.
}
\section{Slots}{

\describe{
\item{\code{constraint}}{the constraint ID string of the constraint.}

\item{\code{nc}}{the number of MIP-format constraints translated from this constraint.}

\item{\code{mat}}{LHS coefficients to use on decision variables.}

\item{\code{rhs}}{RHS values to use.}

\item{\code{dir}}{the imposed constraint between LHS and RHS.}

\item{\code{suspend}}{\code{TRUE} if the constraint is not to be imposed.}
}}

