% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{constraint-class}
\alias{constraint-class}
\title{constraint}
\description{
Represents a set of constriants.
}
\section{Slots}{

\describe{
\item{\code{CONSTRAINT}}{Character. The index of the constraint set.}

\item{\code{mat}}{A matrix representing the left-hand side weights. Has nc rows.}

\item{\code{dir}}{A vector of length nc. Each entry represents a logical operator relating the left-hand side to the right-hand side.}

\item{\code{rhs}}{A vector of length nc. Each entry represents the right-hand side of the constraint.}

\item{\code{nc}}{Numeric. The number of constraints represented in this object.}
}}

