% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset.fatigue}
\alias{dataset.fatigue}
\alias{itempool.fatigue.raw}
\alias{itemattrib.fatigue.raw}
\alias{itemcontent.fatigue.raw}
\alias{constraints.fatigue.raw}
\alias{resp.fatigue.raw}
\alias{itempool.fatigue}
\alias{itemattrib.fatigue}
\alias{constraints.fatigue}
\title{Fatigue dataset}
\description{
Item-based example pool with item contents (95 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool.fatigue} An \code{\linkS4class{item.pool}} object.
  \item \code{itemattrib.fatigue} A data frame containing item attributes.
  \item \code{constraints.fatigue} A list containing 111 constraints.
}

Also, the following datasets are intended for illustrating expected data structures. See examples below.
\itemize{
  \item \code{itempool.fatigue.raw} Item parameters.
  \item \code{itemattrib.fatigue.raw} Item attributes.
  \item \code{itemcontent.fatigue.raw} Item contents.
  \item \code{constraints.fatigue.raw} Constraints.
  \item \code{resp.fatigue.raw} Raw response data.
}
}
\examples{
\dontrun{
write.csv(itempool.fatigue.raw, "itempool.fatigue.csv", row.names = F)
itempool.fatigue = LoadItemPool("itempool.fatigue.csv")

write.csv(itemattrib.fatigue, "itemattrib.fatigue.csv", row.names = F)
itemattrib.fatigue = LoadItemAttrib("itemattrib.fatigue.csv", itempool.fatigue)

write.csv(constraints.fatigue.raw, "constraints.fatigue.csv", row.names = F)
constraints.fatigue = LoadConstraints("constraints.fatigue.csv",
          itempool.fatigue, itemattrib.fatigue)

write.csv(itemcontent.fatigue.raw, "itemcontent.fatigue.csv", row.names = F)
write.table(resp.fatigue.raw, "resp.fatigue.csv", row.names = F, col.names = F, sep = ",")
}
}
\keyword{datasets}
