% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset.reading}
\alias{dataset.reading}
\alias{itempool.reading.raw}
\alias{itemattrib.reading.raw}
\alias{stimattrib.reading.raw}
\alias{constraints.reading.raw}
\alias{itempool.reading}
\alias{itemattrib.reading}
\alias{stimattrib.reading}
\alias{constraints.reading}
\title{Reading dataset}
\description{
Stimulus-based example item pool (303 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool.reading} An \code{\linkS4class{item.pool}} object.
  \item \code{itemattrib.reading} A data frame containing item attributes.
  \item \code{stimattrib.reading} A data frame containing stimulus attributes.
  \item \code{constraints.reading} A list containing 18 constraints.
}

Also, the following datasets are intended for illustrating expected data structures. See examples below.
\itemize{
  \item \code{itempool.reading.raw} Item parameters.
  \item \code{itemattrib.reading.raw} Item attributes.
  \item \code{stimattrib.reading.raw} Item attributes.
  \item \code{constraints.reading.raw} Constraints.
}
}
\examples{
\dontrun{
write.csv(itempool.reading.raw, "itempool.reading.csv", row.names = FALSE)
itempool.reading <- LoadItemPool("itempool.reading.csv")

write.csv(itemattrib.reading.raw, "itemattrib.reading.csv", row.names = FALSE)
itemattrib.reading <- LoadItemAttrib("itemattrib.reading.csv", itempool.reading)

write.csv(stimattrib.reading.raw, "stimattrib.reading.csv", row.names = FALSE)
stimattrib.reading <- LoadStAttrib("stimattrib.reading.csv", itemattrib.reading)

write.csv(constraints.reading.raw, "constraints.reading.csv", row.names = FALSE)
constraints.reading <- LoadConstraints(
  "constraints.reading.csv",
  itempool.reading, itemattrib.reading, stimattrib.reading
)
}
}
\keyword{datasets}
