% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset.science}
\alias{dataset.science}
\alias{itempool.science.raw}
\alias{itemattrib.science.raw}
\alias{constraints.science.raw}
\alias{itempool.science}
\alias{itemattrib.science}
\alias{constraints.science}
\title{Science dataset}
\description{
Item-based example item pool (1000 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool.science} An \code{\linkS4class{item.pool}} object.
  \item \code{itemattrib.science} A data frame containing item attributes.
  \item \code{constraints.science} A list containing 36 constraints.
}

Also, the following datasets are intended for illustrating expected data structures. See examples below.
\itemize{
  \item \code{itempool.science.raw} Item parameters.
  \item \code{itemattrib.science.raw} Item attributes.
  \item \code{constraints.science.raw} Constraints.
}
}
\examples{
\dontrun{
write.csv(itempool.science.raw, "itempool.science.csv", row.names = FALSE)
itempool.science <- loadItemPool("itempool.science.csv")

write.csv(itemattrib.science.raw, "itemattrib.science.csv", row.names = FALSE)
itemattrib.science <- loadItemAttrib("itemattrib.science.csv", itempool.science)

write.csv(constraints.science.raw, "constraints.science.csv", row.names = FALSE)
constraints.science <- loadConstraints(
  "constraints.science.csv",
  itempool.science, itemattrib.science
)
}
}
\keyword{datasets}
