% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{par_fatigue}
\alias{par_fatigue}
\alias{item_attrib_fatigue}
\alias{item_content_fatigue}
\alias{constraints_fatigue}
\alias{raw_fatigue}
\title{par_fatigue}
\description{
An item-based example pool with item contents.
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{par_fatigue} Item parameters.
  \item \code{item_attrib_fatigue} Item attributes.
  \item \code{item_content_fatigue} Item contents.
  \item \code{constraints_fatigue} A constraint set (111 constraints).
  \item \code{raw_fatigue} Raw response data.
}
}
\examples{
\dontrun{
write.csv(par_fatigue, "par_fatigue.csv", row.names = F)
itempool.fatigue = LoadItemPool("par_fatigue.csv")

write.csv(item_attrib_fatigue, "item_attrib_fatigue.csv", row.names = F)
itemattrib.fatigue = LoadItemAttrib("item_attrib_fatigue.csv", itempool.fatigue)

write.csv(constraints_fatigue, "constraints_fatigue.csv", row.names = F)
constraints.fatigue = LoadConstraints("constraints_fatigue.csv", itempool.fatigue, itemattrib.fatigue)
}
}
\keyword{datasets}
