% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_fatigue}
\alias{dataset_fatigue}
\alias{itempool_fatigue_raw}
\alias{itemattrib_fatigue_raw}
\alias{itemcontent_fatigue_raw}
\alias{constraints_fatigue_raw}
\alias{resp_fatigue_raw}
\alias{itempool_fatigue}
\alias{itemattrib_fatigue}
\alias{constraints_fatigue}
\title{Fatigue dataset}
\description{
Item-based example pool with item contents (95 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{itempool_fatigue} An \code{\linkS4class{item_pool}} object.
  \item \code{itemattrib_fatigue} A data frame containing item attributes.
  \item \code{constraints_fatigue} A list containing 111 constraints.
}

Also, the following datasets are intended for illustrating expected data structures. See examples below.
\itemize{
  \item \code{itempool_fatigue_raw} Item parameters.
  \item \code{itemattrib_fatigue_raw} Item attributes.
  \item \code{itemcontent_fatigue_raw} Item contents.
  \item \code{constraints_fatigue_raw} Constraints.
  \item \code{resp_fatigue_raw} Raw response data.
}
}
\examples{
\dontrun{
write.csv(itempool_fatigue_raw, "itempool_fatigue.csv", row.names = FALSE)
itempool_fatigue <- loadItemPool("itempool_fatigue.csv")

write.csv(itemattrib_fatigue, "itemattrib_fatigue.csv", row.names = FALSE)
itemattrib_fatigue <- loadItemAttrib("itemattrib_fatigue.csv", itempool_fatigue)

write.csv(constraints_fatigue_raw, "constraints_fatigue.csv", row.names = FALSE)
constraints_fatigue <- loadConstraints(
  "constraints_fatigue.csv",
  itempool_fatigue, itemattrib_fatigue
)

write.csv(itemcontent_fatigue_raw, "itemcontent_fatigue.csv", row.names = FALSE)
write.table(resp_fatigue_raw, "resp_fatigue.csv", row.names = FALSE, col.names = FALSE, sep = ",")
}
}
\keyword{datasets}
