% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{par_reading}
\alias{par_reading}
\alias{item_attrib_reading}
\alias{stimulus_attrib_reading}
\alias{constraints_reading}
\title{par_reading}
\description{
Stimulus-based example item pool.
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{par_reading} Item parameters.
  \item \code{item_attrib_reading} Item attributes.
  \item \code{stimulus_attrib_reading} Stimulus attributes.
  \item \code{constraints_reading} A constraint set (18 constraints).
}
}
\examples{
\dontrun{
write.csv(par_reading, "par_reading.csv", row.names = F)
itempool.reading = LoadItemPool("par_reading.csv")

write.csv(item_attrib_reading, "item_attrib_reading.csv", row.names = F)
itemattrib.reading = LoadItemAttrib("item_attrib_reading.csv", itempool.reading)

write.csv(stimulus_attrib_reading, "stimulus_attrib_reading.csv", row.names = F)
stimattrib.reading = LoadStAttrib("stimulus_attrib_reading.csv", itemattrib.reading)

write.csv(constraints_reading, "constraints_reading.csv", row.names = F)
constraints.reading = LoadConstraints("constraints_reading.csv",
          itempool.reading, itemattrib.reading, stimattrib.reading)
}
}
\keyword{datasets}
