% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{par_science}
\alias{par_science}
\alias{item_attrib_science}
\alias{constraints_science}
\title{par_science}
\description{
Item-based example item pool (1000 items).
}
\details{
This pool is associated with the following objects:
\itemize{
  \item \code{par_science} Item parameters.
  \item \code{item_attrib_science} Item attributes.
  \item \code{constraints_science} A constraint set (36 constraints).
}
}
\examples{
\dontrun{
write.csv(par_science, "par_science.csv", row.names = F)
itempool.science = LoadItemPool("par_science.csv")

write.csv(item_attrib_science, "item_attrib_science.csv", row.names = F)
itemattrib.science = LoadItemAttrib("item_attrib_science.csv", itempool.science)

write.csv(constraints_science, "constraints_science.csv", row.names = F)
constraints.science = LoadConstraints("constraints_science.csv", itempool.science, itemattrib.science)
}
}
\keyword{datasets}
