% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_functions.R
\docType{methods}
\name{eap}
\alias{eap}
\alias{eap,item_pool-method}
\title{Compute expected a posteriori estimates of theta}
\usage{
eap(object, theta, prior, resp, select = NULL)

\S4method{eap}{item_pool}(object, theta, prior, resp, select = NULL)
}
\arguments{
\item{object}{an \code{\linkS4class{item_pool}} object.}

\item{theta}{the theta grid to use as quadrature points.}

\item{prior}{A prior distribution, a numeric vector for a common prior or a matrix for individualized priors.}

\item{resp}{A numeric matrix of item responses, one row per examinee.}

\item{select}{A vector of indices identifying the items to subset.}
}
\description{
\code{\link{eap}} is a function to compute expected a posteriori estimates of theta.
}
