% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{EAP}
\alias{EAP}
\alias{EAP,test-method}
\alias{EAP,test.cluster-method}
\title{EAP}
\usage{
EAP(object, prior, select = NULL, resetPrior = FALSE)

\S4method{EAP}{test}(object, prior, select = NULL, resetPrior = FALSE)

\S4method{EAP}{test.cluster}(object, prior, select = NULL,
  resetPrior = FALSE)
}
\arguments{
\item{object}{A nice parameter}

\item{prior}{A nice parameter}

\item{select}{A nice parameter}

\item{resetPrior}{A nice parameter}
}
\description{
Generate expected a posteriori estimates of theta

Genereate expected a posteriori estimates of theta
}
