% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{EAP}
\alias{EAP}
\alias{EAP,test-method}
\alias{EAP,test.cluster-method}
\title{Generate expected a posteriori estimates of theta}
\usage{
EAP(object, prior, select = NULL, resetPrior = FALSE)

\S4method{EAP}{test}(object, prior, select = NULL, resetPrior = FALSE)

\S4method{EAP}{test.cluster}(object, prior, select = NULL,
  resetPrior = FALSE)
}
\arguments{
\item{object}{An S4 object of \code{\linkS4class{test}} or \code{\linkS4class{test.cluster}}.}

\item{prior}{A prior distribution, a numeric vector for a common prior or a matrix for individualized priors.}

\item{select}{A vector of indices identifying the items to subset.}

\item{resetPrior}{TRUE to reset the prior distribution for each test when object is of class \code{\linkS4class{test.cluster}}.}
}
\description{
Generate expected a posteriori estimates of theta.
}
