% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_segment}
\alias{find_segment}
\title{Find the segment to which each theta value belongs}
\usage{
find_segment(segment, x)
}
\arguments{
\item{segment}{A numeric vector of segment cuts.}

\item{x}{A numeric vector of theta values.}
}
\description{
Find the segment to which each theta value belongs.
}
