% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cpp_theta_documents.r
\name{find_segment}
\alias{find_segment}
\title{(C++) Classify theta values into segments using cutpoints}
\usage{
find_segment(x, segment)
}
\arguments{
\item{x}{the theta value. This can be a vector.}

\item{segment}{segment cutpoints. Values of \code{-Inf, Inf} are not implied and must be explicitly supplied if intended.}
}
\description{
\code{find_segment()} is a function to classify theta values into segments based on supplied cutpoints.
}
\examples{
cuts <- c(-Inf, -2, 0, 2, Inf)

find_segment(-3, cuts)
find_segment(-1, cuts)
find_segment(1, cuts)
find_segment(3, cuts)
find_segment(seq(-3, 3, 2), cuts)

}
