% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{getSegmentProb}
\alias{getSegmentProb}
\title{(Internal) Convert a theta distribution to segment-wise classification probabilities}
\usage{
getSegmentProb(posterior_sample, simulation_constants)
}
\arguments{
\item{posterior_sample}{a vector containing values in a Markov chain.}

\item{simulation_constants}{a named list containing simulation constants.}
}
\value{
\code{\link{getSegmentProb}} returns a vector of segment-wise classification probabilities.
}
\description{
\code{\link{getSegmentProb}} is an internal function for
converting a theta distribution to segment-wise classification probabilities
This is used for \code{BIGM-BAYESIAN} exposure control.
}
\keyword{internal}
