% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_1pl}
\alias{info_1pl}
\title{Calculate the Fisher information according to the 1pl model}
\usage{
info_1pl(x, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{b}{Numeric. A difficulty parameter value.}
}
\description{
Calculate the Fisher information according to the 1pl model
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}
}
