% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_1pl}
\alias{info_1pl}
\title{Calculate Fisher information at a single theta (1PL)}
\usage{
info_1pl(x, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{b}{Numeric. A difficulty parameter value.}
}
\description{
Calculate the Fisher information at a theta value according to the 1PL model.
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}
}
