% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_3pl}
\alias{info_3pl}
\title{Calculate Fisher information at a single theta (3PL)}
\usage{
info_3pl(x, a, b, c)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A difficulty parameter value.}

\item{c}{Numeric. A guessing parameter value.}
}
\description{
Calculate the Fisher information at a theta value according to the 3PL model.
}
\references{
{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}
}
