% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_gpc}
\alias{info_gpc}
\title{Calculate Fisher information at a single theta (GPC).}
\usage{
info_gpc(x, a, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the Fisher information at a theta value according to the generalizied partial credit model.
}
\references{
{
  \insertRef{muraki_generalized_1992}{TestDesign}
}
}
