% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_gr}
\alias{info_gr}
\title{Calculate the Fisher information according to the graded response model}
\usage{
info_gr(x, a, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of category boundary parameter values.}
}
\description{
Calculate the Fisher information according to the graded response model
}
\references{
{
  \insertRef{samejima_estimation_1969}{Shadow}
}
}
