% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{info_gr}
\alias{info_gr}
\title{Calculate Fisher information at a single theta (GR).}
\usage{
info_gr(x, a, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of category boundary parameter values.}
}
\description{
Calculate the Fisher information at a theta value according to the graded resposne model.
}
\references{
{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
