% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{iparPosteriorSample}
\alias{iparPosteriorSample}
\title{Sample item parameter estimates from their posterior distributions}
\usage{
iparPosteriorSample(pool, nSample = 500)
}
\arguments{
\item{pool}{An \code{\linkS4class{item.pool}} object.}

\item{nSample}{An integer as the number of sampled parameters.}
}
\description{
Sample item parameter estimates from their posterior distributions.
}
