% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_functions.R
\name{iparPosteriorSample}
\alias{iparPosteriorSample}
\title{Generate item parameter samples for Bayesian purposes}
\usage{
iparPosteriorSample(pool, n_sample = 500)
}
\arguments{
\item{pool}{an \code{\linkS4class{item_pool}} object.}

\item{n_sample}{the number of samples to draw.}
}
\value{
\code{\link{iparPosteriorSample}} returns a length-\emph{ni} list of item parameter matrices, with each matrix having \code{n_sample} rows.
}
\description{
\code{\link{iparPosteriorSample}} is a function for generating item parameter samples.
Used for the FB (full-Bayesian) estimation method.
}
\examples{
ipar <- iparPosteriorSample(itempool_bayes, 5)
ipar <- iparPosteriorSample(itempool_science, 5) # no variation

}
