% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{iparPosteriorSample}
\alias{iparPosteriorSample}
\title{Sample item parameter estimates from their posterior distributions}
\usage{
iparPosteriorSample(pool, n_sample = 500)
}
\arguments{
\item{pool}{An \code{\linkS4class{item_pool}} object.}

\item{n_sample}{An integer as the number of sampled parameters.}
}
\description{
Sample item parameter estimates from their posterior distributions.
}
\examples{
ipar <- iparPosteriorSample(itempool_science, 5)

}
