% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.1pl-class}
\alias{item.1pl-class}
\title{An S4 class to represent a 1PL item}
\description{
An S4 class to represent a 1PL item.
}
\section{Slots}{

\describe{
\item{\code{difficulty}}{Numeric. A difficulty parameter value.}
}}

\examples{
item.1 = new("item.1pl", difficulty = 0.5)
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}
}
