% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.gr-class}
\alias{item.gr-class}
\title{An S4 class to represent a graded response item}
\description{
An S4 class to represent a graded response item.
}
\section{Slots}{

\describe{
\item{\code{slope}}{Numeric. A slope parameter value.}

\item{\code{category}}{Numeric. A vector of category boundary values.}

\item{\code{ncat}}{Numeric. The number of response categories.}
}}

\examples{
item.6 = new("item.gr", slope = 1.0, category = c(-2.0, -1.0, 0, 1.0, 2.0), ncat = 6)
}
\references{
{
  \insertRef{samejima_estimation_1969}{oat}
}
}
