% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.pc-class}
\alias{item.pc-class}
\title{An S4 class to represent a partial credit item}
\description{
An S4 class to represent a partial credit item.
}
\section{Slots}{

\describe{
\item{\code{threshold}}{Numeric. A vector of threshold parameter values.}

\item{\code{ncat}}{Numeric. The number of response categories.}
}}

\examples{
item.4 = new("item.pc", threshold = c(-0.5, 0.5), ncat = 3)
}
\references{
{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}
}
