% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_class.R
\docType{class}
\name{item.pool-class}
\alias{item.pool-class}
\title{An S4 class to represent an item pool}
\description{
An S4 class to represent an item pool
}
\section{Slots}{

\describe{
\item{\code{ni}}{a numeric vector of length one to denote the number of items in the item pool}

\item{\code{maxCat}}{a numeric vector of length one to denote the maximum number of response categories across all items in the pool}

\item{\code{index}}{a numeric vector of item sequence numbers}

\item{\code{ID}}{a character vector of item IDs}

\item{\code{model}}{a numeric vector of IRT models by item (1: item.1pl, 2: item.2pl, 3: item.3pl, 4: item.pc, 5: item.gpc, 6: item.gr)}

\item{\code{NCAT}}{a numeric vector of the number of response categories by item}

\item{\code{parms}}{a list of item parameters in the pool}

\item{\code{ipar}}{a matrix of item parameters in the pool}
}}

\examples{
\dontrun{
itemPool.1 = LoadItemPool("C:/itemPool.csv")
}
}
