% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\docType{class}
\name{item_pool-class}
\alias{item_pool-class}
\title{Class 'item_pool': an item pool}
\description{
\code{\linkS4class{item_pool}} is an S4 class to represent an item pool.
}
\details{
See \code{\link{item_pool-operators}} for object manipulation functions.
}
\section{Slots}{

\describe{
\item{\code{ni}}{the number of items in the pool.}

\item{\code{max_cat}}{the maximum number of response categories across the pool.}

\item{\code{index}}{the numeric index of each item.}

\item{\code{id}}{the ID string of each item.}

\item{\code{model}}{the item class name of each item. See \code{\link{item-classes}}.}

\item{\code{NCAT}}{the number of response categories of each item.}

\item{\code{parms}}{a list containing item class objects. See \code{\link{item-classes}}.}

\item{\code{ipar}}{a matrix containing item parameters.}

\item{\code{se}}{a matrix containing item parameter standard errors.}

\item{\code{raw}}{the raw input \code{\link{data.frame}} used in \code{\link{loadItemPool}} to create this object.}

\item{\code{raw_se}}{the raw input \code{\link{data.frame}} used in \code{\link{loadItemPool}} to create this object.}
}}

