% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{lnHyperPars}
\alias{lnHyperPars}
\title{Calculate hyperparameters for log-normal distribution}
\usage{
lnHyperPars(mean, sd)
}
\arguments{
\item{mean}{Mean of the distribution.}

\item{sd}{Standard deviation of the distribution.}
}
\description{
Calculate hyperparameters for log-normal distribution.
}
\examples{
lnHyperPars(.5, 1)

}
