% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{loadConstraints}
\alias{loadConstraints}
\title{Load constraints}
\usage{
loadConstraints(file, pool, item.attrib, st.attrib = NULL)
}
\arguments{
\item{file}{Character. The name of the file containing specifications for constraints.}

\item{pool}{An \code{item.pool} object.}

\item{item.attrib}{A \code{data.frame} containing item attributes. Use \code{\link{loadItemAttrib}} for this.}

\item{st.attrib}{(Optional) A \code{data.frame} containing stimulus attributes. Use \code{\link{loadStAttrib}} for this.}
}
\value{
A list containing the parsed constraints, to be used in \code{\link{ATA}} and \code{\link{Shadow}}.
}
\description{
Read constraints from specified file.
}
\details{
Use \code{vignette("constraints")} for instructions on how to create a constraint set object.
}
\seealso{
\link{dataset_science} for example usage.
}
