% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loading_functions.R
\name{LoadConstraints}
\alias{LoadConstraints}
\title{Load constraints}
\usage{
LoadConstraints(file.csv, pool, ItemAttrib, StAttrib = NULL)
}
\arguments{
\item{file.csv}{Character. The name of the file containing specifications for constraints.}

\item{pool}{An \code{item.pool} object.}

\item{ItemAttrib}{A \code{data.frame} containing item attributes. Use \code{\link{LoadItemAttrib}} for this.}

\item{StAttrib}{(Optional) A \code{data.frame} containing stimulus attributes. Use \code{\link{LoadStAttrib}} for this.}
}
\value{
A list containing the parsed constraints, to be used in \code{\link{ATA}} and \code{\link{Shadow}}.
}
\description{
Read constraints from specified file.
}
\seealso{
\link{dataset.science} for example usage.
}
