% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{loadConstraints}
\alias{loadConstraints}
\title{Load constraints}
\usage{
loadConstraints(object, pool, item_attrib, st_attrib = NULL, file = NULL)
}
\arguments{
\item{object}{Constraint specifications. Can be a file path of a .csv file, or a data.frame. See vignette for the expected format.}

\item{pool}{An \code{item_pool} object. Use \code{\link{loadItemPool}} for this.}

\item{item_attrib}{An \code{item_attrib} object containing item attributes. Use \code{\link{loadItemAttrib}} for this.}

\item{st_attrib}{(Optional) An \code{st_attrib} object containing stimulus attributes. Use \code{\link{loadStAttrib}} for this.}

\item{file}{(Deprecated) Use 'object' above.}
}
\value{
A \code{constraints} object containing the parsed constraints, to be used in \code{\link{Static}} and \code{\link{Shadow}}.
}
\description{
Read constraints from specified file.
}
\details{
Use \code{vignette("constraints")} for instructions on how to create the constraint file.
}
\examples{
## Read from data.frame:
itempool_science    <- loadItemPool(itempool_science_data)
itemattrib_science  <- loadItemAttrib(itemattrib_science_data, itempool_science)
constraints_science <- loadConstraints(constraints_science_data,
  itempool_science, itemattrib_science)

## Read from file: write to tempdir() for illustration and clean afterwards
f <- file.path(tempdir(), "constraints_science.csv")
write.csv(constraints_science_data, f, row.names = FALSE)
constraints_science <- loadConstraints(f,
  itempool_science, itemattrib_science)
file.remove(f)

}
\seealso{
\link{dataset_science} for example usage.
}
