% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\docType{class}
\name{item_attrib-class}
\alias{item_attrib-class}
\alias{loadItemAttrib}
\title{Load item attributes}
\usage{
loadItemAttrib(object, pool, file = NULL)
}
\arguments{
\item{object}{Item attributes. Can be a data.frame or the file path of a .csv file. The content should at least include column 'ID' that matches the item pool.}

\item{pool}{An \code{\linkS4class{item_pool}} object. Use \code{\link{loadItemPool}} for this.}

\item{file}{(Deprecated) Use 'object' above.}
}
\value{
An \code{\linkS4class{item_attrib}} object.
}
\description{
Read item attributes from specified file.
}
\examples{
## Read from data.frame:
itempool_science   <- loadItemPool(itempool_science_data)
itemattrib_science <- loadItemAttrib(itemattrib_science_data, itempool_science)

## Read from file: write to tempdir() for illustration and clean afterwards
f <- file.path(tempdir(), "itemattrib_science.csv")
write.csv(itemattrib_science_data, f, row.names = FALSE)
itemattrib_science <- loadItemAttrib(f, itempool_science)
file.remove(f)

## TestDesign 1.1.0 - Deprecated arguments
\dontrun{
loadItemAttrib(object = "iatt.csv", pool) # is equivalent to
loadItemAttrib(file   = "iatt.csv", pool) # pre 1.1.0
}

}
\seealso{
\link{dataset_science} for example usage.
}
