% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{loadItemAttrib}
\alias{loadItemAttrib}
\title{Load item attributes}
\usage{
loadItemAttrib(file, pool)
}
\arguments{
\item{file}{Character. The name of the file containing item attributes.}

\item{pool}{An \code{\linkS4class{item_pool}} object. Use \code{\link{loadItemPool}} for this.}
}
\value{
A \code{data.frame} containing parsed dataset.
}
\description{
Read item attributes from specified file.
}
\examples{
\donttest{
itempool_science <- loadItemPool("itempool_science.csv")
itemattrib_science <- loadItemAttrib("itemattrib_science.csv", itempool_science)
}
}
\seealso{
\link{dataset_science} for example usage.
}
