% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\docType{class}
\name{item_attrib-class}
\alias{item_attrib-class}
\alias{loadItemAttrib}
\title{Load item attributes}
\usage{
loadItemAttrib(object, pool, file = NULL)
}
\arguments{
\item{object}{Item attributes. Can be a file path of a .csv file, or a data.frame. The content should at least include the column 'ID' matching to the item pool.}

\item{pool}{An \code{\linkS4class{item_pool}} object. Use \code{\link{loadItemPool}} for this.}

\item{file}{(Deprecated) Use 'object' above.}
}
\value{
An \code{\linkS4class{item_attrib}} object.
}
\description{
Read item attributes from specified file.
}
\examples{
## Read from data.frame:
itempool_science   <- loadItemPool(itempool_science_raw)
itemattrib_science <- loadItemAttrib(itemattrib_science_raw, itempool_science)

## Read from file: write to tempdir() for illustration and clean afterwards
f <- file.path(tempdir(), "itemattrib_science.csv")
write.csv(itemattrib_science_raw, f, row.names = FALSE)
itemattrib_science <- loadItemAttrib(f, itempool_science)
file.remove(f)

}
\seealso{
\link{dataset_science} for example usage.
}
