% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loading_functions.R
\name{loadItemAttrib}
\alias{loadItemAttrib}
\title{Load item attributes}
\usage{
loadItemAttrib(file, pool)
}
\arguments{
\item{file}{Character. The name of the file containing item attributes.}

\item{pool}{An \code{\linkS4class{item.pool}} object. Use \code{\link{LoadItemPool}} for this.}
}
\value{
A \code{data.frame} containing parsed dataset.
}
\description{
Read item attributes from specified file.
}
\seealso{
\link{dataset.science} for example usage.
}
