% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_functions.R
\name{LoadItemAttrib}
\alias{LoadItemAttrib}
\title{Load item attributes}
\usage{
LoadItemAttrib(file.csv, pool)
}
\arguments{
\item{file.csv}{Character. The name of the file containing item attributes.}

\item{pool}{An \code{item.pool} object. Use \code{\link{LoadItemPool}} for this.}
}
\value{
A \code{data.frame} containing parsed dataset.
}
\description{
Read item attributes from specified file.
}
