% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loading_functions.R
\name{LoadItemPool}
\alias{LoadItemPool}
\title{Load item paramaters}
\usage{
LoadItemPool(file.csv, ipar = NULL, se.file.csv = NULL)
}
\arguments{
\item{file.csv}{File path of a .csv file containing item parameters. The file content should not have column names.}

\item{ipar}{A data.frame created from a .csv file.}

\item{se.file.csv}{File path of a .csv file containing standard errors.}
}
\value{
An \linkS4class{item.pool} object.
}
\description{
Read item parameters from a .csv file or a data.frame and create an \linkS4class{item.pool} class.
}
\seealso{
\link{dataset.science} for example usage.
}
