% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{loadItemPool}
\alias{loadItemPool}
\title{Load item paramaters}
\usage{
loadItemPool(ipar, ipar_se = NULL, file = NULL, se_file = NULL)
}
\arguments{
\item{ipar}{Item parameters. Can be a file path of a .csv file, or a data.frame. The content should at least include columns 'ID' and 'MODEL'.}

\item{ipar_se}{(Optional) Standard errors. Can be a file path of a .csv file, or a data.frame.}

\item{file}{(Deprecated) Use 'ipar' above.}

\item{se_file}{(Deprecated) Use 'ipar_se' above.}
}
\value{
An \linkS4class{item_pool} object.
}
\description{
Read item parameters from a data.frame or a .csv file and create an \linkS4class{item_pool} class.
}
\examples{
## Read from data.frame:
itempool_science <- loadItemPool(itempool_science_data)

## Read from file: write to tempdir() for illustration and clean afterwards
f <- file.path(tempdir(), "itempool_science.csv")
write.csv(itempool_science_data, f, row.names = FALSE)
itempool_science <- loadItemPool(f)
file.remove(f)

}
\seealso{
\link{dataset_science} for example usage.
}
