% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadItemPool.R
\name{LoadItemPool}
\alias{LoadItemPool}
\title{Load item parameters from a .csv file or a data.frame and create an item.pool class}
\usage{
LoadItemPool(file.csv, ipar = NULL, se.file.csv = NULL)
}
\arguments{
\item{file.csv}{File path of a .csv file containing item parameters. The file content should not have column names.}

\item{ipar}{A data.frame created from a .csv file.}

\item{se.file.csv}{File path of a .csv file containing standard errors.}
}
\value{
An \linkS4class{item.pool} object.
}
\description{
Load item parameters from a .csv file or a data.frame and create an item.pool class
}
\examples{
 
\dontrun{
itemPool.1 = LoadItemPool("C:/sample.csv")
ipar.1 = read.csv("C:/sample.csv")
itemPool.1 = LoadItemPool(ipar = ipar.1)
}
}
