% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{loadStAttrib}
\alias{loadStAttrib}
\title{Load set/stimulus/passage attributes}
\usage{
loadStAttrib(file, item_attrib)
}
\arguments{
\item{file}{Character. The name of the file containing item attributes.}

\item{item_attrib}{A \code{data.frame} containing item attributes. Use \code{\link{loadItemAttrib}} for this.}
}
\value{
A \code{data.frame} containing stimulus attributes.
}
\description{
Read set, stimulus, or passage attributes from specified file.
}
\seealso{
\link{dataset_reading} for example usage.
}
