% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\docType{class}
\name{st_attrib-class}
\alias{st_attrib-class}
\alias{loadStAttrib}
\title{Load set/stimulus/passage attributes}
\usage{
loadStAttrib(object, item_attrib, file = NULL)
}
\arguments{
\item{object}{Set attributes. Can be a file path of a .csv file, or a data.frame. The content should at least include the column 'STID' referring to the same column in item attributes.}

\item{item_attrib}{An \code{\linkS4class{item_attrib}} object containing item attributes. Use \code{\link{loadItemAttrib}} for this.}

\item{file}{(Deprecated) Use 'object' above.}
}
\value{
A \code{\linkS4class{st_attrib}} object containing stimulus attributes.
}
\description{
Read set, stimulus, or passage attributes from specified file.
}
\examples{
## Read from data.frame:
itempool_reading   <- loadItemPool(itempool_reading_data)
itemattrib_reading <- loadItemAttrib(itemattrib_reading_data, itempool_reading)
stimattrib_reading <- loadStAttrib(stimattrib_reading_data, itemattrib_reading)

## Read from file: write to tempdir() for illustration and clean afterwards
f <- file.path(tempdir(), "stimattrib_reading.csv")
write.csv(stimattrib_reading_data, f, row.names = FALSE)
stimattrib_reading <- loadStAttrib(f, itemattrib_reading)
file.remove(f)

## TestDesign 1.1.0 - Deprecated arguments
\dontrun{
loadStAttrib(object = "satt.csv", item_attrib) # is equivalent to
loadStAttrib(file   = "satt.csv", item_attrib) # pre 1.1.0
}

}
\seealso{
\link{dataset_reading} for example usage.
}
