% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_functions.R
\name{LoadStAttrib}
\alias{LoadStAttrib}
\title{Load set/stimulus/passage attributes}
\usage{
LoadStAttrib(file.csv, ItemAttrib)
}
\arguments{
\item{file.csv}{Character. The name of the file containing item attributes.}

\item{ItemAttrib}{A \code{data.frame} containing item attributes. Use \code{\link{LoadItemAttrib}} for this.}
}
\value{
A \code{data.frame} containing stimulus attributes.
}
\description{
Read set attributes from specified file.
}
