% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{logitHyperPars}
\alias{logitHyperPars}
\title{Calculate hyperparameters for logit-normal distribution}
\usage{
logitHyperPars(mean, sd)
}
\arguments{
\item{mean}{Mean of the distribution.}

\item{sd}{Standard deviation of the distribution.}
}
\description{
Calculate hyperparameters for logit-normal distribution.
}
\examples{
logitHyperPars(.5, 1)

}
