% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{MakeItemPoolCluster}
\alias{MakeItemPoolCluster}
\title{MakeItemPoolCluster}
\usage{
MakeItemPoolCluster(pools, names = NULL)
}
\arguments{
\item{pools}{A list of \code{\linkS4class{pool}} objects}

\item{names}{An optional vector of \code{\linkS4class{pool}} names}
}
\description{
Create a \code{\linkS4class{pool.cluster}} object
}
