% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{makeItemPoolCluster}
\alias{makeItemPoolCluster}
\title{Create an item pool cluster object}
\usage{
makeItemPoolCluster(pools, names = NULL)
}
\arguments{
\item{pools}{A list of \code{\linkS4class{item.pool}} objects.}

\item{names}{An optional vector of \code{\linkS4class{item.pool}} names.}
}
\description{
Create a \code{\linkS4class{pool.cluster}} object.
}
