% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{makeTest}
\alias{makeTest}
\alias{makeTest,item.pool-method}
\title{Generate a test object}
\usage{
makeTest(object, theta = seq(-4, 4, 0.1), infoType = "FISHER",
  trueTheta = NULL)

\S4method{makeTest}{item.pool}(object, theta = seq(-4, 4, 0.1),
  infoType = "FISHER", trueTheta = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{item.pool}} object.}

\item{theta}{A grid of theta values.}

\item{infoType}{An information type.}

\item{trueTheta}{An optional vector of true theta values to simulate response data.}
}
\description{
Generate a \code{\linkS4class{test}} object
}
\examples{
test <- makeTest(itempool.science, seq(-3, 3, 1))
}
