% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\docType{methods}
\name{makeTest}
\alias{makeTest}
\alias{makeTest,item.pool-method}
\title{Generate a test object}
\usage{
makeTest(object, theta = seq(-4, 4, 0.1), info.type = "FISHER",
  true.theta = NULL)

\S4method{makeTest}{item.pool}(object, theta = seq(-4, 4, 0.1),
  info.type = "FISHER", true.theta = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{item.pool}} object.}

\item{theta}{A grid of theta values.}

\item{info.type}{An information type.}

\item{true.theta}{An optional vector of true theta values to simulate response data.}
}
\description{
Generate a \code{\linkS4class{test}} object
}
\examples{
test <- makeTest(itempool_science, seq(-3, 3, 1))
}
