% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{MakeTest}
\alias{MakeTest}
\alias{MakeTest,item.pool-method}
\title{An S4 Generic}
\usage{
MakeTest(object, theta, infoType = "FISHER", trueTheta = NULL)

\S4method{MakeTest}{item.pool}(object, theta, infoType = "FISHER",
  trueTheta = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{item.pool}} object}

\item{theta}{A grid of theta values}

\item{infoType}{An information type}

\item{trueTheta}{An optional vector of true theta values to simulate response data}
}
\description{
Generate an \code{\linkS4class{test}} object
}
