% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_operators.R
\docType{methods}
\name{makeTest}
\alias{makeTest}
\alias{makeTest,item_pool-method}
\title{Generate a test object}
\usage{
makeTest(
  object,
  theta = seq(-4, 4, 0.1),
  info_type = "FISHER",
  true_theta = NULL
)

\S4method{makeTest}{item_pool}(
  object,
  theta = seq(-4, 4, 0.1),
  info_type = "FISHER",
  true_theta = NULL
)
}
\arguments{
\item{object}{An \code{\linkS4class{item_pool}} object.}

\item{theta}{A grid of theta values.}

\item{info_type}{An information type.}

\item{true_theta}{An optional vector of true theta values to simulate response data.}
}
\description{
Generate a \code{\linkS4class{test}} object
}
\examples{
test <- makeTest(itempool_science, seq(-3, 3, 1))
}
