% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{MakeTestCluster}
\alias{MakeTestCluster}
\alias{MakeTestCluster,pool.cluster,numeric,numeric-method}
\alias{MakeTestCluster,pool.cluster,numeric,list-method}
\title{An S4 Generic}
\usage{
MakeTestCluster(object, theta, trueTheta)

\S4method{MakeTestCluster}{pool.cluster,numeric,numeric}(object, theta,
  trueTheta)

\S4method{MakeTestCluster}{pool.cluster,numeric,list}(object, theta,
  trueTheta)
}
\arguments{
\item{object}{A nice parameter}

\item{theta}{A nice parameter}

\item{trueTheta}{A nice parameter}
}
\description{
MakeTestCluster

MakeTestCluster
}
