% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_cache_operators.r
\docType{methods}
\name{makeTestCluster}
\alias{makeTestCluster}
\alias{makeTestCluster,item_pool_cluster,numeric,numeric-method}
\alias{makeTestCluster,item_pool_cluster,numeric,list-method}
\title{Create a test cluster object}
\usage{
makeTestCluster(object, theta, true_theta)

\S4method{makeTestCluster}{item_pool_cluster,numeric,numeric}(object, theta, true_theta)

\S4method{makeTestCluster}{item_pool_cluster,numeric,list}(object, theta, true_theta)
}
\arguments{
\item{object}{an \code{\linkS4class{item_pool_cluster}} object.}

\item{theta}{a grid of theta values.}

\item{true_theta}{an optional vector of true theta values to simulate response data.}
}
\description{
\code{\link{makeTestCluster}} is a function for creating a \code{\linkS4class{test_cluster}} object.
This is used to make all necessary data (e.g., item information, response data) prior to the main simulation.
This function is only kept for backwards compatibility.
}
