% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\docType{methods}
\name{makeTestCluster}
\alias{makeTestCluster}
\alias{makeTestCluster,item_pool_cluster,numeric,numeric-method}
\alias{makeTestCluster,item_pool_cluster,numeric,list-method}
\title{Generate a test cluster object}
\usage{
makeTestCluster(object, theta, true_theta)

\S4method{makeTestCluster}{item_pool_cluster,numeric,numeric}(object, theta, true_theta)

\S4method{makeTestCluster}{item_pool_cluster,numeric,list}(object, theta, true_theta)
}
\arguments{
\item{object}{An \code{\linkS4class{item_pool_cluster}} object}

\item{theta}{A grid of theta values}

\item{true_theta}{An optional vector of true theta values to simulate response data}
}
\description{
Generate a \code{\linkS4class{test_cluster}} object
}
