% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{maxinfoplot}
\alias{maxinfoplot}
\title{maxinfoplot}
\usage{
maxinfo_plot(pool, constraints, theta = seq(-3, 3, 0.5))
}
\arguments{
\item{pool}{An \code{\linkS4class{item.pool}} object}

\item{constraints}{A list constraints generated by \code{\link{LoadConstraints}}}

\item{theta}{A theta grid}
}
\description{
Draw a plot of maximum attainable information given the constraints imposed
}
