% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{MLE}
\alias{MLE}
\alias{MLE,test-method}
\title{An S4 Generic}
\usage{
MLE(object, startTheta = NULL, rangeTheta = c(-4, 4), maxIter = 100,
  crit = 0.001, select = NULL)

\S4method{MLE}{test}(object, startTheta = NULL, rangeTheta = c(-4, 4),
  maxIter = 100, crit = 0.001, select = NULL)
}
\arguments{
\item{object}{A nice parameter}

\item{startTheta}{A nice parameter}

\item{rangeTheta}{A nice parameter}

\item{maxIter}{A nice parameter}

\item{crit}{A nice parameter}

\item{select}{A nice parameter}
}
\description{
MLE
}
