% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{MLE}
\alias{MLE}
\alias{MLE,test-method}
\title{An S4 Generic}
\usage{
MLE(object, startTheta = NULL, rangeTheta = c(-4, 4), maxIter = 100,
  crit = 0.001, select = NULL)

\S4method{MLE}{test}(object, startTheta = NULL, rangeTheta = c(-4, 4),
  maxIter = 100, crit = 0.001, select = NULL)
}
\arguments{
\item{object}{A \code{\linkS4class{test}} object}

\item{startTheta}{An optional vector of start theta values}

\item{rangeTheta}{Lower and upper limit of theta values}

\item{maxIter}{Maximum number of iterations}

\item{crit}{Convergence criterion}

\item{select}{A vector of item indexes to subset}
}
\description{
Generate maximum likelihood estimates of theta
}
