% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_class.R
\docType{class}
\name{output_Shadow-class}
\alias{output_Shadow-class}
\title{output_Shadow}
\description{
output_Shadow
}
\section{Slots}{

\describe{
\item{\code{simulee_id}}{Numeric. The index of the simulee.}

\item{\code{true_theta}}{Numeric or NULL. True theta value of the simulee if supplied in advance.}

\item{\code{true_theta_segment}}{Numeric or NULL. Which segment the true theta value is in.}

\item{\code{final_theta_est}}{Numeric. The estimated theta after the last administered item.}

\item{\code{final_se_est}}{Numeric. The standard error of estimation after the last administered item.}

\item{\code{administered_item_index}}{Numeric. A vector of item indices administered at each position.}

\item{\code{administered_item_resp}}{Numeric. A vector of responses at each position.}

\item{\code{administered_stimulus_index}}{Numeric. A vector of stimulus indices administered at each position.}

\item{\code{shadow_test_refreshed}}{Logical. A vector of logical values indicating whether the shadow test was refreshed before administering an item at each position.}

\item{\code{shadow_test_feasible}}{Logical. A vector of logical values indicating whether a feasible solution to the shadow test was available in each position.}

\item{\code{solve_time}}{Numeric. A vector of values indicating the time taken in obtaining a shadow test.}

\item{\code{interim_theta_est}}{Numeric. A vector containing estimated thetas at each position.}

\item{\code{interim_se_est}}{Numeric. A vector containing standard errors at each position.}

\item{\code{theta_segment_index}}{Numeric. A vector containing which segments the estimated thetas were in at each position.}

\item{\code{prior}}{Numeric. A prior distribution.}

\item{\code{prior_par}}{Numeric. The hyper parameters for the prior distribution.}

\item{\code{posterior}}{Numeric. A posterior distribution.}

\item{\code{posterior_sample}}{Numeric. A vector containing MCMC samples.}

\item{\code{likelihood}}{Numeric. A likelihood distribution.}

\item{\code{shadow_test}}{A list of vectors containing item indices of the shadow test at each position.}
}}

