% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_class.R
\docType{class}
\name{output_Shadow_all-class}
\alias{output_Shadow_all-class}
\title{Class 'output_Shadow_all': a set of adaptive assembly solutions}
\description{
\code{\linkS4class{output_Shadow_all}} is an S4 class to represent a set of adaptive assembly solutions.
}
\section{Slots}{

\describe{
\item{\code{output}}{list of \code{\linkS4class{output_Shadow}} objects, containing the assembly results for each participant.}

\item{\code{final_theta_est}}{final theta estimates for each participant.}

\item{\code{final_se_est}}{standard errors of final theta estimates for each participant.}

\item{\code{exposure_rate}}{Exposure rate of each item in the pool.}

\item{\code{usage_matrix}}{The matrix representing which items were used in each item position.}

\item{\code{true_segment_count}}{foo}

\item{\code{est_segment_count}}{foo}

\item{\code{eligibility_stats}}{foo}

\item{\code{check_eligibility_stats}}{foo}

\item{\code{no_fading_eligibility_stats}}{foo}

\item{\code{freq_infeasible}}{foo}

\item{\code{pool}}{the \code{\linkS4class{item_pool}} used in the assembly.}

\item{\code{config}}{the \code{\linkS4class{config_Shadow}} used in the assembly.}

\item{\code{constraints}}{the \code{\linkS4class{constraints}} used in the assembly.}

\item{\code{true_theta}}{the \code{true_theta} argument used in the assembly.}

\item{\code{data}}{the \code{data} argument used in the assembly.}

\item{\code{prior}}{the \code{prior} argument used in the assembly.}

\item{\code{prior_par}}{the \code{prior_par} argument used in the assembly.}
}}

