% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_1pl}
\alias{p_1pl}
\title{Calculate probability at a single theta (1PL)}
\usage{
p_1pl(x, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{b}{Numeric. A difficulty parameter value.}
}
\description{
Calculate the probability of correct response at a theta value, under the 1PL model.
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{TestDesign}
}
}
