% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_2pl}
\alias{p_2pl}
\title{Calculate probability at a single theta (2PL)}
\usage{
p_2pl(x, a, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A difficulty parameter value.}
}
\description{
Calculate the probability of correct response at a theta value, under the 2PL model.
}
\references{
{
  \insertRef{lord_theory_1952}{TestDesign}

  \insertRef{birnbaum_efficient_1957}{TestDesign}

  \insertRef{birnbaum_estimation_1958}{TestDesign}

  \insertRef{birnbaum_further_1958}{TestDesign}
}
}
