% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_3pl}
\alias{p_3pl}
\title{Calculate probability at a single theta (3PL)}
\usage{
p_3pl(x, a, b, c)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{the slope parameter value.}

\item{b}{the difficulty parameter value.}

\item{c}{the guessing parameter value.}
}
\description{
Calculate the probability of correct response at a theta value, under the 3PL model.
}
\references{
{
  \insertRef{birnbaum_latent_1968}{TestDesign}
}
}
