% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_3pl}
\alias{p_3pl}
\title{Calculate the probability of a correct response according to the 3PL model}
\usage{
p_3pl(x, a, b, c)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A difficulty parameter value.}

\item{c}{Numeric. A guessing parameter value.}
}
\description{
Calculate the probability of a correct response according to the 3PL model.
}
\references{
{
  \insertRef{birnbaum_latent_1968}{Shadow}
}
}
