% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gpc}
\alias{p_gpc}
\title{Calculate probability at a single theta (GPC)}
\usage{
p_gpc(x, a, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the probability of correct response at a theta value, under the generalized partial credit model.
}
\references{
{
  \insertRef{muraki_generalized_1992}{TestDesign}
}
}
