% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_gr}
\alias{p_gr}
\title{Calculate probability at a single theta (GR)}
\usage{
p_gr(x, a, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{a}{Numeric. A slope parameter value.}

\item{b}{Numeric. A vector of category boundary parameter values.}
}
\description{
Calculate the probability of correct response at a theta value, under the graded response model.
}
\references{
{
  \insertRef{samejima_estimation_1969}{TestDesign}
}
}
