% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_pc}
\alias{p_pc}
\title{Calculate the response probabilities according to the partial credit model}
\usage{
p_pc(x, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the response probabilities according to the partial credit model.
}
\references{
{
  \insertRef{masters_rasch_1982}{Shadow}
  
  \insertRef{andrich_rating_1978}{Shadow}
}
}
