% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_pc}
\alias{p_pc}
\title{Calculate probability at a single theta (PC)}
\usage{
p_pc(x, b)
}
\arguments{
\item{x}{Numeric. A single theta value.}

\item{b}{Numeric. A vector of threshold parameter values.}
}
\description{
Calculate the probability of correct response at a theta value, under the partial credit model.
}
\references{
{
  \insertRef{masters_rasch_1982}{TestDesign}

  \insertRef{andrich_rating_1978}{TestDesign}
}
}
