% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\docType{methods}
\name{plotCAT}
\alias{plotCAT}
\alias{plotCAT,list-method}
\alias{plotCAT,Shadow.output-method}
\title{Draw an audit trail plot}
\usage{
plotCAT(object, examinee.id = 1, min.theta = -5, max.theta = 5,
  min.score = 0, max.score = 1, z.ci = 1.96, file.pdf = NULL, ...)

\S4method{plotCAT}{list}(object, examinee.id = 1, min.theta = -5,
  max.theta = 5, min.score = 0, max.score = 1, z.ci = 1.96,
  file.pdf = NULL, ...)

\S4method{plotCAT}{Shadow.output}(object, examinee.id = 1,
  min.theta = -5, max.theta = 5, min.score = 0, max.score = 1,
  z.ci = 1.96, file.pdf = NULL, ...)
}
\arguments{
\item{object}{An output object generated by \code{\link{Shadow}}.}

\item{examinee.id}{Numeric ID of the examinee to draw the plot.}

\item{min.theta}{A lower bound of theta.}

\item{max.theta}{An upper bound of theta.}

\item{min.score}{A minimum item score.}

\item{max.score}{A maximum item score.}

\item{z.ci}{A quantile of the normal distribution for confidence intervals.}

\item{file.pdf}{If supplied a filename, save as a PDF file.}

\item{...}{Additional options to be passed on to \code{pdf()}.}
}
\description{
Draw an audit trail plot.
}
