% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{plotCAT}
\alias{plotCAT}
\alias{plotCAT,list-method}
\alias{plotCAT,Shadow.output-method}
\title{Draw an audit trail plot}
\usage{
plotCAT(object, examineeID = 1, PDF = NULL, minTheta = -5,
  maxTheta = 5, minScore = 0, maxScore = 1, zCI = 1.96)

\S4method{plotCAT}{list}(object, examineeID = 1, PDF = NULL,
  minTheta = -5, maxTheta = 5, minScore = 0, maxScore = 1,
  zCI = 1.96)

\S4method{plotCAT}{Shadow.output}(object, examineeID = 1, PDF = NULL,
  minTheta = -5, maxTheta = 5, minScore = 0, maxScore = 1,
  zCI = 1.96)
}
\arguments{
\item{object}{An output object generated by \code{\link{Shadow}}.}

\item{examineeID}{Numeric ID of the examinee to draw the plot.}

\item{PDF}{If supplied a filename, save as a PDF file.}

\item{minTheta}{A lower bound of theta.}

\item{maxTheta}{An upper bound of theta.}

\item{minScore}{A minimum item score.}

\item{maxScore}{A maximum item score.}

\item{zCI}{A quantile of the normal distribution for confidence intervals.}
}
\description{
Draw an audit trail plot.
}
