% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{plotCAT}
\alias{plotCAT}
\alias{plotCAT,Shadow.output-method}
\alias{plotCAT,list-method}
\title{plotCAT}
\usage{
plotCAT(object, minTheta = -5, maxTheta = 5, minScore = 0,
  maxScore = 1, zCI = 1.96, PDF = NULL)

\S4method{plotCAT}{Shadow.output}(object, minTheta = -5, maxTheta = 5,
  minScore = 0, maxScore = 1, zCI = 1.96, PDF = NULL)

\S4method{plotCAT}{list}(object, minTheta = -5, maxTheta = 5,
  minScore = 0, maxScore = 1, zCI = 1.96, PDF = NULL)
}
\arguments{
\item{object}{An output object generated by \code{STA}}

\item{minTheta}{A lower bound of theta}

\item{maxTheta}{An upper bound of theta}

\item{minScore}{A minimum item score}

\item{maxScore}{A maximum item score}

\item{zCI}{A quantile of the normal distribution for confidence intervals}

\item{PDF}{If supplied a filename, save as a PDF file}
}
\description{
Draws an audit trail plot

plotCAT
}
