% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\docType{methods}
\name{plotCAT}
\alias{plotCAT}
\alias{plotCAT,output_Shadow_all-method}
\alias{plotCAT,list-method}
\alias{plotCAT,output_Shadow-method}
\title{(deprecated) Plot audit trail}
\usage{
plotCAT(
  object,
  examinee_id = 1,
  min_theta = -5,
  max_theta = 5,
  min_score = 0,
  max_score = 1,
  z_ci = 1.96,
  file_pdf = NULL,
  ...
)

\S4method{plotCAT}{output_Shadow_all}(
  object,
  examinee_id = 1,
  min_theta = -5,
  max_theta = 5,
  min_score = 0,
  max_score = 1,
  z_ci = 1.96,
  file_pdf = NULL,
  ...
)

\S4method{plotCAT}{list}(
  object,
  examinee_id = 1,
  min_theta = -5,
  max_theta = 5,
  min_score = 0,
  max_score = 1,
  z_ci = 1.96,
  file_pdf = NULL,
  ...
)

\S4method{plotCAT}{output_Shadow}(
  object,
  examinee_id = 1,
  min_theta = -5,
  max_theta = 5,
  min_score = 0,
  max_score = 1,
  z_ci = 1.96,
  file_pdf = NULL,
  ...
)
}
\arguments{
\item{object}{An output object generated by \code{\link{Shadow}}.}

\item{examinee_id}{Numeric ID of the examinee to draw the plot.}

\item{min_theta}{A lower bound of theta.}

\item{max_theta}{An upper bound of theta.}

\item{min_score}{A minimum item score.}

\item{max_score}{A maximum item score.}

\item{z_ci}{A quantile of the normal distribution for confidence intervals.}

\item{file_pdf}{If supplied a filename, save as a PDF file.}

\item{...}{Additional options to be passed on to \code{pdf()}.}
}
\description{
(deprecated) Use \code{\link[TestDesign:plot-methods]{plot}} with \code{type = 'audit'} instead.
}
\examples{
\dontrun{
config <- createShadowTestConfig()
true_theta <- rnorm(1)
solution <- Shadow(config, constraints_science, true_theta)
plotCAT(solution, 1)
}
}
