% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotEligibilityStats}
\alias{plotEligibilityStats}
\title{Draw item eligibility statistics plots}
\usage{
plotEligibilityStats(config, object = NULL, object.no.fading = NULL,
  file = NULL, file.no.fading = NULL, segment = 1, items = c(1),
  file.pdf = NULL, max.rate = 0.25, discard.first = NULL)
}
\arguments{
\item{config}{A \code{\linkS4class{Shadow.config}} object.}

\item{object}{An object containing eligibility statistics generated by \code{\link{Shadow}}.}

\item{object.no.fading}{An object containing eligibility statistics generated without fading.}

\item{file}{The filename of an object containing eligibility statistics generated by \code{\link{Shadow}}.}

\item{file.no.fading}{The filename of an object containing eligibility statistics generated without fading.}

\item{segment}{A theta segment index.}

\item{items}{A vector of item indices to generate the plots.}

\item{file.pdf}{If supplied a filename, save as a PDF file.}

\item{max.rate}{A target item exposure rate.}

\item{discard.first}{A integer identifying the first x simulees to discard as burn-in.}
}
\description{
Draw item eligibility statistics plots.
}
