% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotEligibilityStats}
\alias{plotEligibilityStats}
\title{plotEligibilityStats}
\usage{
plotEligibilityStats(config, object = NULL, objectNoFading = NULL,
  file = NULL, fileNoFading = NULL, segment = 1, items = c(1),
  PDF = NULL, maxRate = 0.25, discardFirst = NULL)
}
\arguments{
\item{config}{A \code{\linkS4class{Shadow.config}} object}

\item{object}{An object containing eligibility statistics generated by \code{\linkS4class{Shadow}}}

\item{objectNoFading}{An object containing eligibility statistics generated without fading}

\item{file}{A file name of an object containing eligibility statistics generated by \code{\linkS4class{Shadow}}}

\item{fileNoFading}{A file name of an object containing eligibility statistics generated without fading}

\item{segment}{A theta segment index}

\item{items}{A vector of item indexes to generate the plots}

\item{PDF}{If supplied a filename, save as a PDF file}

\item{maxRate}{A target item exposure rate}

\item{discardFirst}{A integer identifying the first x simulees to discard as burn-in}
}
\description{
Draw item eligibility statistics plots
}
