% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{plotEligibilityStats}
\alias{plotEligibilityStats}
\title{Draw item eligibility statistics plots}
\usage{
plotEligibilityStats(config, object = NULL, object_no_fading = NULL,
  file = NULL, file_no_fading = NULL, segment = 1, items = c(1),
  file_pdf = NULL, max_rate = 0.25, discard_first = NULL)
}
\arguments{
\item{config}{A \code{\linkS4class{config_Shadow}} object.}

\item{object}{An object containing eligibility statistics generated by \code{\link{Shadow}}.}

\item{object_no_fading}{An object containing eligibility statistics generated without fading.}

\item{file}{The filename of an object containing eligibility statistics generated by \code{\link{Shadow}}.}

\item{file_no_fading}{The filename of an object containing eligibility statistics generated without fading.}

\item{segment}{A theta segment index.}

\item{items}{A vector of item indices to generate the plots.}

\item{file_pdf}{If supplied a filename, save as a PDF file.}

\item{max_rate}{A target item exposure rate.}

\item{discard_first}{A integer identifying the first x simulees to discard as burn-in.}
}
\description{
Draw item eligibility statistics plots.
}
