% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{plotExposure}
\alias{plotExposure}
\alias{plotExposure,list-method}
\title{Draw an item exposure plot}
\usage{
plotExposure(object, maxRate = 0.25, thetaSegment = "Estimated",
  PDF = NULL, ...)

\S4method{plotExposure}{list}(object, maxRate = 0.25,
  thetaSegment = "Estimated", PDF = NULL, ...)
}
\arguments{
\item{object}{An output object generated by \code{\link{Shadow}}.}

\item{maxRate}{A target exposure rate.}

\item{thetaSegment}{True or Estimated theta used to create segments ("Estimated" or "True").}

\item{PDF}{If supplied a filename, save as a PDF file.}

\item{...}{Additional options to be passed on to \code{pdf()}.}
}
\description{
Draw a plot of item exposure rates
}
