% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{plotExposureRateBySegment}
\alias{plotExposureRateBySegment}
\title{Draw exposure rate plots by theta segment}
\usage{
plotExposureRateBySegment(object, config, max.rate = 0.25,
  file.pdf = NULL, width = 7, height = 6, mfrow = c(2, 4))
}
\arguments{
\item{object}{An output object generated by \code{\link{Shadow}}.}

\item{config}{A \code{\linkS4class{Shadow.config}} object.}

\item{max.rate}{A target item exposure rate.}

\item{file.pdf}{If supplied a filename, save as a PDF file.}

\item{width}{Width of the graphics device.}

\item{height}{Height of the graphics device.}

\item{mfrow}{Number of multiple figures defined as c(nrow, ncol).}
}
\description{
Draw exposure rate plots by theta segment.
}
