% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotExposureRateBySegment}
\alias{plotExposureRateBySegment}
\title{plotExposureRateBySegment}
\usage{
plotExposureRateBySegment(object, config, maxRate = 0.25, PDF = NULL,
  width = 7, height = 6, mfrow = c(2, 4))
}
\arguments{
\item{object}{An output object generated by \code{\link{Shadow}}}

\item{config}{An object of class \code{\link{Shadow.config}}}

\item{maxRate}{A target item exposure rate}

\item{PDF}{If supplied a filename, save as a PDF file}

\item{width}{Width of the graphics device}

\item{height}{Height of the graphics device}

\item{mfrow}{Number of multiple figures defined as c(nrow, ncol)}
}
\description{
Draw exposure rate plots by theta segment
}
