% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{plotExposureRateFinal}
\alias{plotExposureRateFinal}
\title{Draw exposure rate plots by final theta segment}
\usage{
plotExposureRateFinal(object, config = NULL, max.rate = 0.25,
  theta = "Estimated", segment.cut = NULL, color = "red",
  file.pdf = NULL, width = 7, height = 6, mfrow = c(2, 4),
  burn = 0, retain = NULL)
}
\arguments{
\item{object}{An output object generated by \code{\link{Shadow}}.}

\item{config}{A \code{\linkS4class{Shadow.config}} object.}

\item{max.rate}{A target item exposure rate.}

\item{theta}{By which theta to generate the plots, either "Estimated" or "True".}

\item{segment.cut}{A vector of cut values defining theta segments.}

\item{color}{A vector of colors.}

\item{file.pdf}{If supplied a filename, save as a PDF file.}

\item{width}{Width of the graphics object.}

\item{height}{Height of the graphics object.}

\item{mfrow}{Number of multiple figures defined as c(nrow, ncol).}

\item{burn}{An integer identifying the first x simulees to discard as burn-in.}

\item{retain}{An optional vector of indices identifying the simulees to retain.}
}
\description{
Draw exposure rate plots by final theta segment.
}
