% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotExposureRateFinal}
\alias{plotExposureRateFinal}
\title{plotExposureRateFinal}
\usage{
plotExposureRateFinal(object, config = NULL, maxRate = 0.25,
  theta = "Estimated", segmentCut = NULL, color = "red",
  PDF = NULL, width = 7, height = 6, mfrow = c(2, 4), burnIn = 0,
  retain = NULL)
}
\arguments{
\item{object}{An output object generated by \code{\link{Shadow}}}

\item{config}{An object of class \code{\link{Shadow.config}}}

\item{maxRate}{A target item exposure rate}

\item{theta}{By which theta to generate the plots, either "Estimated" or "True"}

\item{segmentCut}{A vector of cut values defining theta segments}

\item{color}{A vector of colors}

\item{PDF}{If supplied a filename, save as a PDF file}

\item{width}{Width of the graphics object}

\item{height}{Height of the graphics object}

\item{mfrow}{Number of multiple figures defined as c(nrow, ncol)}

\item{burnIn}{An integer identifying the first x simulees to discard as burn-in}

\item{retain}{An optional vector of indexes identifying the simulees to retain}
}
\description{
Draw exposure rate plots by final theta segment
}
