% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{plotExposureRateFinalFlag}
\alias{plotExposureRateFinalFlag}
\title{Draw item information plots for flagged items by segment}
\usage{
plotExposureRateFinalFlag(object, pool, theta = seq(-3, 3, 0.1),
  flag_from = 0.4, file_pdf = NULL, width = 7, height = 6,
  color = "red", mfrow = c(2, 4))
}
\arguments{
\item{object}{A list object generated by \code{\link{plotExposureRateFinal}}.}

\item{pool}{An \code{\linkS4class{item_pool}} object.}

\item{theta}{A theta grid.}

\item{flag_from}{A flagging criterion.}

\item{file_pdf}{If supplied a filename, save as a PDF file.}

\item{width}{Width of the graphics device.}

\item{height}{Height of the graphics device.}

\item{color}{Plotting color.}

\item{mfrow}{Number of multiple figures defined as c(nrow, ncol).}
}
\description{
Draw item information plots for flagged items by segment.
}
