% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_functions.R
\docType{methods}
\name{plotInfo}
\alias{plotInfo}
\alias{plotInfo,output_Static-method}
\alias{plotInfo,item_pool-method}
\alias{plotInfo,constraints-method}
\title{(deprecated) Plot item/test/pool-level information}
\usage{
plotInfo(
  object,
  theta = seq(-3, 3, 0.1),
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "blue",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(2, 4)
)

\S4method{plotInfo}{output_Static}(
  object,
  theta = seq(-3, 3, 0.1),
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "blue",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(2, 4)
)

\S4method{plotInfo}{item_pool}(
  object,
  theta = seq(-3, 3, 0.1),
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "blue",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(1, 1)
)

\S4method{plotInfo}{constraints}(
  object,
  theta = seq(-3, 3, 0.1),
  info_type = "FISHER",
  plot_sum = TRUE,
  select = NULL,
  color = "black",
  file_pdf = NULL,
  width = 7,
  height = 6,
  mfrow = c(1, 1)
)
}
\arguments{
\item{object}{\itemize{
  \item{\code{\linkS4class{item_pool}}: plot pool-level or item-level information.}
  \item{\code{\linkS4class{output_Static}}: plot test-level information of the assembly solution.}
  \item{\code{\linkS4class{constraints}}: plot attainable information range.}
}}

\item{theta}{the theta grid to use on the x-axis. (default = \code{seq(-3, 3, .1)})}

\item{info_type}{the type of information. Accepts \code{FISHER}. (default = \code{FISHER})}

\item{plot_sum}{used when the \code{object} argument is an \code{\linkS4class{item_pool}} object.
If \code{TRUE} then draw pool-level information, and if \code{FALSE} draw item-level information for every item in the pool. (default = \code{TRUE})}

\item{select}{(optional) used when the \code{object} argument is an \code{\linkS4class{item_pool}} object. Items to select from the pool.}

\item{color}{the color of the curve. (default = \code{blue})}

\item{file_pdf}{(optional) if supplied a filename, save as a PDF file.}

\item{width}{the width of the plot. (default = \code{7})}

\item{height}{the height of the plot. (default = \code{6})}

\item{mfrow}{multi-panel configurations to use. (default = \code{c(2, 4)})}
}
\description{
(deprecated) Use \code{\link[TestDesign:plot-methods]{plot}}.
}
\examples{
subitempool <- subsetItemPool(itempool_science, 1:8)
plot(subitempool)

config <- createStaticTestConfig()
solution <- Static(config, constraints_science)
plot(solution)

}
