% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotInfo}
\alias{plotInfo}
\title{plotInfo}
\usage{
plotInfo(object, theta, infoType = "FISHER", select = NULL,
  pdfFile = NULL, color = "blue", width = 7, height = 6,
  mfrow = c(2, 4))
}
\arguments{
\item{object}{A nice parameter}

\item{theta}{A nice parameter}

\item{infoType}{A nice parameter}

\item{select}{A nice parameter}

\item{pdfFile}{A nice parameter}

\item{color}{A nice parameter}

\item{width}{A nice parameter}

\item{height}{A nice parameter}

\item{mfrow}{A nice parameter}
}
\description{
plotInfo
}
