% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotInfoOverlay}
\alias{plotInfoOverlay}
\title{plotInfoOverlay}
\usage{
plotInfoOverlay(object, theta, infoType = "FISHER", select = NULL,
  PDF = NULL, color = "red", width = 7, height = 6)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{item.pool}}}

\item{theta}{A theta grid}

\item{infoType}{Type of information}

\item{select}{A vector of item indexes to subset}

\item{PDF}{If supplied a filename, save as a PDF file}

\item{color}{Plotting color}

\item{width}{Width of the graphics device}

\item{height}{Height of the graphics device}
}
\description{
Overlay item information plots
}
