% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{plotInfoOverlay}
\alias{plotInfoOverlay}
\title{Overlay item information plots}
\usage{
plotInfoOverlay(object, theta, info_type = "FISHER", select = NULL,
  file_pdf = NULL, color = "red", width = 7, height = 6)
}
\arguments{
\item{object}{An \code{\linkS4class{item_pool}} object.}

\item{theta}{A theta grid.}

\item{info_type}{Type of information.}

\item{select}{A vector of indices identifying the items to subset.}

\item{file_pdf}{If supplied a filename, save as a PDF file.}

\item{color}{Plotting color.}

\item{width}{Width of the graphics device.}

\item{height}{Height of the graphics device.}
}
\description{
Overlay item information plots.
}
