% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{plotRMSE}
\alias{plotRMSE}
\title{Draw RMSE plots}
\usage{
plotRMSE(..., title = NULL, legendTitle = NULL, legendLabels = NULL,
  ltySet = NULL, colSet = NULL, theta = seq(-2, 2, 1))
}
\arguments{
\item{...}{A series of RMSE values}

\item{title}{A plot title}

\item{legendTitle}{A legend title}

\item{legendLabels}{A vector of labels for the series}

\item{ltySet}{A vector of line types for the series}

\item{colSet}{A vector of colors for the series}

\item{theta}{A theta grid}
}
\description{
Draw RMSE plots.
}
