% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{plotRMSE}
\alias{plotRMSE}
\title{Draw RMSE plots}
\usage{
plotRMSE(
  ...,
  title = NULL,
  legend_title = NULL,
  legend_labels = NULL,
  lty_set = NULL,
  col_set = NULL,
  theta = seq(-2, 2, 1)
)
}
\arguments{
\item{...}{A series of RMSE values.}

\item{title}{A plot title.}

\item{legend_title}{A legend title.}

\item{legend_labels}{A vector of labels for the series.}

\item{lty_set}{A vector of line types for the series.}

\item{col_set}{A vector of colors for the series.}

\item{theta}{A theta grid.}
}
\description{
Draw RMSE plots.
}
