% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{plotShadow}
\alias{plotShadow}
\alias{plotShadow,list-method}
\alias{plotShadow,Shadow.output-method}
\title{Draw a shadow test chart}
\usage{
plotShadow(object, constraints, examinee.id = 1,
  sort.by.difficulty = FALSE, file.pdf = NULL, ...)

\S4method{plotShadow}{list}(object, constraints, examinee.id = 1,
  sort.by.difficulty = FALSE, file.pdf = NULL, ...)

\S4method{plotShadow}{Shadow.output}(object, constraints,
  examinee.id = 1, sort.by.difficulty = FALSE, file.pdf = NULL, ...)
}
\arguments{
\item{object}{An output from \code{\link{Shadow}} function.}

\item{constraints}{The constraint object used in obtaining the output.}

\item{examinee.id}{Numeric ID of the examinee to draw the plot.}

\item{sort.by.difficulty}{Sort the items by difficulty.}

\item{file.pdf}{If supplied a filename, save as a PDF file.}

\item{...}{Additional options to be passed on to \code{pdf()}.}
}
\description{
Draw a chart of shadow tests constructed for each simulee. The index of a column represents the position of item administration process, and each column represents the item pool.
}
