% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{plotShadow}
\alias{plotShadow}
\alias{plotShadow,Shadow.output-method}
\alias{plotShadow,list-method}
\title{An S4 generic}
\usage{
plotShadow(object, Constraints, sortByDifficulty = FALSE, PDF = NULL)

\S4method{plotShadow}{Shadow.output}(object, Constraints,
  sortByDifficulty = FALSE, PDF = NULL)

\S4method{plotShadow}{list}(object, Constraints,
  sortByDifficulty = FALSE, PDF = NULL)
}
\arguments{
\item{object}{An output from \code{\link{Shadow}} function.}

\item{Constraints}{The constraint object used in obtaining the output.}

\item{sortByDifficulty}{Sort the items by difficulty.}

\item{PDF}{If supplied a filename, save as a PDF file.}
}
\description{
Draw a chart of shadow tests constructed for each simulee. The index of a column represents the position of item administration process, and each column represents the item pool.
}
