% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\docType{methods}
\name{plotShadow}
\alias{plotShadow}
\alias{plotShadow,output_Shadow_all-method}
\alias{plotShadow,list-method}
\title{(deprecated) Plot a shadow test chart}
\usage{
plotShadow(
  object,
  examinee_id = 1,
  sort_by_difficulty = FALSE,
  file_pdf = NULL,
  simple = FALSE,
  ...
)

\S4method{plotShadow}{output_Shadow_all}(
  object,
  examinee_id = 1,
  sort_by_difficulty = FALSE,
  file_pdf = NULL,
  simple = FALSE,
  ...
)

\S4method{plotShadow}{list}(
  object,
  examinee_id = 1,
  sort_by_difficulty = FALSE,
  file_pdf = NULL,
  simple = FALSE,
  ...
)
}
\arguments{
\item{object}{An output from \code{\link{Shadow}} function.}

\item{examinee_id}{Numeric ID of the examinee to draw the plot.}

\item{sort_by_difficulty}{Sort the items by difficulty. (not implemented)}

\item{file_pdf}{If supplied a filename, save as a PDF file.}

\item{simple}{If \code{TRUE}, simplity the chart by hiding unused items.}

\item{...}{Additional options to be passed on to \code{pdf()}.}
}
\description{
(deprecated) Use \code{\link[TestDesign:plot-methods]{plot}} with \code{type = 'shadow'} instead.
}
\examples{
\dontrun{
config <- createShadowTestConfig()
true_theta <- rnorm(1)
solution <- Shadow(config, constraints_science, true_theta)
plotShadow(solution, 1)
plotShadow(solution, 1, simple = TRUE)
}
}
