% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R, R/shadow_class.R
\docType{methods}
\name{print,constraints-method}
\alias{print,constraints-method}
\alias{print,exposure_rate_plot-method}
\title{Print constraints object}
\usage{
\S4method{print}{constraints}(x, ...)

\S4method{print}{exposure_rate_plot}(x, ...)
}
\arguments{
\item{x}{A \code{exposure_rate_plot} object.}

\item{...}{Further arguments to pass onto \code{\link{print}} function.}
}
\description{
Print \code{\linkS4class{constraints}} object. This is a shortcut to the \code{@constraints} slot.

Print \code{exposure_rate_plot} object. This is a shortcut to the \code{@plot} slot.
}
