% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_class.R
\docType{methods}
\name{print,exposure_rate_plot-method}
\alias{print,exposure_rate_plot-method}
\title{Print exposure rate plot}
\usage{
\S4method{print}{exposure_rate_plot}(x, ...)
}
\arguments{
\item{x}{A \code{exposure_rate_plot} object.}

\item{...}{Further arguments to pass onto \code{\link{print}} function.}
}
\description{
Print \code{exposure_rate_plot} object. This is a shortcut to the \code{@plot} slot.
}
