% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/print_functions.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,item_1PL-method}
\alias{print,item_2PL-method}
\alias{print,item_3PL-method}
\alias{print,item_PC-method}
\alias{print,item_GPC-method}
\alias{print,item_GR-method}
\title{Extension of print() for objects in TestDesign package}
\usage{
\S4method{print}{item_1PL}(x)

\S4method{print}{item_2PL}(x)

\S4method{print}{item_3PL}(x)

\S4method{print}{item_PC}(x)

\S4method{print}{item_GPC}(x)

\S4method{print}{item_GR}(x)
}
\arguments{
\item{x}{an object to print.}
}
\description{
Extension of print() for objects in TestDesign package
}
