% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/print_functions.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,item_1PL-method}
\alias{print,item_2PL-method}
\alias{print,item_3PL-method}
\alias{print,item_PC-method}
\alias{print,item_GPC-method}
\alias{print,item_GR-method}
\alias{print,item_pool-method}
\alias{print,item_attrib-method}
\alias{print,st_attrib-method}
\alias{print,summary_item_attrib-method}
\alias{print,constraints-method}
\alias{print,config_Static-method}
\alias{print,config_Shadow-method}
\alias{print,output_Static-method}
\alias{print,output_Shadow-method}
\alias{print,output_Shadow_all-method}
\alias{print,exposure_rate_plot-method}
\alias{print,summary_item_pool-method}
\alias{print,summary_constraints-method}
\alias{print,summary_output_Static-method}
\alias{print,summary_output_Shadow_all-method}
\title{Extension of print() for objects in TestDesign package}
\usage{
\S4method{print}{item_1PL}(x)

\S4method{print}{item_2PL}(x)

\S4method{print}{item_3PL}(x)

\S4method{print}{item_PC}(x)

\S4method{print}{item_GPC}(x)

\S4method{print}{item_GR}(x)

\S4method{print}{item_pool}(x)

\S4method{print}{item_attrib}(x)

\S4method{print}{st_attrib}(x)

\S4method{print}{summary_item_attrib}(x)

\S4method{print}{constraints}(x)

\S4method{print}{config_Static}(x)

\S4method{print}{config_Shadow}(x)

\S4method{print}{output_Static}(x, index_only = TRUE)

\S4method{print}{output_Shadow}(x)

\S4method{print}{output_Shadow_all}(x)

\S4method{print}{exposure_rate_plot}(x)

\S4method{print}{summary_item_pool}(x)

\S4method{print}{summary_constraints}(x)

\S4method{print}{summary_output_Static}(x, digits = 3)

\S4method{print}{summary_output_Shadow_all}(x, digits = 3)
}
\arguments{
\item{x}{an object to print.}

\item{index_only}{if \code{TRUE} then only print item indices. If \code{FALSE} then print all item attributes. (default = \code{TRUE})}

\item{digits}{minimal number of *significant* digits. See \code{\link{print.default}}.}
}
\description{
Extension of print() for objects in TestDesign package
}
