% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{RE}
\alias{RE}
\title{Calculate Relative Errors}
\usage{
RE(RMSE.foc, RMSE.ref)
}
\arguments{
\item{RMSE.foc}{A vector of RMSE values for the focal group.}

\item{RMSE.ref}{A vector of RMSE values for the reference group.}
}
\description{
Calculate Relative Errors.
}
