% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{RE}
\alias{RE}
\title{Calculate Relative Errors}
\usage{
RE(RMSE_foc, RMSE_ref)
}
\arguments{
\item{RMSE_foc}{A vector of RMSE values for the focal group.}

\item{RMSE_ref}{A vector of RMSE values for the reference group.}
}
\description{
Calculate Relative Errors.
}
