% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{RMSE}
\alias{RMSE}
\title{Calculate Root Mean Squared Error}
\usage{
RMSE(x, y, conditional = TRUE)
}
\arguments{
\item{x}{A vector of values.}

\item{y}{A vector of values.}

\item{conditional}{If \code{TRUE}, calculate RMSE conditional on x.}
}
\description{
Calculate Root Mean Squared Error.
}
