% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{SaveOutput}
\alias{SaveOutput}
\title{Save or print audit trails}
\usage{
SaveOutput(objectList, file = NULL)
}
\arguments{
\item{objectList}{A list of output objects generated from \code{STA}.}

\item{file}{An optional file name as a character string to save the output.}
}
\value{
None
}
\description{
Save or print audit trails for all simulees.
}
