% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{saveOutput}
\alias{saveOutput}
\title{Save or print audit trails}
\usage{
saveOutput(object_list, file = NULL)
}
\arguments{
\item{object_list}{A list of output objects generated from \code{STA}.}

\item{file}{An optional file name as a character string to save the output.}
}
\value{
None
}
\description{
Save or print audit trails for all simulees.
}
