% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{Shadow}
\alias{Shadow}
\alias{Shadow,item.pool-method}
\title{Shadow}
\usage{
Shadow(object, config, trueTheta = NULL, Constraints = NULL,
  prior = NULL, priorPar = NULL, Data = NULL, session = NULL)

\S4method{Shadow}{item.pool}(object, config, trueTheta = NULL,
  Constraints = NULL, prior = NULL, priorPar = NULL, Data = NULL,
  session = NULL)
}
\arguments{
\item{object}{A nice parameter}

\item{config}{A nice parameter}

\item{trueTheta}{A nice parameter}

\item{Constraints}{A nice parameter}

\item{prior}{A nice parameter}

\item{priorPar}{A nice parameter}

\item{Data}{A nice parameter}

\item{session}{A nice parameter}
}
\description{
Shadow
}
