% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\docType{methods}
\name{Shadow}
\alias{Shadow}
\alias{Shadow,item.pool-method}
\title{Run computerized adaptive testing with generalized shadow-test approach}
\usage{
Shadow(object, config, true.theta = NULL, constraints = NULL,
  prior = NULL, prior.par = NULL, data = NULL, session = NULL)

\S4method{Shadow}{item.pool}(object, config, true.theta = NULL,
  constraints = NULL, prior = NULL, prior.par = NULL, data = NULL,
  session = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{item.pool}} object. Use \code{\link{loadItemPool}} for this.}

\item{config}{A \code{\linkS4class{Shadow.config}} object.}

\item{true.theta}{Numeric. A vector of true theta values to be used in simulation.}

\item{constraints}{A list representing optimization constraints. Use \code{\link{loadConstraints}} for this.}

\item{prior}{Numeric. A matrix or a vector containing priors.}

\item{prior.par}{Numeric. A vector of parameters for prior distribution.}

\item{data}{Numeric. A matrix containing item response data.}

\item{session}{Used to communicate with a Shiny session.}
}
\description{
Run computerized adaptive testing with generalized shadow-test approach.
}
\examples{
object <- itempool.science
config <- createShadowTestConfig()
trueTheta <- rnorm(1)
solution <- Shadow(itempool.science, config, trueTheta, constraints.science)
solution$output
}
