% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{Shadow.config-class}
\alias{Shadow.config-class}
\title{Shadow.config}
\description{
An S4 object to store configurations for Shadow Test Assembly (STA).
}
\section{Slots}{

\describe{
\item{\code{itemSelection}}{A list containing item selection criteria.
\itemize{
  \item{\code{method}} The type of criteria. Accepts one of \code{MFI, MPWI, FB, EB}.
  \item{\code{infoType}} The type of information. Accepts \code{FISHER}.
  \item{\code{initialTheta}} 
  \item{\code{fixedTheta}} 
}}

\item{\code{contentBalancing}}{A list containing content balancing options.
\itemize{
  \item{\code{method}} The type of balancing method. Accepts \code{NONE}, or \code{STA}.
}}

\item{\code{MIP}}{A list containing solver options.
\itemize{
  \item{\code{solver}} The type of solver. Accepts one of \code{SYMPHONY, GUROBI, GLPK, LPSOLVE}.
  \item{\code{verbosity}} Verbosity level.
  \item{\code{timeLimit}} Time limit to be passed onto solver. Used in solvers \code{SYMPHONY, GUROBI, GLPK}.
  \item{\code{gapLimit}} Gap limit to be passed onto solver. Used in solvers  \code{SYMPHONY, GUROBI}.
}}

\item{\code{MCMC}}{A list containing Markov-chain Monte Carlo configurations.
\itemize{
  \item{\code{burnIn}} Numeric. The number of chains from the start to discard.
  \item{\code{postBurnIn}}  Numeric. The number of chains to use after discarding the first \code{burnIn} chains.
  \item{\code{thin}} Thin.
  \item{\code{jumpfactor}} JumpFactor.
}}

\item{\code{refreshPolicy}}{A list containing refresh policy for obtaining a new shadow test.
\itemize{
  \item{\code{method}} The type of policy. Accepts one of \code{ALWAYS, POSITION, INTERVAL, THRESHOLD, INTERVAL-THRESHOLD, STIMULUS, SET, PASSAGE}.
  \item{\code{interval}} Integer. Set to 1 to refresh at each position, 2 to refresh at every two positions, and so on.
  \item{\code{threshold}} Numeric. The shadow test is refreshed when the absolute change in theta estimate is greater than this value.
  \item{\code{position}}
}}

\item{\code{exposureControl}}{A list.}

\item{\code{stoppingCriterion}}{A list.}

\item{\code{interimTheta}}{A list containing interim theta estimation options.
\itemize{
  \item{\code{method}} The type of estimation. Accepts one of \code{EAP, EB, FB}.
  \item{\code{priorDist}} The type of prior distribution. Accepts one of \code{NORMAL, UNIF}.
  \item{\code{priorPar}} Distributional parameters for the prior.
  \item{\code{boundML}}
  \item{\code{maxIter}}
  \item{\code{crit}}
}}

\item{\code{finalTheta}}{A list containing final theta estimation options.
\itemize{
  \item{\code{method}} The type of estimation. Accepts one of \code{EAP, EB, FB}.
  \item{\code{priorDist}} The type of prior distribution. Accepts one of \code{NORMAL, UNIF}.
  \item{\code{priorPar}} Distributional parameters for the prior.
  \item{\code{boundML}}
  \item{\code{maxIter}}
  \item{\code{crit}}
}}

\item{\code{thetaGrid}}{Numeric. A vector of theta values to represent the continuum.}

\item{\code{auditTrail}}{Logical.}
}}

