% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATA_class.R, R/IRT_method.R,
%   R/Shadow_functions.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,ATA.config-method}
\alias{show,item.1pl-method}
\alias{show,item.2pl-method}
\alias{show,item.3pl-method}
\alias{show,item.pc-method}
\alias{show,item.gpc-method}
\alias{show,item.gr-method}
\alias{show,item.pool-method}
\alias{show,pool.cluster-method}
\alias{show,Shadow.config-method}
\alias{show,Shadow.output-method}
\title{show}
\usage{
\S4method{show}{ATA.config}(object)

\S4method{show}{item.1pl}(object)

\S4method{show}{item.2pl}(object)

\S4method{show}{item.3pl}(object)

\S4method{show}{item.pc}(object)

\S4method{show}{item.gpc}(object)

\S4method{show}{item.gr}(object)

\S4method{show}{item.pool}(object)

\S4method{show}{pool.cluster}(object)

\S4method{show}{Shadow.config}(object)

\S4method{show}{Shadow.output}(object)
}
\arguments{
\item{object}{A nice parameter}
}
\value{
\code{show} returns an invisible \code{NULL}.
}
\description{
show

show

show

show

show

show

show

show

show

show
}
\details{
Objects from an S4 class (a class defined by a call to
  \code{\link{setClass}}) will be displayed automatically is if by a
  call to \code{show}.  S4 objects that occur as attributes of S3
objects will also be displayed in this form; conversely, S3 objects
encountered as slots in S4 objects will be printed using the S3
convention, as if by a call to \code{\link{print}}.

Methods defined for \code{show} will only be inherited  by simple
inheritance, since otherwise the method would not receive the
complete, original object, with misleading results.  See the
\code{simpleInheritanceOnly} argument to \code{\link{setGeneric}} and
the discussion in \code{\link{setIs}} for the general concept.
}
\examples{
## following the example shown in the setMethod documentation ...
setClass("track", slots = c(x="numeric", y="numeric"))
setClass("trackCurve", contains = "track", slots = c(smooth = "numeric"))

t1 <- new("track", x=1:20, y=(1:20)^2)

tc1 <- new("trackCurve", t1)

setMethod("show", "track",
  function(object)print(rbind(x = object@x, y=object@y))
)
## The method will now be used for automatic printing of t1

t1

\dontrun{  [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
x    1    2    3    4    5    6    7    8    9    10    11    12
y    1    4    9   16   25   36   49   64   81   100   121   144
  [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20]
x    13    14    15    16    17    18    19    20
y   169   196   225   256   289   324   361   400
}
## and also for tc1, an object of a class that extends "track"
tc1

\dontrun{  [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
x    1    2    3    4    5    6    7    8    9    10    11    12
y    1    4    9   16   25   36   49   64   81   100   121   144
  [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20]
x    13    14    15    16    17    18    19    20
y   169   196   225   256   289   324   361   400
}
}
\seealso{
\code{\link{showMethods}} prints all the methods for one or more
  functions.
}
