% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/show_functions.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,item_1PL-method}
\alias{show,item_2PL-method}
\alias{show,item_3PL-method}
\alias{show,item_PC-method}
\alias{show,item_GPC-method}
\alias{show,item_GR-method}
\alias{show,item_pool-method}
\alias{show,pool_cluster-method}
\alias{show,item_attrib-method}
\title{Extension of show() for objects in TestDesign package}
\usage{
\S4method{show}{item_1PL}(object)

\S4method{show}{item_2PL}(object)

\S4method{show}{item_3PL}(object)

\S4method{show}{item_PC}(object)

\S4method{show}{item_GPC}(object)

\S4method{show}{item_GR}(object)

\S4method{show}{item_pool}(object)

\S4method{show}{pool_cluster}(object)

\S4method{show}{item_attrib}(object)
}
\arguments{
\item{object}{an object to display.}
}
\description{
Extension of show() for objects in TestDesign package
}
