% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/show_functions.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,item_1PL-method}
\alias{show,item_2PL-method}
\alias{show,item_3PL-method}
\alias{show,item_PC-method}
\alias{show,item_GPC-method}
\alias{show,item_GR-method}
\alias{show,item_pool-method}
\alias{show,item_pool_cluster-method}
\alias{show,pool_cluster-method}
\alias{show,item_attrib-method}
\alias{show,st_attrib-method}
\alias{show,constraints-method}
\alias{show,summary_item_pool-method}
\alias{show,summary_item_attrib-method}
\alias{show,summary_constraints-method}
\alias{show,config_Static-method}
\alias{show,config_Shadow-method}
\alias{show,output_Static-method}
\alias{show,output_Shadow-method}
\alias{show,output_Shadow_all-method}
\alias{show,summary_output_Static-method}
\alias{show,summary_output_Shadow_all-method}
\alias{show,exposure_rate_plot-method}
\title{Extension of show() for objects in TestDesign package}
\usage{
\S4method{show}{item_1PL}(object)

\S4method{show}{item_2PL}(object)

\S4method{show}{item_3PL}(object)

\S4method{show}{item_PC}(object)

\S4method{show}{item_GPC}(object)

\S4method{show}{item_GR}(object)

\S4method{show}{item_pool}(object)

\S4method{show}{item_pool_cluster}(object)

\S4method{show}{item_attrib}(object)

\S4method{show}{st_attrib}(object)

\S4method{show}{constraints}(object)

\S4method{show}{summary_item_pool}(object)

\S4method{show}{summary_item_attrib}(object)

\S4method{show}{summary_constraints}(object)

\S4method{show}{config_Static}(object)

\S4method{show}{config_Shadow}(object)

\S4method{show}{output_Static}(object)

\S4method{show}{output_Shadow}(object)

\S4method{show}{output_Shadow_all}(object)

\S4method{show}{summary_output_Static}(object)

\S4method{show}{summary_output_Shadow_all}(object)

\S4method{show}{exposure_rate_plot}(object)
}
\arguments{
\item{object}{an object to display.}
}
\description{
Extension of show() for objects in TestDesign package
}
