% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{showConstraints}
\alias{showConstraints}
\title{Show constraints}
\usage{
showConstraints(constraints)
}
\arguments{
\item{constraints}{Output from \code{\link{loadConstraints}}.}
}
\description{
Show constraints. This function is a shortcut to access \code{''} slot.
}
