% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{methods}
\name{showConstraints}
\alias{showConstraints}
\title{(deprecated) Show constraints}
\usage{
showConstraints(constraints)
}
\arguments{
\item{constraints}{a \code{\linkS4class{constraints}} object.}
}
\description{
(deprecated) Use \code{\link[TestDesign:print-methods]{print}} instead.
}
