% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT_method.R, R/Shadow_functions.R
\docType{methods}
\name{simResp}
\alias{simResp}
\alias{simResp,item.1pl,numeric-method}
\alias{simResp,item.2pl,numeric-method}
\alias{simResp,item.3pl,numeric-method}
\alias{simResp,item.pc,numeric-method}
\alias{simResp,item.gpc,numeric-method}
\alias{simResp,item.gr,numeric-method}
\alias{simResp,item.pool,numeric-method}
\alias{simResp,pool.cluster,numeric-method}
\alias{simResp,pool.cluster,list-method}
\title{An S4 generic and its methods to simulate responses}
\usage{
simResp(object, theta)

\S4method{simResp}{item.1pl,numeric}(object, theta)

\S4method{simResp}{item.2pl,numeric}(object, theta)

\S4method{simResp}{item.3pl,numeric}(object, theta)

\S4method{simResp}{item.pc,numeric}(object, theta)

\S4method{simResp}{item.gpc,numeric}(object, theta)

\S4method{simResp}{item.gr,numeric}(object, theta)

\S4method{simResp}{item.pool,numeric}(object, theta)

\S4method{simResp}{pool.cluster,numeric}(object, theta)

\S4method{simResp}{pool.cluster,list}(object, theta)

\S4method{simResp}{pool.cluster,list}(object, theta)
}
\arguments{
\item{object}{An instance of an item class.}

\item{theta}{A vector of theta values.}
}
\value{
Simulated responses.
}
\description{
An S4 generic and its methods to simulate responses
}
\examples{
item.1 = new("item.1pl", difficulty = 0.5)
sim.item.1 = simResp(item.1, seq(-3, 3, 1))
item.2 = new("item.2pl", slope = 1.0, difficulty = 0.5)
sim.item.2 = simResp(item.2, seq(-3, 3, 1))
item.3 = new("item.3pl", slope = 1.0, difficulty = 0.5, guessing = 0.2)
sim.item.3 = simResp(item.3, seq(-3, 3, 1))
item.4 = new("item.pc", threshold = c(-1, 0, 1), ncat = 4)
sim.item.4 = simResp(item.4, seq(-3, 3, 1))
item.5 = new("item.gpc", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
sim.item.5 = simResp(item.5, seq(-3, 3, 1))
item.6 = new("item.gr", slope = 0.9, category = c(-1, 0 , 1), ncat = 4)
sim.item.6 = simResp(item.6, seq(-3, 3, 1))
\dontrun{
itemPool.1 = LoadItemPool("C:/item_par.csv")
sim.itemPool.1 = simResp(itemPool.1, seq(-3, 3, 1))
}
}
\references{
{
  \insertRef{rasch_probabilistic_1960}{Shadow}
}

{
  \insertRef{lord_theory_1952}{Shadow}
  
  \insertRef{birnbaum_efficient_1957}{Shadow}
  
  \insertRef{birnbaum_estimation_1958}{Shadow}
  
  \insertRef{birnbaum_further_1958}{Shadow}
}

{
  \insertRef{birnbaum_latent_1968}{Shadow}
}

{
  \insertRef{masters_rasch_1982}{Shadow}
  
  \insertRef{andrich_rating_1978}{Shadow}
}

{
  \insertRef{muraki_generalized_1992}{Shadow}
}

{
  \insertRef{samejima_estimation_1969}{Shadow}
}
}
