% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_functions.R, R/shadow_functions.R
\docType{methods}
\name{simResp}
\alias{simResp}
\alias{simResp,item_1PL,numeric-method}
\alias{simResp,item_2PL,numeric-method}
\alias{simResp,item_3PL,numeric-method}
\alias{simResp,item_PC,numeric-method}
\alias{simResp,item_GPC,numeric-method}
\alias{simResp,item_GR,numeric-method}
\alias{simResp,item_pool,numeric-method}
\alias{simResp,item_pool_cluster,numeric-method}
\alias{simResp,item_pool_cluster,list-method}
\title{Simulate item response data}
\usage{
simResp(object, theta)

\S4method{simResp}{item_1PL,numeric}(object, theta)

\S4method{simResp}{item_2PL,numeric}(object, theta)

\S4method{simResp}{item_3PL,numeric}(object, theta)

\S4method{simResp}{item_PC,numeric}(object, theta)

\S4method{simResp}{item_GPC,numeric}(object, theta)

\S4method{simResp}{item_GR,numeric}(object, theta)

\S4method{simResp}{item_pool,numeric}(object, theta)

\S4method{simResp}{item_pool_cluster,numeric}(object, theta)

\S4method{simResp}{item_pool_cluster,list}(object, theta)

\S4method{simResp}{item_pool_cluster,list}(object, theta)
}
\arguments{
\item{object}{an \code{\link{item}} or an \code{\linkS4class{item_pool}} object.}

\item{theta}{theta values to use.}
}
\value{
\describe{
  \item{\code{\link{item}} object:}{\code{\link{simResp}} returns a length \emph{nq} vector containing simulated item response data.}
  \item{\code{\linkS4class{item_pool}} object:}{\code{\link{simResp}} returns a (\emph{nq}, \emph{ni}) matrix containing simulated item response data.}
}
}
\description{
\code{\link{simResp}} is a function to simulate item response data.
}
\details{
\describe{
  \item{\emph{notations}}{\itemize{
    \item{\emph{nq} denotes the number of theta values.}
    \item{\emph{ni} denotes the number of items in the \code{\linkS4class{item_pool}} object.}
  }}
}
}
\examples{

item_1    <- new("item_1PL", difficulty = 0.5)
item_2    <- new("item_2PL", slope = 1.0, difficulty = 0.5)
item_3    <- new("item_3PL", slope = 1.0, difficulty = 0.5, guessing = 0.2)
item_4    <- new("item_PC", threshold = c(-1, 0, 1), ncat = 4)
item_5    <- new("item_GPC", slope = 1.2, threshold = c(-0.8, -1.0, 0.5), ncat = 4)
item_6    <- new("item_GR", slope = 0.9, category = c(-1, 0, 1), ncat = 4)

sim_item_1 <- simResp(item_1, seq(-3, 3, 1))
sim_item_2 <- simResp(item_2, seq(-3, 3, 1))
sim_item_3 <- simResp(item_3, seq(-3, 3, 1))
sim_item_4 <- simResp(item_4, seq(-3, 3, 1))
sim_item_5 <- simResp(item_5, seq(-3, 3, 1))
sim_item_6 <- simResp(item_6, seq(-3, 3, 1))
sim_pool   <- simResp(itempool_science, seq(-3, 3, 1))

}
\references{
Rasch, G. (1960). \emph{Probabilistic models for some intelligence and attainment tests.}
Copenhagen: Danish Institute for Educational Research.

Lord, F. M. (1952). A theory of test scores (Psychometric Monograph No. 7). Richmond, VA: Psychometric Corporation.

Birnbaum, A. (1957). \emph{Efficient design and use of tests of mental ability for various decision-making problems}
(Series Report No. 58-16. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). \emph{On the estimation of mental ability}
(Series Report No. 15. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1958). \emph{Further considerations of efficiency in tests of a mental ability}
(Series Report No. 17. Project No. 7755-23).
Randolph Air Force Base, TX: USAF School of Aviation Medicine.

Birnbaum, A. (1968). Some latent trait models and their use in inferring an examinee's ability.
In Lord, F. M., Novick, M. R. (eds.), \emph{Statistical Theories of Mental Test Scores}, 397–479.
Reading, MA: Addison-Wesley.

Masters, G. N. (1982). A Rasch model for partial credit scoring.
\emph{Psychometrika, 47}(2), 149–174.

Andrich, D. (1978). A rating formulation for ordered response categories.
\emph{Psychometrika, 43}(4), 561–573.

Muraki, E. (1992). A generalized partial credit model: Application of an EM algorithm.
\emph{Applied Psychological Measurement, 16}(2), 159–176.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores.
\emph{Psychometrika Monograph, 17}.
}
