% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{STA}
\alias{STA}
\title{STA}
\usage{
STA(Constraints, objective, solver = "Symphony", xmat = NULL,
  xdir = NULL, xrhs = NULL, maximize = TRUE, mps = FALSE,
  lp = FALSE, verbosity = -2, time_limit = 5, gap_limit = -1, ...)
}
\arguments{
\item{Constraints}{A nice parameter}

\item{objective}{A nice parameter}

\item{solver}{A nice parameter}

\item{xmat}{A nice parameter}

\item{xdir}{A nice parameter}

\item{xrhs}{A nice parameter}

\item{maximize}{A nice parameter}

\item{mps}{A nice parameter}

\item{lp}{A nice parameter}

\item{verbosity}{A nice parameter}

\item{time_limit}{A nice parameter}

\item{gap_limit}{A nice parameter}

\item{...}{Some dots}
}
\value{
An output
}
\description{
Do fancy stuff.
}
