% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_functions.R
\name{subsetItemPool}
\alias{subsetItemPool}
\title{subsetItemPool}
\usage{
subsetItemPool(pool, select = NULL)
}
\arguments{
\item{pool}{A nice parameter}

\item{select}{A nice parameter}
}
\description{
subsetItemPool
}
