% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_class.R
\name{subsetItemPool}
\alias{subsetItemPool}
\title{Create a subset of an item pool object}
\usage{
subsetItemPool(pool, select = NULL)
}
\arguments{
\item{pool}{An \code{\linkS4class{item_pool}} object.}

\item{select}{A vector of indices identifying the items to subset.}
}
\description{
Create a subset of an \code{\linkS4class{item_pool}} object.
}
\examples{
subitempool <- subsetItemPool(itempool_science, 1:100)
}
