% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_functions.R
\name{subsetTest}
\alias{subsetTest}
\title{Create a subset of a test object}
\usage{
subsetTest(test, select = NULL)
}
\arguments{
\item{test}{An \code{\linkS4class{test}} object.}

\item{select}{A vector of item indices to subset.}
}
\description{
Create a subset of a test object.
}
\examples{
test <- makeTest(itempool_science, seq(-3, 3, 1))
subtest <- subsetTest(test, 1:100)
}
