% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R, R/static_class.R,
%   R/shadow_class.R
\docType{methods}
\name{summary-method}
\alias{summary-method}
\alias{summary,item_pool-method}
\alias{summary,item_attrib-method}
\alias{summary,output_Static-method}
\title{summary}
\usage{
\S4method{summary}{item_pool}(object)

\S4method{summary}{item_attrib}(object)

\S4method{summary}{output_Static}(object)

\S4method{summary}{output_Shadow_all}(object, ...)
}
\arguments{
\item{object}{An \linkS4class{item_pool} object.}
}
\description{
summary

summary

summary

summary
}
