% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R, R/summary_functions.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,item_pool-method}
\alias{summary,item_attrib-method}
\alias{summary,output_Static-method}
\alias{summary,output_Shadow_all-method}
\title{Extension of summary() for objects in TestDesign package}
\usage{
\S4method{summary}{item_pool}(object)

\S4method{summary}{item_attrib}(object)

\S4method{summary}{output_Static}(object)

\S4method{summary}{output_Shadow_all}(object)
}
\arguments{
\item{object}{an object to summarize.}
}
\description{
Extension of summary() for objects in TestDesign package
}
