% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_class.R
\docType{class}
\name{test-class}
\alias{test-class}
\title{An S4 class to represent a test}
\description{
An S4 class to represent a test.
}
\section{Slots}{

\describe{
\item{\code{pool}}{An \code{\linkS4class{item_pool}} object.}

\item{\code{theta}}{A theta grid.}

\item{\code{prob}}{A list of item response probabilities.}

\item{\code{info}}{A matrix of item information values.}

\item{\code{true_theta}}{An optional vector of true theta values.}

\item{\code{data}}{An optional matrix of item responses.}
}}

