% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_cache_class.r
\docType{class}
\name{test-class}
\alias{test-class}
\title{Class 'test': data cache for simulations}
\description{
\code{\linkS4class{test}} is an S4 class to represent data cache for running simulations.
Despite the name, this class does not represent a test and is not related to a test. That is, test length is not stored in this class.
This class is only kept for backwards compatibility.
The functionality of this class is superseded by \code{\linkS4class{simulation_data_cache}}.
}
\section{Slots}{

\describe{
\item{\code{pool}}{the \code{\linkS4class{item_pool}} object.}

\item{\code{theta}}{the theta grid to use as quadrature points.}

\item{\code{prob}}{the list containing item response probabilities.}

\item{\code{info}}{the matrix containing item information values.}

\item{\code{true_theta}}{(optional) the true theta values.}

\item{\code{data}}{(optional) the matrix containing item responses.}
}}

