% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_class.R
\docType{class}
\name{test-class}
\alias{test-class}
\title{Class 'test': data for test assembly}
\description{
\code{\linkS4class{test}} is an S4 class to represent data for test assembly.
}
\section{Slots}{

\describe{
\item{\code{pool}}{the \code{\linkS4class{item_pool}} object.}

\item{\code{theta}}{the theta grid to use as quadrature points.}

\item{\code{prob}}{the list containing item response probabilities.}

\item{\code{info}}{the matrix containing item information values.}

\item{\code{true_theta}}{(optional) the true theta values.}

\item{\code{data}}{(optional) the matrix containing item responses.}
}}

