% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shadow_class.R
\docType{class}
\name{test.cluster-class}
\alias{test.cluster-class}
\title{test.cluster}
\description{
test.cluster
}
\section{Slots}{

\describe{
\item{\code{nt}}{Numeric.}

\item{\code{tests}}{A list.}

\item{\code{names}}{Character.}
}}

