% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data_cache_class.r
\docType{class}
\name{test_cluster-class}
\alias{test_cluster-class}
\title{Class 'test_cluster': data cache for simulations}
\description{
\code{\linkS4class{test_cluster}} is an S4 class to represent data cache for running simulations.
Despite the name, this class does not represent a series of tests and is not related to a series of tests. That is, test length is not stored in this class.
This class is only kept for backwards compatibility.
}
\section{Slots}{

\describe{
\item{\code{nt}}{the number of \code{\linkS4class{test}} objects in this cluster.}

\item{\code{tests}}{the list containing \code{\linkS4class{test}} objects.}

\item{\code{names}}{test ID strings for each \code{\linkS4class{test}} object.}
}}

