% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_class.R
\docType{class}
\name{test_cluster-class}
\alias{test_cluster-class}
\title{Class 'test_cluster': data for test assembly}
\description{
\code{\linkS4class{test_cluster}} is an S4 class to represent data for test assembly.
}
\section{Slots}{

\describe{
\item{\code{nt}}{the number of \code{\linkS4class{test}} objects in this cluster.}

\item{\code{tests}}{the list containing \code{\linkS4class{test}} objects.}

\item{\code{names}}{test ID strings for each \code{\linkS4class{test}} object.}
}}

