% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_EAP}
\alias{theta_EAP}
\title{Calculate an EAP estimate of theta for one examinee}
\usage{
theta_EAP(theta_grid, item_parm, resp, ncat, model, prior, prior_parm)
}
\arguments{
\item{theta_grid}{An equi-spaced theta grid.}

\item{item_parm}{A numeric matrix of item parameters.}

\item{resp}{a numeric vector containing item responses.}

\item{ncat}{A numeric vector of the number of response categories by item.}

\item{model}{A numeric vector indicating the IRT models of each item (1: 1PL, 2: 2PL, 3: 3PL, 4: PC, 5: GPC, 6: GR).}

\item{prior}{The type of prior distribution (1: normal, 2: uniform).}

\item{prior_parm}{A numeric vector of hyperparameters for the prior distribution, c(mu, sigma) or c(ll, ul).}
}
\description{
Calculate an expected a posterior estimate of theta for one examinee.
}
