% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_EAP_matrix}
\alias{theta_EAP_matrix}
\title{theta_EAP_matrix}
\usage{
theta_EAP_matrix(theta_grid, item_parm, Resp, ncat, model, prior,
  prior_parm)
}
\arguments{
\item{theta_grid}{A nice parameter}

\item{item_parm}{A nice parameter}

\item{Resp}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}

\item{prior}{A nice parameter}

\item{prior_parm}{A nice parameter}
}
\description{
theta_EAP_matrix
}
