% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_EB_single}
\alias{theta_EB_single}
\title{Calculate an empirical Bayes estimate of theta for a single item}
\usage{
theta_EB_single(nx, theta_init, theta_prop, item_parm, resp, ncat, model,
  prior, prior_parm)
}
\arguments{
\item{nx}{a number of MCMC draws}

\item{theta_init}{an initial theta estimate}

\item{theta_prop}{SD of the proposal distribution}

\item{item_parm}{a numeric matrix of item parameters}

\item{resp}{a numeric vector item responses}

\item{ncat}{a numeric vector of the number of response categories by item}

\item{model}{a numeric vector of the IRT model by item (1: 1pl, 2: 2pl, 3: 3pl, 4: pc, 5: gpc, 6: gr)}

\item{prior}{prior distribution (1: normal, 2: uniform)}

\item{prior_parm}{a numeric vector of hyper parameters for the prior distribution, c(mu, sigma) or c(ll, ul)}
}
\description{
Calculate an empirical Bayes estimate of theta for a single item
}
