% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theta_FB_single}
\alias{theta_FB_single}
\title{theta_FB_single}
\usage{
theta_FB_single(nx, theta_init, theta_prop, item_mcmc, item_init, resp,
  ncat, model, prior, prior_parm)
}
\arguments{
\item{nx}{A nice parameter}

\item{theta_init}{A nice parameter}

\item{theta_prop}{A nice parameter}

\item{item_mcmc}{A nice parameter}

\item{item_init}{A nice parameter}

\item{resp}{A nice parameter}

\item{ncat}{A nice parameter}

\item{model}{A nice parameter}

\item{prior}{A nice parameter}

\item{prior_parm}{A nice parameter}
}
\description{
theta_FB_single
}
