% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loading_functions.R
\name{UpdateConstraints}
\alias{UpdateConstraints}
\title{Update constraints}
\usage{
UpdateConstraints(object, on = NULL, off = NULL)
}
\arguments{
\item{object}{a list object returned from \code{\link{LoadConstraints}}}

\item{on}{a vector of constraints index to turn on.}

\item{off}{a vector of constraints index to turn off.}
}
\value{
An updated list of constraints to be used in \code{\link{ATA}} and \code{\link{Shadow}}.
}
\description{
Update the onstraints list
}
