% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{updateConstraints}
\alias{updateConstraints}
\title{Update constraints}
\usage{
updateConstraints(object, on = NULL, off = NULL)
}
\arguments{
\item{object}{a list object returned from \code{\link{loadConstraints}}}

\item{on}{a vector of constraints index to turn on.}

\item{off}{a vector of constraints index to turn off.}
}
\value{
An updated list of constraints to be used in \code{\link{ATA}} and \code{\link{Shadow}}.
}
\description{
Update the onstraints list
}
\examples{
constraints_science2 <- updateConstraints(constraints_science, off = 32:36)
constraints_science3 <- updateConstraints(constraints_science, on = 32:36)

}
