% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{updateConstraints}
\alias{updateConstraints}
\title{Update constraints}
\usage{
updateConstraints(object, on = NULL, off = NULL)
}
\arguments{
\item{object}{A \code{\linkS4class{constraints}} object from \code{\link{loadConstraints}}.}

\item{on}{a vector of constraints index to turn on.}

\item{off}{a vector of constraints index to turn off.}
}
\value{
An updated \code{\linkS4class{constraints}} object, to be used in \code{\link{Static}} and \code{\link{Shadow}}.
}
\description{
Update the constraints list
}
\examples{
constraints_science2 <- updateConstraints(constraints_science, off = 32:36)
constraints_science3 <- updateConstraints(constraints_science, on = 32:36)

}
