% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{updateConstraints}
\alias{updateConstraints}
\title{(deprecated) Update constraints}
\usage{
updateConstraints(object, on = NULL, off = NULL)
}
\arguments{
\item{object}{a \code{\linkS4class{constraints}} object from \code{\link{loadConstraints}}.}

\item{on}{constraint indices to mark as active.}

\item{off}{constraint indices to mark as inactive.}
}
\value{
\code{\link{updateConstraints}} returns the updated \code{\linkS4class{constraints}} object.
}
\description{
Use \code{\link{toggleConstraints}} instead.
}
\examples{
\dontrun{
constraints_science2 <- updateConstraints(constraints_science, off = 32:36)
constraints_science3 <- updateConstraints(constraints_science, on = 32:36)
}
}
